@@ -1,30 +1,19 @@
 {
+  "name": "@rodewitsch/carbone",
+  "description": "Fast, Simple and Powerful report generator. Injects JSON and produces PDF, DOCX, XLSX, ODT, PPTX, ODS, ...!",
+  "version": "5.0.0-beta",
+  "bin": "bin/carbone",
+  "main": "./lib",
+  "types": "types/lib/index.d.ts",
+  "bundleDependencies": false,
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/rodewitsch/carbone"
+  },
   "author": {
     "name": "Fork of the original work of David Grelaud"
   },
-  "bin": {
-    "carbone": "bin/carbone"
-  },
-  "types": "types/lib/index.d.ts",
-  "bundleDependencies": false,
-  "dependencies": {
-    "dayjs": "^1.11.7",
-    "debug": "=2.6.9",
-    "fastq": "^1.15.0",
-    "html2json": "^1.0.2",
-    "moxie-zip": "=0.0.3",
-    "timsort": "=0.3.0",
-    "xml-js": "^1.6.11",
-    "yauzl": "=2.10.0"
-  },
-  "deprecated": false,
-  "description": "Fast, Simple and Powerful report generator. Injects JSON and produces PDF, DOCX, XLSX, ODT, PPTX, ODS, ...!",
-  "devDependencies": {
-    "mocha": "^9.2.2"
-  },
-  "engines": {
-    "node": ">= 0.10.0"
-  },
+  "license": "SEE LICENSE IN LICENSE.md",
   "homepage": "https://carbone.io",
   "keywords": [
     "report",
@@ -36,13 +25,25 @@
     "ods",
     "json"
   ],
-  "license": "Apache-2.0",
-  "main": "./lib",
-  "name": "@rodewitsch/carbone",
   "scripts": {
-    "lint": "eslint ./*/**.js",
-    "lint:fix": "eslint ./*/**.js --fix",
-    "test": "node ./node_modules/mocha/bin/mocha test --timeout 100000 --exit"
+    "test": "./node_modules/.bin/mocha test --timeout 100000 --exit",
+    "lint": "eslint ./lib/**.js ./test/**.js ./formatters/**.js",
+    "lint:fix": "eslint ./lib/**.js ./test/**.js ./formatters/**.js --fix"
   },
-  "version": "4.4.2"
+  "dependencies": {
+    "dayjs": "=1.11.11",
+    "dayjs-timezone-iana-plugin": "=0.1.0",
+    "debug": "=4.3.5",
+    "html2json": "^1.0.2",
+    "which": "=2.0.2",
+    "xml-js": "^1.6.11",
+    "yauzl": "=2.10.0",
+    "yazl": "=2.5.1"
+  },
+  "devDependencies": {
+    "mocha": "=9.2.2"
+  },
+  "engines": {
+    "node": ">= 12.14.0"
+  }
 }
