@@ -1,13 +1,10 @@
 var path  = require('path');
 var fs = require('fs');
 var params = require('./params');
-var zipWriter = require('moxie-zip').ZipWriter;
 var yauzl = require('yauzl');
 var unzipEmbeddedFileTypes = ['.xlsx', '.ods'];
-
-var nodeVersion = process.versions.node.split('.');
-var isNewBufferDeprecated = (parseInt(nodeVersion[0], 10) === 4 && parseInt(nodeVersion[1], 10) >= 5)
-                          || parseInt(nodeVersion[0], 10) > 5;
+var yazl = require('yazl');
+var debug = require('debug')('carbone');
 
 var file = {
 
@@ -16,23 +13,36 @@ var file = {
    * @param  {String}   filePath file path
    * @param  {Function} callback(err, isZipped)
    */
-  isZipped: function (filePath, callback) {
-    var _buf = isNewBufferDeprecated === true ? Buffer.allocUnsafe(10) : new Buffer(10, 'hex');
-    if (Buffer.isBuffer(filePath)) {
-      const buffer = Buffer.from(filePath, 0, 10).toString();
-      callback(null, (buffer.slice(0, 2).toString() === 'PK'));
-    } else {
-      fs.open(filePath, 'r', function (err, fd) {
-        if (err) {
-          return callback(err, false);
-        }
-        fs.read(fd, _buf, 0, 10, 0, function (err, bytesRead, buffer) {
-          fs.close(fd, function () {
-            callback(err, (buffer.slice(0, 2).toString() === 'PK'));
-          });
+  isZipped : function (filePath, callback) {
+    var _buf = Buffer.allocUnsafe(10);
+    fs.open(filePath, 'r', function (err, fd) {
+      if (err) {
+        return callback(err, false);
+      }
+      fs.read(fd, _buf, 0, 10, 0, function (err, bytesRead, buffer) {
+        fs.close(fd, function () {
+          callback(err, (buffer.slice(0, 2).toString() === 'PK'));
         });
       });
+    });
+  },
+
+  /**
+   * Get specific file in template
+   *
+   * TODO test it
+   *
+   * @param  {Object}    template
+   * @param  {String}    filename
+   * @return {Object|null}
+   */
+  getTemplateFile : function (template, filename) {
+    for (var i = 0; i < template.files.length; i++) {
+      if (template.files[i].name === filename) {
+        return template.files[i];
+      }
     }
+    return null;
   },
 
   /**
@@ -59,7 +69,7 @@ var file = {
       zipfile.on('entry', function (entry) {
         var _unzippedFile = {
           name : entry.fileName,
-          data : new Buffer(0)
+          data : Buffer.from([])
         };
         _unzippedFiles.push(_unzippedFile);
         if (/\/$/.test(entry.fileName)) {
@@ -92,18 +102,32 @@ var file = {
 
   /**
    * Zip a group of files
-   * @param  {Array}    files                 files is an array of files ['name':'filename', 'buffer':Buffer]
+   * @param  {Array}    files                 files is an array of files ['name':'filename', 'data':Buffer]
    * @param  {Function} callback(err, result) result is a buffer (the zip file)
    */
   zip : function (files, callback) {
-    var zip = new zipWriter();
+    var _buffer = [];
+    var _zip = new yazl.ZipFile();
+    _zip.outputStream.on('data', function (data) {
+      _buffer.push(data);
+    });
+    _zip.outputStream.on('error', function (err) {
+      debug('Error when building zip file ' + err);
+    });
+    _zip.outputStream.on('end', function () {
+      var _finalBuffer = Buffer.concat(_buffer);
+      callback(null, _finalBuffer);
+    });
     for (var i = 0; i < files.length; i++) {
       var _file = files[i];
-      zip.addData(_file.name, _file.data);
+      if (_file.name.endsWith('/') === true) {
+        _zip.addEmptyDirectory(_file.name);
+      }
+      else {
+        _zip.addBuffer(Buffer.from(_file.data), _file.name);
+      }
     }
-    zip.toBuffer(function (buf) {
-      callback(null, buf);
-    });
+    _zip.end();
   },
 
   /**
@@ -118,13 +142,16 @@ var file = {
       embeddings : [],
       files      : []
     };
+    if (templateId instanceof Object) {
+      // accept already dezipped files (for test purpose)
+      return callback(null, templateId);
+    }
     // security, remove access on parent directory
     // if (/\.\./.test(path.dirname(templateId)) === true) {
     //   return callback('access forbidden');
     // }
     // and then use path instead of resolve
-    var _templateFile;
-    if (!Buffer.isBuffer(templateId)) _templateFile = path.resolve(params.templatePath, templateId); else _templateFile = templateId;
+    var _templateFile = path.resolve(params.templatePath, templateId);
     file.isZipped(_templateFile, function (err, isZipped) {
       if (err) {
         return callback(err, _template);
@@ -166,6 +193,9 @@ var file = {
     if (this._checkWordInFilename(template, 'ppt/')) {
       return 'pptx';
     }
+    if (this._checkMimetypeFile(template, 'application/vnd.oasis.opendocument.graphics')) {
+      return 'odg';
+    }
     if (this._checkMimetypeFile(template, 'application/vnd.oasis.opendocument.text')) {
       return 'odt';
     }
@@ -174,6 +204,12 @@ var file = {
     }
     if (this._checkMimetypeFile(template, 'application/vnd.oasis.opendocument.presentation')) {
       return 'odp';
+    }
+    if (this._checkMimetypeFile(template, 'application/vnd.adobe.indesign-idml-package')) {
+      return 'idml';
+    }
+    if (template.files.length === 1 && /office:mimetype="application\/vnd\.oasis\.opendocument\.text"/.test(template.files[0].data) === true) {
+      return 'fodt';
     }
     if (this._isXHTMLFile(template)) {
       return 'xhtml';
@@ -197,9 +233,8 @@ var file = {
    */
   _isXMLFile : function (template) {
     for (var i = 0; i < template.files.length; i++) {
-      var _trimmedTemplate = template.files[i].data.trim();
-
-      if (_trimmedTemplate.startsWith('<')) {
+      var _template = template.files[i].data;
+      if (/^\s*</.test(_template) === true) {
         return true;
       }
     }
@@ -213,11 +248,9 @@ var file = {
    */
   _isXHTMLFile : function (template) {
     for (var i = 0; i < template.files.length; i++) {
-      var _trimmedTemplate = template.files[i].data.trim();
-      var _xmlnsRegex = /<html xmlns="/gm;
-
-      if (_trimmedTemplate.startsWith('<!DOCTYPE')) {
-        if (_xmlnsRegex.test(_trimmedTemplate)) {
+      var _template = template.files[i].data;
+      if (/^\s*<!DOCTYPE/i.test(_template) === true) {
+        if (/<html\s+xmlns="/i.test(_template) === true) {
           return true;
         }
       }
@@ -232,16 +265,11 @@ var file = {
    */
   _isHTMLFile : function (template) {
     for (var i = 0; i < template.files.length; i++) {
-      var _trimmedTemplate = template.files[i].data.trim();
-      var _htmlRegex = /<html/gm;
-
-      if (_trimmedTemplate.startsWith('<!DOCTYPE')) {
-        if (_htmlRegex.test(_trimmedTemplate)) {
-          return true;
-        }
+      var _template = template.files[i].data;
+      if (/<html/i.test(_template) === true) {
+        return true;
       }
     }
-
     return false;
   },
 
@@ -288,7 +316,9 @@ var file = {
     }
     else {
       if (report.files.length !== 1) {
-        throw Error('This report is not zipped and does not contain exactly one file');
+        // only for test purpose when a template is not zipped
+        return callback(null, report);
+        // throw Error('This report is not zipped and does not contain exactly one file');
       }
       return callback(null, report.files[0].data);
     }
@@ -349,10 +379,10 @@ function zipFiles (filesToZip, callback) {
     }
     if (Buffer.isBuffer(_file.data) === false) {
       try {
-        _file.data = new Buffer(_file.data, 'utf8');
+        _file.data = Buffer.from(_file.data, 'utf8');
       }
       catch (e) {
-        _file.data = new Buffer('', 'utf8');
+        _file.data = Buffer.from('', 'utf8');
       }
     }
     _previousParentName = _file.parent;
