@@ -1,4 +1,3 @@
-/* eslint-disable eqeqeq */
 const locale   = require('./_locale.js');
 const currency = require('./_currency.js');
 
@@ -9,6 +8,8 @@ const currency = require('./_currency.js');
  * for one report in `options.currencyRates` of `Carbone.render` or globally with `Carbone.set`
  *
  * `convCurr()`  without parameters converts automatically from `options.currencySource` to `options.currencyTarget`.
+ *
+ * @version 1.2.0
  *
  * @exampleContext {"currency": { "source":"EUR", "target":"USD", "rates": { "EUR":1, "USD":2 }} }
  * @example [10                  ]
@@ -38,6 +39,8 @@ function convCurr (d, target, source) {
  * Round a number
  *
  * Same as toFixed(2) but it rounds number correclty `round(1.05, 1) = 1.1`
+ *
+ * @version 1.2.0
  *
  * @example [10.05123  , 2  ]
  * @example [1.05      , 1  ]
@@ -90,14 +93,19 @@ function _format (value, format, precision = 3) {
 }
 
 /**
- * Format number according to the locale
+ * Format number according to the locale.
+ * Applying a number of decimals depends on the report type:
+ * - For ODS/XLSX, the number of decimals has to be formatted based on the text editor.
+ * - For the other type of files, the number of decimals depends on the `precision` parameter passed to the formatter.
+ *
+ * @version 1.2.0
  *
  * @exampleContext {"lang":"en-us"}
  * @example ["10"          ]
  * @example ["1000.456"    ]
  *
  * @param  {Number} d          Number to format
- * @param  {Number} precision  [optional] Number of decimal
+ * @param  {Number} precision  [optional] Number of decimals
  * @return {String} return     converted values
  */
 function formatN (d, precision) {
@@ -109,6 +117,9 @@ function formatN (d, precision) {
 }
 
 /**
+ *
+ * @version 1.2.0
+ *
  * Format currency numbers
  *
  * Currencies are defined by the locale (`options.lang`). It can be overwritten by
@@ -163,7 +174,9 @@ function formatC (d, precisionOrFormat) {
     return _formatFn(_valueRaw,
       _currencyInfo.symbol,
       _currencyInfo.minSymbol,
+      // eslint-disable-next-line eqeqeq
       (d != 1 ? _currencyInfo.major + 's' : _currencyInfo.major),
+      // eslint-disable-next-line eqeqeq
       (d != 1 ? _currencyInfo.minor + 's' : _currencyInfo.minor),
       _currencyInfo.name
     );
@@ -173,6 +186,8 @@ function formatC (d, precisionOrFormat) {
 
 /**
  * Add two numbers
+ *
+ * @version 1.2.0
  *
  * @example [1000.4  ,  2  ]
  * @example ["1000.4", "2" ]
@@ -186,9 +201,12 @@ function add (d, value) {
   }
   return d;
 }
+add.isAcceptingMathExpression = true;
 
 /**
  * Substract two numbers
+ *
+ * @version 1.2.0
  *
  * @example [1000.4  ,  2  ]
  * @example ["1000.4", "2" ]
@@ -202,9 +220,12 @@ function sub (d, value) {
   }
   return d;
 }
+sub.isAcceptingMathExpression = true;
 
 /**
  * Multiply two numbers
+ *
+ * @version 1.2.0
  *
  * @example [1000.4  ,  2  ]
  * @example ["1000.4", "2" ]
@@ -218,9 +239,12 @@ function mul (d, value) {
   }
   return d;
 }
+mul.isAcceptingMathExpression = true;
 
 /**
  * Divide two numbers
+ *
+ * @version 1.2.0
  *
  * @example [1000.4   ,  2  ]
  * @example ["1000.4" , "2" ]
@@ -234,6 +258,7 @@ function div (d, value) {
   }
   return d;
 }
+div.isAcceptingMathExpression = true;
 
 module.exports = {
   formatN  : formatN,
@@ -248,6 +273,7 @@ module.exports = {
   /**
    * Converts a number to an INT
    * @deprecated
+   * @version 1.0.0 deprecated
    *
    * @return {Number}
    */
@@ -258,6 +284,7 @@ module.exports = {
   /**
    * Converts a number with English specifications (decimal separator is '.')
    * @deprecated
+   * @version 1.0.0 deprecated
    *
    * @return {String}
    */
@@ -268,6 +295,7 @@ module.exports = {
   /**
    * Converts a number into string, keeping only <nb> decimals
    * @deprecated
+   * @version 1.0.0 deprecated
    *
    * @param  {Number} nb
    * @return {String}
@@ -279,6 +307,7 @@ module.exports = {
   /**
    * Converts a number with French specifications (decimal separator is ',')
    * @deprecated
+   * @version 1.0.0 deprecated
    *
    * @return {String}
    */
