@@ -1,7 +1,6 @@
 var extracter = require('./extracter');
 var parser = require('./parser');
 var helper = require('./helper');
-var timSort = require('timsort');
 
 var builder = {
 
@@ -19,51 +18,56 @@ var builder = {
    *                                }
    * @param  {Function}     callback
    */
-  buildXML: function (xml, data, options, callback) {
-    if (typeof (options) === 'function') {
+  buildXML : function (xml, data, options, callback) {
+    if (typeof(options) === 'function') {
       callback = options;
       options = {};
     }
 
-    // find translation markers {t()} and translate
-    parser.translate(xml, options, function (err, xmlWithoutTranslation) {
-      // find declared variable {#myVar }
-      parser.findVariables(xmlWithoutTranslation, options.existingVariables, function (err, xmlWithoutVariable, variables) {
-        // find markers { d. } or { c. }
-        parser.findMarkers(xmlWithoutVariable, function (err, xmlWithoutMarkers, markers) {
-          // exit if there is no markers
-          if (markers.length === 0) {
-            return callback(null, xmlWithoutVariable);
+    // find declared variable {#myVar }
+    parser.findVariables(xml, options.existingVariables, function (err, xmlWithoutVariable, variables) {
+      // find markers { d. } or { c. }
+      parser.findMarkers(xmlWithoutVariable, function (err, xmlWithoutMarkers, markers) {
+        // exit if there is no markers
+        if (markers.length === 0) {
+          return callback(null, xmlWithoutVariable);
+        }
+        parser.preprocessMarkers(markers, variables, function (err, preprocessedMarkers) {
+          var _xmlParts = [];
+          var _xmlResult = '';
+          var _builder;
+          if (options.isDebugActive === true) {
+            options.debugInfo.markers = [...options.debugInfo.markers, ...preprocessedMarkers.map((marker) => '{'+marker.name.replace(/^_root\./, '')+'}' )];
           }
-          parser.preprocessMarkers(markers, variables, function (err, preprocessedMarkers) {
-            var _xmlParts = [];
-            try {
-              var _dynamicDescriptor = extracter.splitMarkers(preprocessedMarkers);
-              var _descriptor = extracter.splitXml(xmlWithoutMarkers, _dynamicDescriptor);
-              _descriptor = extracter.buildSortedHierarchy(_descriptor);
-              _descriptor = extracter.deleteAndMoveNestedParts(_descriptor);
-              var _builder = builder.getBuilderFunction(_descriptor, options.formatters);
-              var _obj = {
-                d: data,
-                c: options.complement
-              };
-              _xmlParts = _builder(_obj, options, helper);
-            }
-            catch (e) {
-              return callback(e, null);
-            }
-            var _xmlResult = builder.assembleXmlParts(_xmlParts, 20); // TODO -> adapt the depth of the sort according to the maximum depth in _xmlParts
-            return callback(null, _xmlResult);
-          });
+          try {
+            var _dynamicDescriptor = extracter.splitMarkers(preprocessedMarkers);
+            var _descriptor = extracter.splitXml(xmlWithoutMarkers, _dynamicDescriptor);
+            _descriptor = extracter.buildSortedHierarchy(_descriptor);
+            _descriptor = extracter.deleteAndMoveNestedParts(_descriptor);
+            var _builder = builder.getBuilderFunction(_descriptor, options.formatters);
+            var _obj = {
+              d : data,
+              c : options.complement
+            };
+            options.stopPropagation = false;
+            _xmlParts = _builder(_obj, options, helper, _builder.builderDictionary);
+            _xmlResult = builder.assembleXmlParts(_xmlParts, 20, _builder.builderDictionary); // TODO -> adapt the depth of the sort according to the maximum depth in _xmlParts
+          }
+          catch (e) {
+            return callback(e, null);
+          }
+          return callback(null, _xmlResult);
         });
       });
     });
+
   },
 
   /**
    * Creates a string of the formatter(s) function(s) call of the attribute
    * INTERNAL USE ONLY
    *
+   * @param  {Function} getSafeValue generate safe value accessor
    * @param  {String} attr. Example 'd.number'
    * @param  {Array} formatters. Example [ 'int', 'toFixed(2)' ]
    * @param  {Object} existingFormatters
@@ -72,8 +76,22 @@ var builder = {
    *                                                  the property canInjectXML = true
    * @return {String}. Example 'toFixed(int(d.number), 2)'
    */
-  getFormatterString: function (varName, contextName, formatters, existingFormatters, onlyFormatterWhichInjectXML) {
+  getFormatterString : function (getSafeValue, varName, contextName, formatters, existingFormatters, onlyFormatterWhichInjectXML) {
     var _lineOfCodes = [];
+
+    function getInjectedVariable (variable) {
+      var _injectedArgument = getSafeValue(variable);
+      if (variable.startsWith('.') === true) {
+        var _nbPoint = 0;
+        for (; variable[_nbPoint] === '.'; _nbPoint++) {
+          // count number of point nothing
+        }
+        var _dynamicVariable = variable.slice(_nbPoint);
+        _injectedArgument = 'helper.getValueOfPath( '+contextName+'.parentsData['+(_nbPoint-1)+'], '+getSafeValue(_dynamicVariable)+')';
+      }
+      return _injectedArgument;
+    }
+
     for (var i = 0; i < formatters.length; i++) {
       var _formatter = formatters[i];
       // Percent-Encode the commas when the paramater is string with single quotes: ',' to '%2c'
@@ -87,23 +105,18 @@ var builder = {
       if (_indexFirstParenthesis !== -1) {
         _functionStr = _formatter.slice(0, _indexFirstParenthesis);
         var _indexLastParenthesis = _formatter.lastIndexOf(')'); // TODO error if last Parenthesis is not found
-        var _arguments = _formatter.slice(_indexFirstParenthesis + 1, _indexLastParenthesis).split(/ *, */);
+        var _arguments = _formatter.slice(_indexFirstParenthesis+1, _indexLastParenthesis).split(/ *, */);
         for (var j = 0; j < _arguments.length; j++) {
           if (_arguments[j] !== '') {
             // remove existing quotes (everything is converted to a string for the momemt. It should change in the future)
             // Percent-Decode the commas to obtain the initial string argument: '%2c' to ','
             var _argument = _arguments[j].replace(/^ *'?/, '').replace(/'? *$/, '').replace(/%2c/g, ',');
-            var _injectedArgument = `'${_argument}'`;
-            if (_argument.startsWith('.') === true) {
-              var _nbPoint = 0;
-              for (; _argument[_nbPoint] === '.'; _nbPoint++) {
-                // count number of point nothing
-              }
-              var _dynamicVariable = _argument.slice(_nbPoint);
-              // TODO warn if undefined variable
-              _injectedArgument = `helper.getValueOfPath( ${contextName}.parentsData[${_nbPoint - 1}], '${_dynamicVariable}') `;
+            if (existingFormatters?.[_functionStr]?.isAcceptingMathExpression === true) {
+              _argumentStr += ', ' + parser.parseMathematicalExpression(_argument, getInjectedVariable)
             }
-            _argumentStr += `, ${_injectedArgument}`;
+            else {
+              _argumentStr += ', ' + getInjectedVariable(_argument);
+            }
           }
         }
       }
@@ -112,14 +125,14 @@ var builder = {
       }
       if (existingFormatters[_functionStr] === undefined) {
         var _alternativeFnName = helper.findClosest(_functionStr, existingFormatters);
-        throw Error('Formatter "' + _functionStr + '" does not exist. Do you mean "' + _alternativeFnName + '"?');
+        throw Error('Formatter "'+_functionStr+'" does not exist. Do you mean "'+_alternativeFnName+'"?');
       }
-      if ((existingFormatters[_functionStr].canInjectXML === true && onlyFormatterWhichInjectXML === true)
+      if ( (existingFormatters[_functionStr].canInjectXML === true && onlyFormatterWhichInjectXML === true)
         || (existingFormatters[_functionStr].canInjectXML !== true && onlyFormatterWhichInjectXML !== true)) {
-        _lineOfCodes.push(varName + ' = formatters.' + _functionStr + '.call(' + contextName + ', ' + varName + _argumentStr + ');\n');
+        _lineOfCodes.push(varName +' = formatters.' + _functionStr + '.call(' + contextName + ', ' + varName + _argumentStr + ');\n');
       }
     }
-    var _str = _lineOfCodes.join('if(' + contextName + '.stopPropagation === false){\n');
+    var _str = _lineOfCodes.join('if('+contextName+'.stopPropagation === false){\n');
     var _nbClosingBraces = _lineOfCodes.length - 1;
     while (_nbClosingBraces-- > 0) {
       _str += '}';
@@ -130,6 +143,8 @@ var builder = {
   /**
    * build the condition string which will be used to filter rows
    *
+   * @param {Function} getSafeVar   : generate safe variable
+   * @param {Function} getSafeValue : generate safe value accessor
    * @param {array} conditions : array of conditions. Example [{'left':'sort', 'operator':'>', 'right':'10'}]
    * @param {string} codeIfTrue : code to insert if the condition is true. Example 'row = true'
    * @param {string} prefix (optional) : if we need to create local variable, this prefix will be added in front of the variable.
@@ -137,16 +152,16 @@ var builder = {
    * @param {string} forceObjectTested (optional) : use forceObjectTested instead of _condition.left.parent
    * @return {string}. Example : 'if(d.number['sort'] > 10) { row = true }'
    */
-  getFilterString: function (conditions, codeIfTrue, prefix, inverseCondition, forceObjectTested) {
+  getFilterString : function (getSafeVar, getSafeValue,  conditions, codeIfTrue, prefix, inverseCondition, forceObjectTested) {
     prefix = prefix || '';
-    if (!codeIfTrue || !(conditions instanceof Array) || conditions.length === 0) {
+    if (!codeIfTrue || !(conditions instanceof Array) || conditions.length ===0) {
       return '';
     }
     var _str = '';
     var _alreadyDeclaredFilter = {};
     var _declareStr = '';
     _str += 'if(';
-    _str += (inverseCondition === true) ? '!(' : '(';
+    _str += (inverseCondition === true) ? '!(':'(';
     for (var c = 0; c < conditions.length; c++) {
       var _condition = conditions[c];
       var _attr = _condition.left.attr;
@@ -158,26 +173,31 @@ var builder = {
       var _objectSeparatorIndex = _attr.indexOf('.');
       if (_objectSeparatorIndex !== -1) {
         var _obj = _attr.slice(0, _objectSeparatorIndex);
-        _attr = _attr.slice(_objectSeparatorIndex + 1);
-        var _subObjName = '_' + prefix + _obj + '_filter';
+        _attr = _attr.slice(_objectSeparatorIndex+1);
+        var _subObjName = '_'+prefix+_obj+'_filter';
         if (_alreadyDeclaredFilter[_subObjName] === undefined) {
-          _declareStr += 'var ' + _subObjName + '=' + _objName + '["' + _obj + '"];\n';
+          _declareStr += 'var '+getSafeVar(_subObjName)+'='+getSafeVar(_objName)+'['+getSafeValue(_obj)+'];\n';
         }
         _objName = _subObjName;
       }
       if (_attr === 'i') {
-        var _rightOperator = _condition.right;
+        var _rightOperator = parseInt(_condition.right, 10);
         var _arrayName = _condition.left.parent;
         if (parseInt(_condition.right, 10) < 0) {
-          _rightOperator = _arrayName + '_array_length ' + _condition.right;
+          _rightOperator = getSafeVar(_arrayName) + '_array_length ' +  parseInt(_condition.right, 10);
         }
-        _str += _arrayName + '_i ' + _condition.operator + _rightOperator;
+        _str += getSafeVar(_arrayName) + '_i '+_condition.operator + _rightOperator;
       }
       else {
         if (_alreadyDeclaredFilter[_objName] === undefined) {
-          _str += _objName + ' && ';
+          _str += getSafeVar(_objName) + ' && ';
         }
-        _str += _objName + '["' + _attr + '"]' + _condition.operator + _condition.right;
+        _str += getSafeVar(_objName)+'['+getSafeValue(_attr)+']';
+        if (_condition.right === 'true' || _condition.right === 'false') {
+          // for boolean, convert data to string to accept both cases ('true' == true) and (true == true) for backward compatibility with Carbone v1 & v2
+          _str += "+''";
+        }
+        _str += _condition.operator + getSafeValue(helper.removeQuote(_condition.right));
       }
       if (c < conditions.length - 1) {
         _str += ' && ';
@@ -185,7 +205,7 @@ var builder = {
       _alreadyDeclaredFilter[_objName] = true;
     }
     _str += ')){\n ';
-    _str += codeIfTrue + ';\n';
+    _str += codeIfTrue +';\n';
     _str += ' }';
 
     return _declareStr + _str;
@@ -196,29 +216,32 @@ var builder = {
    *
    * @param {array} arrayOfStringPart :  special array which is generated by the function returned by getBuilderFunction
    * @param {integer} sortDepth : description
+   * @param {array} builderDictionary : xml part coming from template
    * @return {string} final result
    */
-  assembleXmlParts: function (arrayOfStringPart, sortDepth) {
+  assembleXmlParts : function (arrayOfStringPart, sortDepth, builderDictionary) {
     var _res = '';
     builder.sortXmlParts(arrayOfStringPart);
     var _rowInfo = [];
     var _prevDepth = 0;
-    var _prevPart = { pos: [] };
+    var _prevPart = {pos : []};
     var _arrayLevel = 0;
     var _loopIds = {};
     var _hideBlock = 0;
+    var _prevHideBlock = 0;
 
     // TODO MANAGE  sortDepth
     sortDepth = (sortDepth) ? sortDepth : 1;
     // init _rowInfo
     for (var loop = 0; loop <= sortDepth; loop++) {
       _rowInfo.push({
-        xmlPos: 0,
-        rowShow: 0
+        xmlPos  : 0,
+        rowShow : 0
       });
     }
     for (var i = 0; i < arrayOfStringPart.length; i++) {
       var _part = arrayOfStringPart[i];
+      var _str = '';
       // Get count value if needed
       builder.getLoopIteration(_loopIds, _part);
       // keep parts if positions are not the same as the last one OR if the it the beginning of an array
@@ -228,16 +251,26 @@ var builder = {
       else {
         continue;
       }
+      _prevHideBlock = _hideBlock;
       if (_part.hide !== undefined) {
         _hideBlock += (_hideBlock && 1) | _part.hide;
         if (_hideBlock < 0) {
           _hideBlock = 0;
         }
       }
-      if (_hideBlock > 0) {
+      if (_hideBlock > 0 && _prevHideBlock > 0) {
         continue;
       }
-      _res += _part.str;
+      // include "before" XML string all the time, except at the end of if block
+      if (!(_hideBlock === 0 && _prevHideBlock > 0)) {
+        _str = _part.bef !== undefined ? builderDictionary[_part.bef] : '';
+      }
+      // include "after" XML string and the value to inject (_part.str) all the time, except at the beginning of if block
+      if (!(_hideBlock > 0 && _prevHideBlock === 0)) {
+        _str += _part.str;
+        _str += _part.aft !== undefined ? builderDictionary[_part.aft] : '';
+      }
+      _res += _str;
 
       var _depth = _part.pos.length;
       if (_prevDepth < _depth) {
@@ -251,20 +284,20 @@ var builder = {
       // TODO. Protection if sortDepth is too short
       if (_rowInfo[_arrayLevel] === undefined) {
         _rowInfo[_arrayLevel] = {
-          xmlPos: 0,
-          rowShow: 0
+          xmlPos  : 0,
+          rowShow : 0
         };
       }
       _rowInfo[_arrayLevel].rowShow |= _part.rowShow;
-      if (_part.rowStart === true) {
-        _rowInfo[_arrayLevel].xmlPos = _res.length - _part.str.length;
+      if (_part.rowStart===true) {
+        _rowInfo[_arrayLevel].xmlPos = _res.length - _str.length;
       }
-      else if (_part.rowEnd === true) {
+      else if (_part.rowEnd===true) {
         if (_rowInfo[_arrayLevel].rowShow === 0) {
           _res = _res.slice(0, _rowInfo[_arrayLevel].xmlPos);
         }
         if (_arrayLevel > 0) {
-          _rowInfo[_arrayLevel - 1].rowShow |= _rowInfo[_arrayLevel].rowShow;
+          _rowInfo[_arrayLevel-1].rowShow |= _rowInfo[_arrayLevel].rowShow;
         }
         _rowInfo[_arrayLevel].rowShow = 0;
       }
@@ -279,7 +312,7 @@ var builder = {
    * @param {Object} loopIds  Stored loop IDs with their start value
    * @param {Object} part     Xml part
    */
-  getLoopIteration: function (loopIds, part) {
+  getLoopIteration : function (loopIds, part) {
     var _rowColumnRegex = /__COUNT_(\d*?)_(\d*?)__/;
     var _rowColumnMatch = _rowColumnRegex.exec(part.str);
 
@@ -303,9 +336,9 @@ var builder = {
    *
    * @param {array} arrayToSort : array of objects
    */
-  sortXmlParts: function (arrayToSort) {
+  sortXmlParts : function (arrayToSort) {
 
-    function compare(a, b) {
+    function compare (a, b) {
       var i = 0;
       var _a = a.pos[i];
       var _b = b.pos[i];
@@ -325,15 +358,7 @@ var builder = {
       return _aLength - _bLength;
     }
 
-    // TimSort is bugged: https://github.com/mziccard/node-timsort/issues/14
-    // While we wait for this fix we use default sort when TimSort crashes
-
-    try {
-      return timSort.sort(arrayToSort, compare);
-    }
-    catch (e) {
-      return arrayToSort.sort(compare);
-    }
+    return arrayToSort.sort(compare);
   },
 
   /**
@@ -345,29 +370,87 @@ var builder = {
    * @param {string} nextAttrName : next attribute name of the descriptor we will visit
    * @param {function} execute(exitedArrayName) : called for each array we are leaving.
    */
-  forEachArrayExit: function (currentlyVisitedArrays, objDependencyDescriptor, nextAttrName, execute) {
+  forEachArrayExit : function (currentlyVisitedArrays, objDependencyDescriptor, nextAttrName, execute) {
     var _currentParentAttr = '';
     if (objDependencyDescriptor[nextAttrName] !== undefined) {
       _currentParentAttr = objDependencyDescriptor[nextAttrName].parent;
     }
-    if (currentlyVisitedArrays.length > 0) {
+    if (currentlyVisitedArrays.length>0) {
       var _firstParentOfTypeArray = _currentParentAttr;
       // go up in the tree until we meet an array
       while (_firstParentOfTypeArray !== '' && objDependencyDescriptor[_firstParentOfTypeArray].type !== 'array') {
         _firstParentOfTypeArray = objDependencyDescriptor[_firstParentOfTypeArray].parent;
       }
-      var _lastArrayName = currentlyVisitedArrays[currentlyVisitedArrays.length - 1];
-      var _nextArrayXmlDepth = objDependencyDescriptor[nextAttrName].depth || objDependencyDescriptor[_firstParentOfTypeArray].depth || -1;
+      var _lastArrayName = currentlyVisitedArrays[currentlyVisitedArrays.length-1];
+      var _nextArrayXmlDepth = objDependencyDescriptor[nextAttrName].depth;
+      if (_nextArrayXmlDepth === undefined) {
+        _nextArrayXmlDepth = -1;
+        if (objDependencyDescriptor[_firstParentOfTypeArray] !== undefined && objDependencyDescriptor[_firstParentOfTypeArray].depth !== undefined) {
+          _nextArrayXmlDepth = objDependencyDescriptor[_firstParentOfTypeArray].depth;
+        }
+      }
       var _lastArrayXmlDepth = objDependencyDescriptor[_lastArrayName].depth || Number.MAX_SAFE_INTEGER;
-      while (_firstParentOfTypeArray !== _lastArrayName && currentlyVisitedArrays.length > 0 && _nextArrayXmlDepth <= _lastArrayXmlDepth) {
+      while (_firstParentOfTypeArray !== _lastArrayName && currentlyVisitedArrays.length>0 && _nextArrayXmlDepth <= _lastArrayXmlDepth) {
         execute(_lastArrayName);
         currentlyVisitedArrays.pop();
-        if (currentlyVisitedArrays.length > 0) {
-          _lastArrayName = currentlyVisitedArrays[currentlyVisitedArrays.length - 1];
+        if (currentlyVisitedArrays.length>0) {
+          _lastArrayName = currentlyVisitedArrays[currentlyVisitedArrays.length-1];
           _lastArrayXmlDepth = objDependencyDescriptor[_lastArrayName].depth || Number.MAX_SAFE_INTEGER;
         }
       }
     }
+  },
+
+  /**
+   * Generate safe and unique variable names generator for builder
+   *
+   * @return {Function} a function which generate variable names
+   */
+  generateSafeJSVariable : function () {
+    let _dictionary = new Map();
+    return (jsVariable) => {
+      let _alreadyCreatedVar = _dictionary.get(jsVariable);
+      if (_alreadyCreatedVar !== undefined) {
+        return _alreadyCreatedVar;
+      }
+      let _newVar = '_gV' + _dictionary.size;
+      _dictionary.set(jsVariable, _newVar);
+      return _newVar;
+    };
+  },
+
+  /**
+   * Generate a functions which store and get values coming from insecure
+   * source (XMl, carbone markers in document, formatters variable, ...) in a dictionary
+   *
+   * @param  {String} dictionaryName name of the dictionary in the builder function
+   */
+  generateSafeJSValueAccessor : function (dictionaryName) {
+    let _dictionaryIndex = new Map();
+    let _dictionary      = [];
+
+    function getIndex (xmlOrConstantValue) {
+      let _dictIndex = _dictionaryIndex.get(xmlOrConstantValue);
+      if (_dictIndex !== undefined) {
+        return _dictIndex;
+      }
+      _dictIndex = (_dictionary.push(xmlOrConstantValue) - 1);
+      _dictionaryIndex.set(xmlOrConstantValue, _dictIndex);
+      return _dictIndex;
+    }
+    return {
+      // get index where the value or XML part is stored in the dictionary
+      getIndex : getIndex,
+      // get the JS code to access securely to a dictionary's value in the builder function
+      get      : function (xmlOrConstantValue) {
+        let _dictIndex = getIndex(xmlOrConstantValue);
+        return dictionaryName + '[' + _dictIndex + ']';
+      },
+      // get the dictionary
+      getDictionary : function () {
+        return _dictionary;
+      }
+    };
   },
 
   /**
@@ -377,14 +460,14 @@ var builder = {
    * @param {object} descriptor : data descriptor computed by the analyzer
    * @return {function} f
    */
-  getBuilderFunction: function (descriptor, existingFormatters) {
+  getBuilderFunction : function (descriptor, existingFormatters) {
     var that = this;
     // declare an object which will contain all the code (by section) of the generated function
     var _code = {
-      init: '',
-      prev: '',
-      main: '',
-      add: function () {
+      init : '',
+      prev : '',
+      main : '',
+      add  : function () {
         var _lastIndex = arguments.length - 1;
         var _str = arguments[_lastIndex];
         for (var i = 0; i < _lastIndex; i++) {
@@ -392,14 +475,18 @@ var builder = {
         }
       }
     };
-    _code.add('prev', addIfNotExist.toString() + '\n' + removeFrom.toString() + '\n');
+    var _getSafeVar     = that.generateSafeJSVariable();
+    var _dictionaryName = '_dictionary';
+    var _safeAccessor   = that.generateSafeJSValueAccessor(_dictionaryName);
+    var _getSafeValue   = _safeAccessor.get;
+    var _getXMLStrIndex = _safeAccessor.getIndex;
+    _code.add('prev', addIfNotExist.toString()+'\n'+removeFrom.toString()+'\n');
     _code.add('init', "var _strResult = '';\n");
-    _code.add('init', 'var _root = (data !== null)?data:{};\n');
+    _code.add('init', 'var '+_getSafeVar('_root') + '= (data !== null)?data:{};\n');
     _code.add('init', 'var _strPart = {};\n');
     _code.add('init', 'var _strParts = [];\n');
     _code.add('init', 'var _xmlPos = [0];\n');
-    _code.add('init', 'var _formatterOptions = { lang : options.lang, currency : options.currency, stopPropagation : false, enum : options.enum, extension : options.extension };\n');
-    _code.add('init', 'var formatters = options.formatters;\n'); // used by getFormatterString
+    _code.add('init', 'var formatters = context.formatters;\n'); // used by getFormatterString
 
     var _atLeastOneSpecialIterator = false;
     // _str += "var _nestedArrayParams = [];\n";
@@ -415,35 +502,36 @@ var builder = {
     if (_staticData.before !== '') {
       _code.add('main', '_strPart = {\n');
       _code.add('main', "  'pos' : [0],\n");
-      _code.add('main', "  'str' : '" + _staticData.before + "'\n");
+      _code.add('main', "  'str' : '',\n");
+      _code.add('main', "  'bef' : "+_getXMLStrIndex(_staticData.before) + '\n');
       _code.add('main', '};\n');
       _code.add('main', '_strParts.push(_strPart);\n');
     }
 
     // no iterator template optimisation
     var _noIteratorTemplate = true;
-    for (var exitedArrayName in _dynamicData) {
-      if (_dynamicData[exitedArrayName].iterators !== undefined && _dynamicData[exitedArrayName].iterators.length > 0) {
+    for (var exitedArrayName in _dynamicData ) {
+      if (_dynamicData[exitedArrayName].iterators !== undefined && _dynamicData[exitedArrayName].iterators.length > 0 ) {
         _noIteratorTemplate = false;
         // TODO break;
       }
     }
-    if (_noIteratorTemplate) {
+    if ( _noIteratorTemplate ) {
       _code.add('main', 'var _registeredXml = {};\n');
     }
 
     // declare all variables
     for (var _objNameDeclaration in _dynamicData) {
       // the root "_" object is already declared above
-      if (_objNameDeclaration !== '_root') {
-        _code.add('init', 'var ' + _objNameDeclaration + ' = {};\n');
+      if (_objNameDeclaration !== '_root' ) {
+        _code.add('init', 'var '+_getSafeVar(_objNameDeclaration)+' = {};\n');
       }
     }
 
     // For each object, generate the code
     for (var _objIndex = 0; _objIndex < _hierarchy.length; _objIndex++) {
       var _objName = _hierarchy[_objIndex];
-      var _realObjName = _dynamicData[_objName].name;
+      var _realObjName = _getSafeValue(_dynamicData[_objName].name);
       var _type = _dynamicData[_objName].type;
       var _objParentName = _dynamicData[_objName].parent;
       var _objParentNames = _dynamicData[_objName].parents.reverse();
@@ -454,134 +542,150 @@ var builder = {
 
       // close nested for-loop brackets
       that.forEachArrayExit(_nestedArray, _dynamicData, _objName, function (exitedArrayName) {
-        _code.add('prev', 'main', '}\n');
+        _code.add('prev','main', '}\n');
         if (_dynamicData[exitedArrayName].iterators !== undefined) {
-          _nbCustomIterator -= _dynamicData[exitedArrayName].iterators.length - 1;
+          _nbCustomIterator -= _dynamicData[exitedArrayName].iterators.length-1;
         }
       });
 
       // Object type
       if (_type === 'object') {
         // declare any nested object
-        if (_objName !== '_root') {
-          _code.add('prev', 'main', _objName + '=' + '(' + _objParentName + ' !== undefined)?' + _objParentName + '[\'' + _realObjName + '\']:{};\n');
+        if (_objName!=='_root') {
+          _code.add('prev','main', _getSafeVar(_objName)+'='+'('+_getSafeVar(_objParentName)+' instanceof Object)?'+_getSafeVar(_objParentName)+'['+_realObjName+']:{};\n');
         }
       }
       else if (_type === 'objectInArray') {
-        var _arrayOfObjectName = _objName + '_array';
-        var _arrayOfObjectIndexName = _objName + '_i';
-        _code.add('prev', 'main', 'var ' + _arrayOfObjectName + '=' + '(' + _objParentName + ' !== undefined)? ' + _objParentName + '[\'' + _realObjName + '\']:[];\n');
+        var _arrayOfObjectNameG = _getSafeVar(_objName)+'_array';
+        var _arrayOfObjectIndexNameG = _getSafeVar(_objName)+'_i';
+        _code.add('prev','main', 'var ' + _arrayOfObjectNameG+'='+'('+_getSafeVar(_objParentName)+' instanceof Object)? '+_getSafeVar(_objParentName)+'['+_realObjName+']:[];\n');
         var _conditionToFindObject = _dynamicData[_objName].conditions || [];
-        var _objNameTemp = _objName + '_temp';
+        var _objNameTemp = _objName+'_temp';
 
-        _code.add('prev', 'main', 'var ' + _arrayOfObjectName + '_length = (' + _arrayOfObjectName + ' instanceof Array)?' + _arrayOfObjectName + '.length : 0;\n');
-        _code.add('prev', 'main', 'for (var ' + _arrayOfObjectIndexName + ' = 0; ' + _arrayOfObjectIndexName + ' < ' + _arrayOfObjectName + '_length ; ' + _arrayOfObjectIndexName + '++) {\n');
-        _code.add('prev', 'main', '  var ' + _objNameTemp + ' = ' + _arrayOfObjectName + '[' + _arrayOfObjectIndexName + '];\n');
-        _code.add('prev', 'main', that.getFilterString(_conditionToFindObject, _objName + '=' + _arrayOfObjectName + '[' + _arrayOfObjectIndexName + '];\n break;\n', _objNameTemp, false, _objNameTemp));
+        _code.add('prev', 'main', _getSafeVar(_objName) + '={};\n');
+        _code.add('prev', 'main', 'var '+_arrayOfObjectNameG+'_length = ('+_arrayOfObjectNameG+' instanceof Array)?'+_arrayOfObjectNameG+'.length : 0;\n');
+        _code.add('prev', 'main', 'for (var '+_arrayOfObjectIndexNameG+' = 0; '+_arrayOfObjectIndexNameG+' < '+_arrayOfObjectNameG+'_length ; '+_arrayOfObjectIndexNameG+'++) {\n');
+        _code.add('prev', 'main', '  var '+_getSafeVar(_objNameTemp)+' = '+_arrayOfObjectNameG+'['+_arrayOfObjectIndexNameG+'];\n');
+        _code.add('prev', 'main', that.getFilterString(_getSafeVar, _getSafeValue, _conditionToFindObject, _getSafeVar(_objName)+'='+_arrayOfObjectNameG+'['+_arrayOfObjectIndexNameG+'];\n break;\n', _objNameTemp, false, _objNameTemp));
         _code.add('prev', 'main', '}\n');
       }
       // Array type
       else if (_type === 'array') {
         var _posStart = _dynamicData[_objName].position.start;
         var _posEnd = _dynamicData[_objName].position.end;
-        var _arrayIndexName = _objName + '_i';
-        var _arrayName = _objName + '_array';
+        var _arrayIndexNameG = _getSafeVar(_objName)+'_i';
+        var _arrayNameG = _getSafeVar(_objName)+'_array';
         var _iterators = _dynamicData[_objName].iterators;
+        var _repeater = _dynamicData[_objName].repeater;
+        var _repeaterGlobalCounter = _getSafeVar(_objName)+'_repeatCount';
         var _containSpecialIterator = _dynamicData[_objName].containSpecialIterator;
 
 
 
-        _code.add('main', '_xmlPos[' + (_nbCustomIterator + _arrayDepth * 2 - 2) + '] = ' + _posStart + ';\n');
-        _code.add('main', '_xmlPos[' + (_nbCustomIterator + _arrayDepth * 2 - 1) + '] = ' + _posStart + ';\n');
+        _code.add('main', '_xmlPos['+(_nbCustomIterator+_arrayDepth*2-2)+'] = '+_posStart+';\n');
+        _code.add('main', '_xmlPos['+(_nbCustomIterator+_arrayDepth*2-1)+'] = '+_posStart+';\n');
 
         // declare any nested object
-        if (_objName !== '_root') {
-          _code.add('prev', 'main', 'var ' + _arrayName + '=' + '(' + _objParentName + ' instanceof Object) ?' + _objParentName + '[\'' + _realObjName + '\']:[];\n');
+        if (_objName!=='_root') {
+          _code.add('prev', 'main', 'var '+_arrayNameG+'='+'('+ _getSafeVar(_objParentName)+' instanceof Object) ?'+ _getSafeVar(_objParentName)+'['+_realObjName+']:[];\n');
         }
         else {
-          _code.add('prev', 'main', 'var ' + _arrayName + '=' + _objName + ';\n');
+          _code.add('prev', 'main', 'var '+_arrayNameG+'='+ _getSafeVar(_objName)+';\n');
         }
         // keep the end of the array
-        _lastArrayEnd = (_posEnd > _lastArrayEnd) ? _posEnd : _lastArrayEnd;
+        _lastArrayEnd = (_posEnd>_lastArrayEnd)? _posEnd : _lastArrayEnd;
 
         // add xml which are before the array repetition section
         if (_beforeStr !== undefined && _beforeStr !== '') {
           _code.add('main', '_strPart = {\n');
-          _code.add('main', "  'pos' :  _xmlPos.slice(0, " + (_nbCustomIterator + _arrayDepth * 2 - 1) + '),\n');
-          _code.add('main', "  'str' : '" + _beforeStr + "'\n");
+          _code.add('main', "  'pos' :  _xmlPos.slice(0, "+(_nbCustomIterator+_arrayDepth*2-1)+'),\n');
+          _code.add('main', "  'str' : '',\n");
+          _code.add('main', "  'bef' : "+_getXMLStrIndex(_beforeStr) + '\n');
           _code.add('main', '};\n');
           _code.add('main', '_strParts.push(_strPart);\n');
         }
         // store array name
         _nestedArray.push(_objName);
-        var _missingIteratorArrayName = _arrayName + '_missingIterators';
-        if (_containSpecialIterator === true) {
-          _missingIteratorArrayName = _arrayName + '_missingIterators';
-          _code.add('main', 'var ' + _missingIteratorArrayName + '=[];\n');
+        var _missingIteratorArrayNameG = _arrayNameG+'_missingIterators';
+        if (_containSpecialIterator===true) {
+          _missingIteratorArrayNameG = _arrayNameG+'_missingIterators';
+          _code.add('main'       , 'var '+_missingIteratorArrayNameG+'=[];\n');
         }
 
-        _code.add('prev', 'main', 'var ' + _arrayName + '_length = 0;\n');
-        _code.add('prev', 'main', 'if (' + _arrayName + ' instanceof Array) {\n');
-        _code.add('prev', 'main', _arrayName + '_length = ' + _arrayName + '.length;\n');
-        _code.add('prev', 'main', '} else if (' + _arrayName + ' instanceof Object) {\n');
-        _code.add('prev', 'main', _arrayName + ' = Object.entries(' + _arrayName + ').map((a) => { return {att : a[0], val : a[1] } });\n');
-        _code.add('prev', 'main', _arrayName + '_length = ' + _arrayName + '.length;\n');
-        _code.add('prev', 'main', '}\n');
+        _code.add('prev','main', 'var '+_arrayNameG+'_length = 0;\n');
+        _code.add('prev','main', 'if ('+_arrayNameG+' instanceof Array) {\n');
+        _code.add('prev','main', _arrayNameG+'_length = '+_arrayNameG+'.length;\n');
+        _code.add('prev','main', '} else if ('+_arrayNameG+' instanceof Object) {\n');
+        _code.add('prev','main', _arrayNameG+' = Object.entries('+_arrayNameG+').map((a) => { return {att : a[0], val : a[1] } });\n');
+        _code.add('prev','main', _arrayNameG+'_length = '+_arrayNameG+'.length;\n');
+        _code.add('prev','main', '}\n');
 
-        _code.add('prev', 'main', 'for (var ' + _arrayIndexName + ' = 0; ' + _arrayIndexName + ' < ' + _arrayName + '_length ');
-        if (_containSpecialIterator === true) {
-          _code.add('main', '  || ' + _missingIteratorArrayName + '.length>0');
+        if ( _repeater ) {
+          _code.add('prev','main', 'var '+_repeaterGlobalCounter+' = 0;\n');
+          _nestedArray.push(_objName);
         }
-        _code.add('prev', 'main', '  ; ' + _arrayIndexName + '++) {\n');
-        _code.add('prev', 'main', '  var ' + _objName + ' = ' + _arrayName + '[' + _arrayIndexName + '];\n');
+        _code.add('prev','main', 'for (var '+_arrayIndexNameG+' = 0; '+_arrayIndexNameG+' < '+_arrayNameG+'_length ');
+        if (_containSpecialIterator===true) {
+          _code.add(       'main', '  || '+_missingIteratorArrayNameG+'.length>0');
+        }
+        _code.add('prev','main', '  ; '+_arrayIndexNameG+'++) {\n');
+        _code.add('prev','main', '  var '+ _getSafeVar(_objName)+' = '+_arrayNameG+'['+_arrayIndexNameG+'];\n');
 
         // use a custom iterator?
         for (var it = 0; it < _iterators.length; it++) {
           var _iterator = _iterators[it];
-          if (it > 0) {
+          var _iteratorObjG = _getSafeValue(_iterator.obj);
+          var _iteratorAttrG = _getSafeValue(_iterator.attr);
+          if (it>0) {
             _nbCustomIterator++;
           }
           if (_iterator.attr === 'i') {
-            _code.add('main', '  _xmlPos[' + (_nbCustomIterator + _arrayDepth * 2 - 1) + '] = ' + _arrayIndexName + ';\n');
+            _code.add('main', '  _xmlPos['+(_nbCustomIterator+_arrayDepth*2-1)+'] = '+_arrayIndexNameG+';\n');
+            if ( _repeater ) {
+              var _repeaterAttr = _getSafeValue(_repeater);
+              var _repeaterName = _getSafeVar(_objName)+'_repeat';
+              _code.add('prev', 'main', ' let '+_repeaterName+' = '+_getSafeVar(_objName)+'['+_repeaterAttr+'];\n');
+              _code.add('prev', 'main', ' if ( '+_repeaterName+' > 200 ) { throw Error("The repeater cannot be above 200"); }\n');
+              _code.add('main', 'while( '+_repeaterName+'-- > 0) {\n');
+              _code.add('main', '  _xmlPos['+(_nbCustomIterator+_arrayDepth*2-1)+'] = '+_repeaterGlobalCounter+'++;\n');
+            }
           }
           else {
-            var _iteratorName = _objName + '_it';
-            var _iteratorObjName = _objName + '_itObj';
-            _code.add('prev', 'main', ' var ' + _iteratorName + ' = 0;\n'); // TODO return errors if the iterator is not defined
-
-            if (_containSpecialIterator === true) {
-              _code.add('main', 'if(' + _objName + ' !== undefined){\n');
+            var _iteratorNameG = _getSafeVar(_objName)+'_it';
+            var _iteratorObjNameG = _getSafeVar(_objName)+'_itObj';
+            _code.add('prev', 'main', ' var '+_iteratorNameG+' = 0;\n'); // TODO return errors if the iterator is not defined
+            if (_containSpecialIterator===true) {
+              _code.add('main', 'if('+ _getSafeVar(_objName) +' !== undefined){\n');
             }
             // the iterator is inside an object
             if (_iterator.obj !== undefined) {
-              _code.add('prev', 'main', ' var ' + _iteratorObjName + ' = ' + _objName + "['" + _iterator.obj + "'];\n");
-              _code.add('prev', 'main', ' if(' + _iteratorObjName + ' !== undefined) {\n');
-              _code.add('prev', 'main', '   ' + _iteratorName + ' = ' + _iteratorObjName + "['" + _iterator.attr + "'];\n");
-              _code.add('prev', 'main', ' }\n');
+              _code.add('prev','main', ' var '+_iteratorObjNameG+' = '+_getSafeVar(_objName)+'['+_iteratorObjG+'];\n');
+              _code.add('prev','main', ' if('+_iteratorObjNameG+' !== undefined) {\n');
+              _code.add('prev','main', '   '+_iteratorNameG+' = '+_iteratorObjNameG+'['+_iteratorAttrG+'];\n');
+              _code.add('prev','main', ' }\n');
             }
             else {
-              _code.add('prev', 'main', ' ' + _iteratorName + ' = ' + _objName + "['" + _iterator.attr + "'];\n");
+              _code.add('prev','main', ' '+_iteratorNameG+' = '+_getSafeVar(_objName)+'['+_iteratorAttrG+'];\n');
             }
-            if (_containSpecialIterator === true) {
-              _code.add('main', '  removeFrom(' + _missingIteratorArrayName + ', ' + _iteratorName + ');\n');
+            if (_containSpecialIterator===true) {
+              _code.add('main', '  removeFrom('+_missingIteratorArrayNameG+', '+_iteratorNameG+');\n');
               _code.add('main', '}else{\n');
-              _code.add('main', '  ' + _iteratorName + ' = ' + _missingIteratorArrayName + '.pop();\n');
+              _code.add('main', '  '+_iteratorNameG+' = '+_missingIteratorArrayNameG+'.pop();\n');
               _code.add('main', '}\n');
             }
-            _code.add('main', '  _xmlPos[' + (_nbCustomIterator + _arrayDepth * 2 - 1) + '] = ' + _iteratorName + ';\n');
+            _code.add('main', '  _xmlPos['+(_nbCustomIterator+_arrayDepth*2-1)+'] = '+_iteratorNameG+';\n');
             // count the number of custom iterator
-            // _str += "  _xmlPos["+(_nbCustomIterator+_arrayDepth*2-1)+"] = "+_arrayIndexName+";\n";
+            // _str += "  _xmlPos["+(_nbCustomIterator+_arrayDepth*2-1)+"] = "+_arrayIndexNameG+";\n";
             if (_iterator.isSpecial === true) {
               _atLeastOneSpecialIterator = true;
-              var _allIteratorArrayName = _iteratorName + '_wanted';
-              _code.main = _code.main.replace(_missingIteratorArrayName + '=[]', _missingIteratorArrayName + '=' + _allIteratorArrayName + '.slice()');
-              _code.add('init', 'var ' + _allIteratorArrayName + '=[];\n');
-              _code.add('prev', ' addIfNotExist(' + _allIteratorArrayName + ', ' + _iteratorName + ');\n');
+              var _allIteratorArrayNameG = _iteratorNameG + '_wanted';
+              _code.main = _code.main.replace(_missingIteratorArrayNameG+'=[]', _missingIteratorArrayNameG+'='+_allIteratorArrayNameG+'.slice()');
+              _code.add('init', 'var '+_allIteratorArrayNameG+'=[];\n');
+              _code.add('prev', ' addIfNotExist('+_allIteratorArrayNameG+', '+_iteratorNameG+');\n');
             }
           }
         }
       }
-
       // Generate code which will concatenate each xml parts
       for (var i = 0; i < _xmlParts.length; i++) {
         var _xmlPart = _xmlParts[i];
@@ -596,13 +700,13 @@ var builder = {
           _keepHighestPosition = _xmlPart.pos;
         }
         // create the xml part and set its absolute position in the xml
-        _code.add('main', '_xmlPos[' + (_nbCustomIterator + _partDepth * 2) + '] = ' + _xmlPart.pos + ';\n');
+        _code.add('main', '_xmlPos['+(_nbCustomIterator+_partDepth*2)+'] = '+_xmlPart.pos+';\n');
         _code.add('main', '_strPart = {\n');
-        _code.add('main', "  'pos' : _xmlPos.slice(0, " + (_nbCustomIterator + _partDepth * 2 + 1) + '),\n');
-        _code.add('main', "  'str'    : ''\n");
+        _code.add('main', "  'pos' : _xmlPos.slice(0, "+(_nbCustomIterator+_partDepth*2+1)+'),\n');
+        _code.add('main', "  'str' : ''\n");
         _code.add('main', '};\n');
         if (_xmlPart.before) {
-          _code.add('main', "_strPart.str = '" + _xmlPart.before + "';\n");
+          _code.add('main', '_strPart.bef = '+_getXMLStrIndex(_xmlPart.before) + ';\n');
         }
         if (_xmlPart.array === 'start') {
           _code.add('main', '_strPart.rowStart = true;\n');
@@ -614,75 +718,44 @@ var builder = {
         if (_dataAttr) {
           // handle conditions
           _code.add('main', '_strPart.rowShow = true;\n');
-          _code.add('main', that.getFilterString(_conditions, '_strPart.rowShow = false', _objName, true));
-          _code.add('main', 'if(' + _dataObj + ' !== undefined && ' + _dataObj + ' !== null){\n');
-          _code.add('main', 'var _str = ' + _dataObj + "['" + _dataAttr + "']" + ';\n');
+          _code.add('main', that.getFilterString(_getSafeVar, _getSafeValue, _conditions, '_strPart.rowShow = false', _getSafeVar(_objName), true));
+          _code.add('main', 'var _str = ' + _getSafeVar(_dataObj) + ' !== undefined &&  ' + _getSafeVar(_dataObj) + ' !== null ? ' + _getSafeVar(_dataObj) + '[' + _getSafeValue(_dataAttr) + ']' + ' : undefined ;\n');
           // TODO optimize avoid using all this options for all formatters
-          _code.add('main', '_formatterOptions.stopPropagation = false;\n');
-          _code.add('main', '_formatterOptions.isConditionTrue = null;\n');
-          _code.add('main', '_formatterOptions.isAndOperator = null;\n');
-          _code.add('main', '_formatterOptions.isHidden = null;\n');
-          _code.add('main', '_formatterOptions.parentsData = [' + _dataObj + ', ' + _objParentNames.join(',') + '];\n');
-          _code.add('main', that.getFormatterString('_str', '_formatterOptions', _formatters, existingFormatters, false));
+          _code.add('main', 'context.stopPropagation = false;\n');
+          _code.add('main', 'context.isConditionTrue = null;\n');
+          _code.add('main', 'context.isAndOperator = null;\n');
+          _code.add('main', 'context.isHidden = null;\n');
+          _code.add('main', 'context.parentsData = ['+_getSafeVar(_dataObj)+', '+_objParentNames.map(_getSafeVar).join(',')+'];\n');
+          _code.add('main', that.getFormatterString(_getSafeValue, '_str', 'context', _formatters, existingFormatters, false));
           // replace null or undefined value by an empty string
-          _code.add('main', 'if(_str === null || _str === undefined || /^undefined\:.+/.test(_str)) {\n');
+          _code.add('main', 'if(_str === null || _str === undefined) {\n');
           _code.add('main', "  _str = '';\n");
           _code.add('main', '};\n');
-          _code.add('main', 'if (_formatterOptions.isHidden !== null){\n');
-          _code.add('main', '  _strPart.hide = _formatterOptions.isHidden;\n');
+          _code.add('main', 'if (context.isHidden !== null){\n');
+          _code.add('main', '  _strPart.hide = context.isHidden;\n');
           _code.add('main', '}\n');
           // escape special characters for xml if the data is a string
           _code.add('main', "if(typeof(_str) === 'string') {\n");
           _code.add('main', "  _str = _str.replace(/&/g, '&amp;').replace(/</g, '&lt;').replace(/>/g, '&gt;').replace(/[\\u0000-\\u0008]|[\\u000B-\\u000C]|[\\u000E-\\u001F]/g, '');\n");
           _code.add('main', '};\n');
           // insert formatters which can inject XML, so after .replace(/</g, '&lt;') ... etc
-          _code.add('main', that.getFormatterString('_str', '_formatterOptions', _formatters, existingFormatters, true));
+          _code.add('main', that.getFormatterString(_getSafeValue, '_str', 'context', _formatters, existingFormatters, true));
 
-          _code.add('main', "_strPart.str += (_strPart.rowShow !== false)?_str:''" + ';\n');
-          _code.add('main', '}\n');
+          _code.add('main', "_strPart.str += (_strPart.rowShow !== false)?_str:''"+';\n');
         }
-        _code.add('main', `var posNumber = posNumber || 0;\n`);
         if (_xmlPart.after) {
-          _code.add('main', "_strPart.str += '" + _xmlPart.after + "';\n");
-          _code.add('main', `posNumber = +(_strPart.str.match(/r=".*?(\\d*)"/) || [])[1] || 0;\n`);
-          if (_arrayIndexName) {
-            _code.add('main', `_strPart.str = _strPart.str.replace(/(r=".*?)(\\d+)"/g, \`$1\${posNumber+${_arrayIndexName || 0}\}"\`);\n`);
-          }
+          _code.add('main', '_strPart.aft = ' + _getXMLStrIndex(_xmlPart.after) + ';\n');
         }
         // push the xml part in a array (this array will be sorted at the end)
         // add this if there is an iterator ( template is using i+1 form )
-        if (_noIteratorTemplate === false) {
-          if (_arrayIndexName) {
-            _code.add('main', `var formulas =  _strPart.str.match(/<f.*?>.*?<\\/f>/g); \n`);
-            _code.add('main', `if(formulas && formulas.length) { 
-              formulas.forEach(formula => {
-              var newFormula = formula;
-              var dataCells = formula.match(/((\\w+\\d+:\\w+\\d+)|(\\w+\\d+))/g);
-              dataCells.forEach(item => {
-                if(item.indexOf(':')!=-1) {
-                  var cells = item.split(':').map(cell => {
-                    var [group, col, row] = cell.match(/(\\w+)(\\d+)/); 
-                    return {value: cell, row , col }
-                  });
-                  if(cells[0].row == cells[1].row) {
-                    newFormula = newFormula.replace(item, item.replace(new RegExp(posNumber, 'g'), posNumber+${_arrayIndexName || 0}));
-                  } 
-                } else {
-                  newFormula = newFormula.replace(item, item.replace(posNumber, posNumber+${_arrayIndexName || 0}));
-                }
-              })
-              _strPart.str = _strPart.str.replace(formula, newFormula);
-            })
-             }\n`);
-            _code.add('main', `_strPart.str = _strPart.str.replace(/(r=".*?)(\\d+)"/g, \`$1\${posNumber+${_arrayIndexName || 0}\}"\`);\n`);
-          }
+        if ( _noIteratorTemplate === false ) {
           _code.add('main', '_strParts.push(_strPart);\n');
           // when no iterator only push when the part position is new or rowshow is true
         }
         else {
-          _code.add('main', 'if(_registeredXml[' + _xmlPart.pos + '] !== true || _strPart.rowShow === true ) {\n');
+          _code.add('main', 'if(_registeredXml['+_xmlPart.pos+'] !== true || _strPart.rowShow === true ) {\n');
           _code.add('main', '  _strParts.push(_strPart);\n');
-          _code.add('main', '  _registeredXml[' + _xmlPart.pos + '] = true;\n');
+          _code.add('main', '  _registeredXml['+_xmlPart.pos+'] = true;\n');
           _code.add('main', '};\n');
         }
       }
@@ -694,8 +767,9 @@ var builder = {
     }
     if (_staticData.after !== '') {
       _code.add('main', '_strPart = {\n');
-      _code.add('main', "  'pos' : [" + (_keepHighestPosition + 1) + '],\n');
-      _code.add('main', "  'str' : '" + _staticData.after + "'\n");
+      _code.add('main', "  'pos' : ["+(_keepHighestPosition+1)+'],\n');
+      _code.add('main', "  'str' : '',\n");
+      _code.add('main', "  'aft' : "+ _getXMLStrIndex(_staticData.after) + '\n');
       _code.add('main', '};\n');
       _code.add('main', '_strParts.push(_strPart);\n');
     }
@@ -709,10 +783,12 @@ var builder = {
     // The function is built, we compile it and check errors in the same time
     var _fn;
     try {
-      _fn = new Function('data', 'options', 'helper', _codeAssembled);
+      _fn = new Function('data', 'context', 'helper', _dictionaryName, _codeAssembled);
+      _fn.builderDictionary = _safeAccessor.getDictionary();
     }
     catch (err) {
-      throw new Error('getSubstitutionFunction: Impossible to generate the XML builder.\n' + err + '\n--------------------------------\n' + _codeAssembled + '\n--------------------------------\n');
+      console.log('Execution plan error; ' + err + '\n' + _codeAssembled);
+      throw new Error('Impossible to prepare execution plan. Please, contact the support');
     }
     return _fn;
   }
@@ -732,7 +808,7 @@ module.exports = builder;
  * @param  {Array}  arrB
  * @return {Boolean}  true if the two arrays are equals, false otherwise
  */
-function isArrayEqual(arrA, arrB) {
+function isArrayEqual (arrA, arrB) {
   var _arrALength = arrA.length;
   var _arrBLength = arrB.length;
   if (_arrALength !== _arrBLength) {
@@ -751,7 +827,7 @@ function isArrayEqual(arrA, arrB) {
  * @param {Array} arr   array
  * @param {Mixed} value value
  */
-function addIfNotExist(arr, value) {
+function addIfNotExist (arr, value) {
   for (var i = 0; i < arr.length; i++) {
     var _value = arr[i];
     if (_value === value) {
@@ -767,9 +843,9 @@ function addIfNotExist(arr, value) {
  * @param  {Mixed} value value
  * @return {Array}       array modified
  */
-function removeFrom(arr, value) {
+function removeFrom (arr, value) {
   for (var i = 0; i < arr.length; i++) {
-    if (arr[i] === value) {
+    if (arr[i]===value) {
       arr.splice(i, 1);
       return arr;
     }
