@@ -1,11 +1,12 @@
 var path  = require('path');
 var fs = require('fs');
+var os = require('os');
 var helper = require('./helper');
 var spawn = require('child_process').spawn;
 var params = require('./params');
 var debug = require('debug')('carbone:converter');
+var which = require('which');
 
-var pythonFile = path.join(__dirname, 'converter.py');
 
 /* Factories object */
 var conversionFactory = {};
@@ -27,6 +28,8 @@ var converterOptions = {
   delayBeforeKill : 500
 };
 
+/* get the total memory available on the system (unit: MB) */
+const totalMemoryAvailableMB = os.totalmem() / 1024 / 1024;
 
 var pythonErrors = {
   1   : 'Global error',
@@ -34,7 +37,6 @@ var pythonErrors = {
   400 : 'Could not open document',
   401 : 'Could not convert document'
 };
-
 
 var converter = {
 
@@ -98,13 +100,14 @@ var converter = {
       var _factory = conversionFactory[i];
       // if a factory is running
       if (_factory && (_factory.pythonThread !== null || _factory.officeThread !== null)) {
+        clearTimeout(_factory.timeoutId);
         _factory.exitCallback = factoryExitFn;
         // kill Python thread first.
         if (_factory.pythonThread !== null) {
-          _factory.pythonThread.kill();
+          _factory.pythonThread.kill('SIGKILL');
         }
         if (_factory.officeThread !== null) {
-          _factory.officeThread.kill();
+          _factory.officeThread.kill('SIGKILL');
           helper.rmDirRecursive(_factory.userCachePath);
         }
       }
@@ -129,29 +132,46 @@ var converter = {
    * Convert a document
    *
    * @param {string} inputFile : absolute path to the source document
-   * @param {string} outputType : destination type among all types returned by getTypes
-   * @param {string} formatOptions : options passed to convert
-   * @param {function} callback : function(buffer) result
+   * @param {string} outputType : destination type of format.js (ex. writer_pdf_Export for PDF)
+   * @param {string} formatOptions : options string passed to convert
+   * @param {string} outputFile : outputFile to generate
+   * @param {function} callback(err, outputFile)
    */
-  convertFile : function (inputFile, outputType, formatOptions, callback, returnLink, outputFile) {
+  convertFile : function (inputFile, outputType, formatOptions, outputFile, callback) {
 
     if (isLibreOfficeFound === false) {
       return callback('Cannot find LibreOffice. Document conversion cannot be used');
     }
 
-    var _output = helper.getUID();
     var _job = {
       inputFilePath  : inputFile,
-      outputFilePath : outputFile || path.join(params.tempPath, _output),
+      outputFilePath : outputFile,
       outputFormat   : outputType,
       formatOptions  : formatOptions || '',
-      returnLink     : returnLink || false,
       callback       : callback,
       nbAttempt      : 0,
       error          : null
     };
     jobQueue.push(_job);
     executeQueue();
+  },
+
+  /**
+   * Do we need to restart LibreOffice?
+   *
+   * Temporal fix for memory leaks of LibreOffice 6+
+   *
+   * @param  {Objecct} params
+   * @param  {Integer} availableMemory system available memory
+   * @param  {Integer} nbReports       nb reborts computed by one factory
+   * @return {Boolean}                 true if LibreOffice must be restarted, false otherwise
+   */
+  shouldTheFactoryBeRestarted : function (params, availableMemory, nbReports) {
+    const _percentageFactoryMemoryLoaded = nbReports * params.factoryMemoryFileSize * 100 / availableMemory;
+    if (_percentageFactoryMemoryLoaded < params.factoryMemoryThreshold || params.factoryMemoryThreshold === 0) {
+      return false;
+    }
+    return true;
   }
 
 };
@@ -164,9 +184,13 @@ var converter = {
 
 /**
  * Add a LibreOffice + Python factory (= 2 threads)
+ *
+ * WARNING: the callback must be used only by converter.init()
+ *
  * @param {function} callback : function() called when the factory is ready to convert documents.
  */
 function addConversionFactory (callback) {
+  debug('ask to add a conversion factory');
   // find a free factory
   var _prevFactory = {};
   var _startListenerID = -1;
@@ -183,30 +207,45 @@ function addConversionFactory (callback) {
   }
   // maximum of factories reached
   if (_startListenerID === -1) {
-    return callback();
+    if (callback) {
+      callback();
+    }
+    return;
   }
   var _uniqueName = helper.getUID();
 
   // generate a unique path to a fake user profile. We cannot start multiple instances of LibreOffice if it uses the same user cache
   var _userCachePath = path.join(params.tempPath, '_office_' + _uniqueName);
-  if (_prevFactory && _prevFactory.userCachePath !== undefined) {
+  if (_prevFactory) {
     // re-use previous directory if possible (faster restart)
-    _userCachePath = _prevFactory.userCachePath;
+    if (_prevFactory.userCachePath !== undefined) {
+      _userCachePath = _prevFactory.userCachePath;
+    }
+    // If soffice crashes as soon as it was started, the callback of the previous starting process must be passed to the new started factory
+    // On Linux, it happens when LibreOffice creates its directory for the first time (oosplash seems to hide this)
+    if (_prevFactory.readyCallback) {
+      callback = _prevFactory.readyCallback;
+    }
   }
+  // generate a URL in LibreOffice's format so that it's portable across OSes:
+  // see: https://wiki.openoffice.org/wiki/URL_Basics
+  var _userCacheURL = convertToURL(_userCachePath);
 
   // generate a unique pipe name
   var _pipeName = params.pipeNamePrefix + '_' +_uniqueName;
   var _connectionString = 'pipe,name=' + _pipeName + ';urp;StarOffice.ComponentContext';
   var _officeParams = ['--headless', '--invisible', '--nocrashreport', '--nodefault', '--nologo', '--nofirststartwizard', '--norestore',
-    '--quickstart', '--nolockcheck', '--accept='+_connectionString, '-env:UserInstallation=file://'+_userCachePath ];
+    '--quickstart', '--nolockcheck', '--accept='+_connectionString, '-env:UserInstallation='+_userCacheURL ];
 
   // save unique name
   activeFactories.push(_pipeName);
 
   var _officeThread = spawn(converterOptions.sofficeExecPath, _officeParams);
   _officeThread.on('close', generateOnExitCallback(_startListenerID, false, _pipeName));
+  debug('office thread started with PID ' + _officeThread.pid);
 
-  var _pythonThread = spawn(converterOptions.pythonExecPath, [pythonFile, '--pipe', _pipeName]);
+  var _pythonThread = spawn(converterOptions.pythonExecPath, [params.pythonPath, '--pipe', _pipeName]);
+  debug('python thread started with PID ' + _pythonThread.pid);
   _pythonThread.on('close', generateOnExitCallback(_startListenerID, true, _pipeName));
   _pythonThread.stdout.on('data', generateOnDataCallback(_startListenerID));
   _pythonThread.stderr.on('data', function (err) {
@@ -223,7 +262,9 @@ function addConversionFactory (callback) {
       pythonThread  : _pythonThread,
       isReady       : false,
       isConverting  : false,
-      readyCallback : callback
+      readyCallback : callback,
+      nbrReports    : 0,
+      timeoutId     : null
     };
     conversionFactory[_startListenerID] = _factory;
   }
@@ -232,6 +273,26 @@ function addConversionFactory (callback) {
   }
 }
 
+/**
+ * Kill one LibreOffice factory
+ *
+ * @param  {Object} factory
+ */
+function killFactory (factory) {
+  if (factory.isReady === false) {
+    return;
+  }
+  factory.isReady      = false;
+  factory.isConverting = false;
+  factory.nbrReports   = 0;
+  clearTimeout(factory.timeoutId);
+  if (factory.officeThread !== null) {
+    factory.officeThread.kill('SIGKILL');
+  }
+  else if (factory.pythonThread !== null) {
+    factory.pythonThread.kill('SIGKILL');
+  }
+}
 
 /**
  * Generate a callback which is used to handle thread error and exit
@@ -254,6 +315,7 @@ function generateOnExitCallback (factoryID, isPythonProcess, factoryUniqueName) 
     // the factory cannot receive jobs anymore
     _factory.isReady = false;
     _factory.isConverting = false;
+    clearTimeout(_factory.timeoutId);
 
     // if the Python process died...
     if (isPythonProcess === true) {
@@ -267,7 +329,7 @@ function generateOnExitCallback (factoryID, isPythonProcess, factoryUniqueName) 
       _otherThreadToKill = _factory.pythonThread;
     }
 
-    debug('process '+_processName+' of factory '+factoryID+' died ' + error);
+    debug('process '+_processName+' (PID ' +_factory.pid+ ') of factory '+factoryID+' died ' + error);
 
     // if both processes Python and Office are off...
     if (_factory.pythonThread === null && _factory.officeThread === null) {
@@ -278,6 +340,20 @@ function generateOnExitCallback (factoryID, isPythonProcess, factoryUniqueName) 
     }
     else {
       _otherThreadToKill.kill('SIGKILL');
+      // Fixes #12
+      // SIGKILL to make sure everything is off
+      //
+      // Be careful, LibreOffice has two threads oosplash (parent) -> soffice (child) if launched with "soffice".
+      // On Linux, we decided to launch the child process directly to simplify the thread management.
+      // Otherwise, only oosplash is killed if SIGKILL is sent. In that case:
+      //   - The child "soffice" is still alive and stdin, stdout and stderr of are not closed automatically.
+      //   - The event "spawn.close" is received only if stdin, stdout and stderr are closed. So carbone hangs indefinitely :(
+      // When killing the oospash parent process, we should close stdin, stdout and stderr and kill the child thread soffice ourself (like "pkill soffice")
+      //
+      // Also, we could use SIGTERM. In that case, oosplash (parent) sends the signal to its child... but this signal is not powerful enough to 
+      // guarantee a shutdown. If LibreOffice hangs, we could wait forever.
+      //
+      // It is easier to only launch directly soffice.bin directly on Linux (see below)
     }
   };
 }
@@ -290,24 +366,17 @@ function generateOnExitCallback (factoryID, isPythonProcess, factoryUniqueName) 
 function whenFactoryIsCompletelyOff (factory) {
   // if Carbone is not shutting down
   if (isAutoRestartActive === true) {
-    // if there is an error while converting a document, let's try another time
-    var _job = factory.currentJob;
-    if (_job) {
-      if (_job.nbAttempt < params.attempts) {
-        factory.currentJob = null;
-        jobQueue.push(_job);
-      }
-      else {
-        _job.callback('Could not convert the document', null);
-      }
+    if (factory.currentJob) {
+      // if there is an error while converting a document, let's try another time
+      factory.currentJob.error = new Error('Could not convert the document');
     }
+    onCurrentJobEnd(factory);
     // avoid restarting too early
-    setTimeout(function () {
-      addConversionFactory(executeQueue);
-    },50);
+    setTimeout(addConversionFactory, 50);
   }
   // else if Carbone is shutting down and there is an exitCallback
   else {
+    // TODO delete async
     // delete office files synchronously (we do not care because Carbone is shutting down) when office is dead
     helper.rmDirRecursive(factory.userCachePath);
     if (factory.exitCallback) {
@@ -327,48 +396,44 @@ function generateOnDataCallback (factoryID) {
   return function (data) {
     var _factory = conversionFactory[factoryID];
     data = data.toString();
-    if (data !== '204') { // Document converted
-      var _job = _factory.currentJob;
-      _factory.currentJob = null;
-      _factory.isConverting = false;
-      if (_job) {
-        _job.error = (pythonErrors[data] !== undefined) ? pythonErrors[data]  : null;
-        if (_job.returnLink===true || _job.error !== null) {
-          if (_job.error && _job.nbAttempt < params.attempts) {
-            debug(_job.error + ': attempt '+_job.nbAttempt, _job);
-            _job.error = null;
-            jobQueue.push(_job);
-          }
-          else {
-            _job.callback(_job.error);
-          }
-          executeQueue();
-        }
-        else {
-          fs.readFile(_job.outputFilePath, function (err, content) {
-            _job.callback(_job.error, content);
-            fs.unlink(_job.outputFilePath, function (err) {
-              if (err) {
-                debug('cannot remove file' + _job.outputFilePath);
-              }
-            });
-            executeQueue();
-          });
-        }
-      }
-    }
-    else if (data === '204') { // Ready to receive conversion
+    // Ready to receive document conversion
+    if (data === '204') {
       debug('factory '+factoryID+' ready');
       _factory.isReady = true;
       if (_factory.readyCallback) {
         _factory.readyCallback();
+        // void readyCallback to avoid calling it twice when the factory object is re-used.
+        _factory.readyCallback = null;
       }
-      executeQueue();
+      return executeQueue();
     }
-    else {
-      debug('it should not print this message');
+    // Document converted with or without errors
+    if (_factory.currentJob) {
+      _factory.currentJob.error = (pythonErrors[data] !== undefined) ? new Error(pythonErrors[data]) : null;
     }
+    onCurrentJobEnd(_factory);
   };
+}
+
+/**
+ * Called when the job is finished
+ *
+ * @param  {Object} factory factory object
+ */
+function onCurrentJobEnd (factory) {
+  var _job = factory.currentJob;
+  factory.currentJob = null;
+  factory.isConverting = false;
+  clearTimeout(factory.timeoutId);
+  if (_job && _job.callback instanceof Function) {
+    // save the number of report converted to check the memory level of the LO process
+    // if it reach a threshold, the LO process is killed
+    if (converter.shouldTheFactoryBeRestarted(params, totalMemoryAvailableMB, ++factory.nbrReports) === true) {
+      killFactory(factory);
+    }
+    _job.callback(_job.error, _job.outputFilePath);
+  }
+  executeQueue();
 }
 
 
@@ -382,7 +447,7 @@ function executeQueue () {
   }
   // if there is no active factories, start them
   if (activeFactories.length < params.factories) {
-    addConversionFactory(executeQueue);
+    addConversionFactory();
     return;
   }
   for (var i in conversionFactory) {
@@ -408,43 +473,180 @@ function sendToFactory (factory, job) {
   factory.pythonThread.stdin.write('--format="'+job.outputFormat+'" --input="'+job.inputFilePath+'" --output="'+job.outputFilePath+'" --formatOptions="'+job.formatOptions+'"\n');
   // keep the number of attempts to convert this file
   job.nbAttempt++;
+  // Timeout to kill long conversions
+  if (params.converterFactoryTimeout > 0) {
+    clearTimeout(factory.timeoutId); // by security
+    factory.timeoutId = setTimeout(function () {
+      job.nbAttempt = params.attempts; // do not retry
+      job.error = new Error('Document conversion timeout reached ('+params.converterFactoryTimeout+' ms)');
+      killFactory(factory);
+      onCurrentJobEnd(factory);
+    }, params.converterFactoryTimeout);
+  }
+}
+
+
+/**
+ * Error for path
+ *
+ * @param {[type]} message [description]
+ */
+function PathError (message) {
+  this.name = 'PathError';
+  this.code = 'PathError';
+  this.message = message || 'Failed to convert path';
+  if (typeof Error.captureStackTrace === 'function') {
+    Error.captureStackTrace(this, PathError);
+  }
+}
+PathError.prototype = new Error();
+
+/**
+ * Convert an absolute path to an absolute URL understood by LibreOffice and
+ *  OpenOffice. This is necessary because LO/OO use a cross-platform path format
+ *  that does not match paths understood natively by OSes.
+ * If the input is already a URL, it is returned as-is.
+ *
+ * @param {string} inputPath - An absolute path to convert to a URL.
+ * @returns {string} A string suitable for use with LibreOffice as an absolute file path URL.
+ */
+function convertToURL (inputPath) {
+  // Guard clause: if it already looks like a URL, keep it that way.
+  if (inputPath.slice(0, 8) === 'file:///') {
+    return inputPath;
+  }
+  if (!path.isAbsolute(inputPath)) {
+    throw new PathError('Paths to convert must be absolute');
+  }
+  // Split into parts so that we can join into a URL:
+  var _normalizedPath = path.normalize(inputPath);
+  // (Use both delimiters blindly - we're aiming for maximum compatibility)
+  var _pathComponents = _normalizedPath.split(/[\\/]/);
+  // Make sure there is no leading empty element, since we always add a leading "/" anyway.
+  if (_pathComponents[0] === '') {
+    _pathComponents.shift();
+  }
+  var outputURL = 'file:///' + _pathComponents.join('/');
+  return outputURL;
 }
 
 
 /**
  * Detect If LibreOffice and python are available at startup
  */
-function detectLibreOffice () {
+function detectLibreOffice (additionalPaths) {
+  function _findBundledPython (sofficePath, pythonName) {
+    if (!sofficePath) {
+      return null;
+    }
+    // Try finding a Python binary shipped alongside the soffice binary,
+    //  either in its actual directory, or - if it's a symbolic link -
+    //  in the directory it points to.
+    var _sofficeActualDirectory;
+    var _symlinkDestination;
+    try {
+      _symlinkDestination = path.resolve(path.dirname(sofficePath), fs.readlinkSync(sofficePath));
+      // Assume symbolic link, will throw in case it's not:
+      _sofficeActualDirectory = path.dirname(_symlinkDestination);
+    }
+    catch (errorToIgnore) {
+      // Not a symlink.
+      _sofficeActualDirectory = path.dirname(sofficePath);
+    }
+    // Check for the Python binary in the actual soffice path:
+    try {
+      return which.sync(pythonName, { path : _sofficeActualDirectory });
+    }
+    catch (errorToIgnore) {
+      // No bundled Python found.
+      return null;
+    }
+  }
+
+  function _findBinaries (paths, pythonName, sofficeName) {
+    var _whichPython;
+    var _whichSoffice;
+    // Look for the soffice binary - first in the well-known paths, then in
+    //  the system PATH. On Linux, this prioritizes "upstream" (TDF) packages
+    //  over distro-provided ones from the OS' repository.
+    _whichSoffice = which.sync(sofficeName, { path : paths.join(':'), nothrow : true }) || which.sync(sofficeName, { nothrow : true }) || null;
+    // Check for a Python binary bundled with soffice, fall back to system-wide:
+    // This is a bit more complex, since we deal with some corner cases.
+    // 1. Hopefully use the python from the original soffice package, same dir
+    //  (this might fail on Mac if python is not in MacOS/, but in Resources/).
+    // 1a. Corner case: on Linux, if soffice was in /usr/bin/soffice and NOT
+    //  a symlink, then we would hit /usr/bin/python, which is probably python2.
+    //  This is why we try with python3 first, to defend against this.
+    // 2. Try finding it in any of the well-known paths - this might result in
+    //  using Python from *another install* of LibreOffice, but it should be ok.
+    //  This is only attempted if the paths exist on this system to avoid
+    //  a fallback to system PATH that "which" does when passed an empty string.
+    // 3. Fall back to system python (hopefully named python3).
+    _whichPython = _findBundledPython(_whichSoffice, 'python3') ||
+                  _findBundledPython(_whichSoffice, 'python') ||
+                  (paths.length > 0 && which.sync('python3', { path : paths.join(':'), nothrow : true })) ||
+                  (paths.length > 0 && which.sync('python', { path : paths.join(':'), nothrow : true })) ||
+                  which.sync('python3', { nothrow : true }) ||
+                  which.sync('python', { nothrow : true }) || null;
+    return {
+      soffice : _whichSoffice,
+      python  : _whichPython
+    };
+  }
+
+  function _listProgramDirectories (basePath, pattern) {
+    try {
+      return fs.readdirSync(basePath).filter(function _isLibreOfficeDirectory (dirname) {
+        return pattern.test(dirname);
+      }).map(function _buildFullProgramPath (dirname) {
+        return path.join(basePath, dirname, 'program');
+      });
+    }
+    catch (errorToIgnore) {
+      return [];
+    }
+  }
+
+  var _pathsToCheck = additionalPaths || [];
+  // overridable file names to look for in the checked paths:
+  var _pythonName = 'python';
+  var _sofficeName = 'soffice';
+  var _linuxDirnamePattern = /^libreoffice\d+\.\d+$/;
+  var _windowsDirnamePattern = /^LibreOffice( \d+(?:\.\d+)*?)?$/i;
 
   if (process.platform === 'darwin') {
-    // it is better to use the python bundled with LibreOffice
-    var _path = 'MacOS';
-    if (fs.existsSync('/Applications/LibreOffice.app/Contents/' + _path + '/python') === false) {
-      _path = 'Resources';
-    }
-
-    if (fs.existsSync('/Applications/LibreOffice.app/Contents/' + _path + '/python') === true) {
-      isLibreOfficeFound = true;
-      converterOptions.pythonExecPath = '/Applications/LibreOffice.app/Contents/' + _path + '/python';
-      converterOptions.sofficeExecPath = '/Applications/LibreOffice.app/Contents/MacOS/soffice';
-    }
+    _pathsToCheck = _pathsToCheck.concat([
+      // It is better to use the python bundled with LibreOffice:
+      '/Applications/LibreOffice.app/Contents/MacOS',
+      '/Applications/LibreOffice.app/Contents/Resources'
+    ]);
   }
   else if (process.platform === 'linux') {
-    var directories = fs.readdirSync('/opt');
-    var patternLibreOffice = /libreoffice\d+\.\d+/;
-    for (var i = 0; i < directories.length; i++) {
-      if (patternLibreOffice.test(directories[i]) === true) {
-        var directoryLibreOffice = '/opt/' + directories[i] + '/program';
-        if (fs.existsSync(directoryLibreOffice + '/python') === true) {
-          converterOptions.pythonExecPath = directoryLibreOffice + '/python';
-          converterOptions.sofficeExecPath = directoryLibreOffice + '/soffice';
-          isLibreOfficeFound = true;
-        }
-      }
-    }
+    // on Linux, avoid oosplash parent process to simplify SIGKILL propagation. Launch directly soffice.bin.
+    // Fixes #12
+    _sofficeName = 'soffice.bin';
+    // The Document Foundation packages (.debs, at least) install to /opt,
+    // into a directory named after the contained LibreOffice version.
+    // Add any existing directories that match this to the list.
+    _pathsToCheck = _pathsToCheck.concat(_listProgramDirectories('/opt', _linuxDirnamePattern));
+  }
+  else if (process.platform === 'win32') {
+    _pathsToCheck = _pathsToCheck
+      .concat(_listProgramDirectories('C:\\Program Files', _windowsDirnamePattern))
+      .concat(_listProgramDirectories('C:\\Program Files (x86)', _windowsDirnamePattern));
+    _pythonName = 'python.exe';
   }
   else {
-    debug('your platform is not supported yet');
+    debug('your platform "%s" is not supported yet', process.platform);
+  }
+
+  // Common logic for all OSes: perform the search and save results as options:
+  var _foundPaths = _findBinaries(_pathsToCheck, _pythonName, _sofficeName);
+  if (_foundPaths.soffice) {
+    debug('LibreOffice found: soffice at %s, python at %s', _foundPaths.soffice, _foundPaths.python);
+    isLibreOfficeFound = true;
+    converterOptions.pythonExecPath = _foundPaths.python;
+    converterOptions.sofficeExecPath = _foundPaths.soffice;
   }
 
   if (isLibreOfficeFound === false) {
@@ -452,8 +654,14 @@ function detectLibreOffice () {
   }
 }
 
-
 detectLibreOffice();
+
+['SIGINT', 'SIGHUP', 'SIGQUIT'].forEach(function (signal) {
+  process.on(signal, function () {
+    converter.exit();
+  });
+});
+
 
 process.on('exit', function () {
   converter.exit();
