@@ -9,7 +9,7 @@ var extracter = {
    * @param {array} markers : array of markers with their position
    * @return {object} descriptor
    */
-  splitMarkers: function (markers) {
+  splitMarkers : function (markers) {
     var _res = {};
     var _visitedArray = {};
     for (var i = 0; i < markers.length; i++) {
@@ -25,6 +25,7 @@ var extracter = {
       var _arrayName = '';
       var _iterators = [];
       var _prevChar = '';
+      var _repeater = '';
       var _isFormatter = false;
       var _isFormatterParenthesis = false;
       var _isString = false;
@@ -44,7 +45,7 @@ var extracter = {
         }
 
         // if we enter in a formatter and we are not between two parenthesis of a formatter (the character ":" is allowed within the parenthesis)
-        if (_char === ':' && _isFormatterParenthesis === false) {
+        if (_char===':' && _isFormatterParenthesis === false) {
           _isFormatter = true;
           // if we enter in a second formatter, keep the previous one. Ex. three nested formaters : formater1:formater2:formater3
           if (_formatterStr !== '') {
@@ -53,23 +54,23 @@ var extracter = {
           }
         }
         // if we are in a formatter, and if we enter in a parenthesis
-        else if (_isFormatter === true && _char === '(') {
+        else if (_isFormatter === true && _char==='(') {
           _formatterStr += _char;
           _isFormatterParenthesis = true;
         }
         // if we are in a formatter and we exit
-        else if (_isFormatter === true && _char === ')') {
+        else if (_isFormatter === true && _char===')') {
           _formatterStr += _char;
           _isFormatterParenthesis = false;
           _isFormatter = false;
           _formaters.push(_formatterStr);
           _formatterStr = '';
         }
-        else if (_isFormatter === true && (_char !== ' ' || _isString === true)) {
+        else if (_isFormatter === true && (_char!==' ' || _isString===true) ) {
           _formatterStr += _char;
         }
         // if we enter into the brakets of an array
-        else if (_char === '[') {
+        else if (_char==='[') {
           _isIteratorUsedInCurrentArray = false;
           _isArray = true;
           _uniqueMarkerName += _word;
@@ -90,36 +91,36 @@ var extracter = {
             _objectNameWhereStartMultiplePoints = _prevMarker;
           }
           if (_parents.length === 0) {
-            throw Error('Cannot access parent object in "' + _marker + '" (too high)');
+            throw Error('Cannot access parent object in "'+_marker+'" (too high)');
           }
-          if (_uniqueMarkerName === _parents[_parents.length - 1]) {
+          if (_uniqueMarkerName === _parents[_parents.length-1]) {
             // special case when coming from an array
             _parents.pop();
           }
           _uniqueMarkerName = _parents.pop();
           _prevMarker = _uniqueMarkerName;
           if (_parents.length > 0) {
-            _prevMarker = _parents[_parents.length - 1];
+            _prevMarker = _parents[_parents.length-1];
           }
         }
         // if we enter into an object
-        else if (_char === '.' && _isArray === false) {
+        else if (_char ==='.' && _isArray === false) {
           if (_objectNameWhereStartMultiplePoints !== '') {
             // when coming from .., set _prevMarker as usual
             _prevMarker = _uniqueMarkerName;
           }
           _uniqueMarkerName += _word;
           _uniqueMarkerName = this._cleanUniqueMarkerName(_uniqueMarkerName);
-          if (_prevMarker !== '' && _parents[_parents.length - 1] !== _prevMarker) {
+          if (_prevMarker !== '' && _parents[_parents.length-1] !== _prevMarker) {
             _parents.push(_prevMarker);
           }
           if (!_res[_uniqueMarkerName]) {
-            _res[_uniqueMarkerName] = {
-              name: _word,
-              type: 'object',
-              parent: _prevMarker,
-              parents: _parents.slice(0),
-              xmlParts: []
+            _res[_uniqueMarkerName]={
+              name     : _word,
+              type     : 'object',
+              parent   : _prevMarker,
+              parents  : _parents.slice(0),
+              xmlParts : []
             };
           }
           _prevMarker = _uniqueMarkerName;
@@ -127,40 +128,49 @@ var extracter = {
           _isFormatter = false;
         }
         // if this is not a whitespace, keep the word
-        else if (_char !== ' ' || _isString === true) {
+        else if (_char!==' ' || _isString===true) {
           _word += _char;
         }
 
         if (_isArray === true) {
-          // detect array condition
-          var _result = /([^,!]+)(!=|=|>|<)([\s\S]+)[,\]]$/.exec(_word);
+          // detect array condition, or direct access in arrays such as myAarray[10]
+          var _result = /([^,!]+)(!=|=|>|<)([\s\S]+)[,\]]$/.exec(_word) || /(\d+)\]$/.exec(_word);
+          // detect direct
           if (_result instanceof Array) {
+            if (_result.length === 2) {
+              // if this is a direct access such as "myArray[10]", adapt the result as if it was "myArray[i=10]"
+              _result = [0, 'i', '=', _result[1]];
+            }
             var _operator = (_result[2] === '=') ? '==' : _result[2];
             var _conditionObj = {
-              left: { parent: _uniqueMarkerName, attr: _result[1] },
-              operator: _operator,
-              right: _result[3]
+              left     : {parent : _uniqueMarkerName, attr : _result[1]},
+              operator : _operator,
+              right    : _result[3]
             };
             _conditions.push(_conditionObj);
             _word = '';
           }
-          else if (_word.length > 1 && (_char === ',' || _char === ']')) {
+          else if (_word.length > 1 && (_char===',' || _char===']')) {
             _isIteratorUsedInCurrentArray = true; // +1 or not
+            if (_word[0] === '*') {
+              _repeater = _word.slice(1, -1);
+              _word = _char;
+            }
           }
           // detect array iteration
-          if (_word.substr(-2) === '+1') {
-            var _iteratorNormal = _word.slice(0, -2).replace(',', '');
-            _iterators.push({ str: _iteratorNormal });
+          if (_word.substr(-2)==='+1') {
+            var _iteratorNormal = _word.slice(0,-2).replace(',','');
+            _iterators.push({str : _iteratorNormal});
             _word = '';
           }
-          else if (_word.substr(-2) === '++') {
-            var _iteratorCustom = _word.slice(0, -2).replace(',', '');
-            _iterators.push({ str: _iteratorCustom, isSpecial: true });
+          else if (_word.substr(-2)==='++') {
+            var _iteratorCustom = _word.slice(0,-2).replace(',','');
+            _iterators.push({str : _iteratorCustom, isSpecial : true});
             _word = '';
           }
         }
         // if we exit the bracket of an array
-        if (_char === ']' && _isArray === true) {
+        if (_char===']' && _isArray===true) {
 
           // If it was the first part of the array (i without "i+1")
           if (_iterators.length === 0) {
@@ -174,23 +184,23 @@ var extracter = {
               // correction of previously detected conditions
               for (var k = 0; k < _conditions.length; k++) {
                 var _condition = _conditions[k];
-                if (_condition.left.parent === _uniqueMarkerName) {
+                if (_condition.left.parent===_uniqueMarkerName) {
                   _condition.left.parent += '$';
                 }
               }
               _uniqueMarkerName += '$';
             }
-            if (_prevMarker !== '' && _parents[_parents.length - 1] !== _prevMarker) {
+            if (_prevMarker !== '' && _parents[_parents.length-1] !== _prevMarker) {
               _parents.push(_prevMarker);
             }
             if (_isIteratorUsedInCurrentArray === false) {
               var _filterUniqueName = '';
               var h = _conditions.length - 1;
-              for (; h >= 0; h--) {
+              for (; h >= 0 ; h--) {
                 var _conditionToFindObject = _conditions[h];
                 _filterUniqueName += helper.cleanJavascriptVariable(_conditionToFindObject.left.attr + '__' + _conditionToFindObject.right);
                 // consider only conditions of this array, thus only last rows which match with unique array name
-                if (_conditionToFindObject.left.parent !== _uniqueMarkerName) {
+                if (_conditionToFindObject.left.parent !== _uniqueMarkerName ) {
                   break;
                 }
               }
@@ -201,20 +211,20 @@ var extracter = {
               _uniqueMarkerName += _filterUniqueName;
               _uniqueMarkerName = this._cleanUniqueMarkerName(_uniqueMarkerName);
               // correction of previously detected conditions
-              for (var m = 0; m < _conditionsToFindObjectInArray.length; m++) {
+              for (var m = 0; m < _conditionsToFindObjectInArray.length ; m++) {
                 _conditionsToFindObjectInArray[m].left.parent += _filterUniqueName;
               }
             }
             // create the new array
             if (!_res[_uniqueMarkerName]) {
-              _res[_uniqueMarkerName] = {
-                name: _arrayName,
-                type: (_isIteratorUsedInCurrentArray === true) ? 'array' : 'objectInArray',
-                parent: _prevMarker,
-                parents: _parents.slice(0),
-                position: {},
-                iterators: [],
-                xmlParts: []
+              _res[_uniqueMarkerName]={
+                name      : _arrayName,
+                type      : (_isIteratorUsedInCurrentArray === true) ? 'array' : 'objectInArray',
+                parent    : _prevMarker,
+                parents   : _parents.slice(0),
+                position  : {},
+                iterators : [],
+                xmlParts  : []
               };
             }
             if (_isIteratorUsedInCurrentArray === true && _res[_uniqueMarkerName].position.start === undefined) {
@@ -230,6 +240,10 @@ var extracter = {
           else {
             if (_res[_uniqueMarkerName].position.end === undefined) {
               _res[_uniqueMarkerName].position.end = _markerPos;
+              if (_repeater) {
+                _res[_uniqueMarkerName].repeater = _repeater;
+                _repeater = '';
+              }
               var _prevIteratorStr = '';
               for (var it = 0; it < _iterators.length; it++) {
                 var _iterator = _iterators[it];
@@ -237,14 +251,14 @@ var extracter = {
                 if (_prevIteratorStr !== _iteratorStr) {
                   var _iteratorInfo = {};
                   // if the iterator is in a sub-object. Ex: movie.sort+1
-                  if (/\./.test(_iteratorStr) === true) {
+                  if (/\./.test(_iteratorStr)===true) {
                     var _splitIterator = _iteratorStr.split('.');
-                    _iteratorInfo = { obj: _splitIterator[0], attr: _splitIterator[1] };
+                    _iteratorInfo = {obj : _splitIterator[0],attr : _splitIterator[1]};
                   }
                   else {
-                    _iteratorInfo = { attr: _iteratorStr };
+                    _iteratorInfo = {attr : _iteratorStr};
                   }
-                  if (_iterator.isSpecial === true) {
+                  if (_iterator.isSpecial===true) {
                     _iteratorInfo.isSpecial = true;
                     _res[_uniqueMarkerName].containSpecialIterator = true;
                   }
@@ -261,7 +275,7 @@ var extracter = {
         }
 
         // remove whitespaces
-        if (_char !== ' ') {
+        if (_char!==' ') {
           _prevChar = _char;
         }
       }
@@ -279,10 +293,11 @@ var extracter = {
           _objParent = _uniqueMarkerName;
         }
         var _xmlPart = {
-          attr: _word,
-          formatters: _formaters,
-          obj: _objParent,
-          pos: _markerPos
+          attr       : _word,
+          formatters : _formaters,
+          obj        : _objParent,
+          pos        : _markerPos,
+          posOrigin  : _markerPos
         };
         if (_conditions.length > 0) {
           _xmlPart.conditions = _conditions;
@@ -297,7 +312,7 @@ var extracter = {
    * Remove dash from string to avoir errors
    * @param {String} markerName Marker name to clean
    */
-  _cleanUniqueMarkerName: function (markerName) {
+  _cleanUniqueMarkerName : function (markerName) {
     return markerName.replace(/-/g, '');
   },
 
@@ -305,33 +320,42 @@ var extracter = {
    * Sort xml parts. Used in splitXml
    * @param  {Array} xmlParts (modified by the function)
    */
-  sortXmlParts: function (xmlParts) {
-    xmlParts.sort(function (a, b) {
-      if (a.pos > b.pos) {
-        return 1;
+  sortXmlParts : function (xmlParts) {
+    xmlParts.sort(function (a,b) {
+      if (Math.trunc(a.pos) > Math.trunc(b.pos)) {
+        return  1;
       }
-      if (a.pos < b.pos) {
+      if (Math.trunc(a.pos) < Math.trunc(b.pos)) {
         return -1;
       }
       // When two arrays are incremented in the same time (d.tab[i+1].subtab[i+1]), pos is the same...
       // I use the length of the string to garuantee that the loop "d.tab" is before "d.tab.subtab".
-      if (a.array === 'start' && b.array === 'start') {
+      if (a.array==='start' && b.array==='start') {
         return (a.obj.length - b.obj.length);
       }
-      if (a.array === 'end' && b.array === 'end') {
+      if (a.array==='end'   && b.array==='end'  ) {
         return (b.obj.length - a.obj.length);
       }
-      // other basic cases:
-      if (a.array === 'start') {
+      // after detecting arrays and conditional blocks, some markers may have moved in XML
+      // It could create conflicts (many markers at the same XML position).
+      // So, we sort them using their original position
+      if (a.posOrigin > b.posOrigin) {
+        return  1;
+      }
+      if (a.posOrigin < b.posOrigin) {
         return -1;
       }
-      if (b.array === 'start') {
-        return 1;
+      // other basic cases:
+      if (a.array==='start') {
+        return -1;
       }
-      if (a.array === 'end') {
-        return 1;
+      if (b.array==='start') {
+        return  1;
       }
-      if (b.array === 'end') {
+      if (a.array==='end')   {
+        return  1;
+      }
+      if (b.array==='end')   {
         return -1;
       }
       return 0;
@@ -345,13 +369,14 @@ var extracter = {
    * @param {object} descriptor : descriptor generated by the function splitMarkers
    * @return {object} descriptor with xml inside
    */
-  splitXml: function (xml, descriptor) {
+  splitXml : function (xml, descriptor) {
+    const MARKER_PRESENCE_CHAR_REGEX = /\uFFFF/g;
     var _res = {
-      staticData: {
-        before: '',
-        after: ''
+      staticData : {
+        before : '',
+        after  : ''
       },
-      dynamicData: descriptor
+      dynamicData : descriptor
     };
 
     findAndSetValidPositionOfConditionalBlocks(xml, descriptor);
@@ -368,7 +393,7 @@ var extracter = {
 
     // if all parts are empty, return all the xml
     if (_allSortedParts.length === 0) {
-      _res.staticData.before = xml.slice(0);
+      _res.staticData.before = xml.slice(0).replace(MARKER_PRESENCE_CHAR_REGEX,'');
     }
 
     // Slice the xml
@@ -378,15 +403,35 @@ var extracter = {
     var _currentlyVisitedArray = [];
     var _lastVisitedArrayPos = {};
 
+    // avoid having two (or more) markers at the same XML position for the final sort in builder.assembleXmlParts
+    // fix this using float position.
+    for (let i = 0; i < _allSortedParts.length; i++) {
+      let _part = _allSortedParts[i];
+      if (Math.trunc(_part.pos) === Math.trunc(_prevPart.pos)) {
+        // Why adding 1/64? to avoid rounding problems of floats (http://0.30000000000000004.com)
+        // This let us the possibility to have 64 markers at the same position. It should enough for all cases.
+        _part.pos = _prevPart.pos + 1/64;
+        // fix also beginning and ending of arrays
+        if (_part.array === 'start') {
+          descriptor[_part.obj].position.start = _part.pos;
+        }
+        else if (_part.array === 'end') {
+          descriptor[_part.obj].position.end = _part.pos;
+        }
+      }
+      _prevPart = _part;
+    }
+
+    _prevPart = {};
     for (var i = 0; i < _allSortedParts.length; i++) {
       var _part = _allSortedParts[i];
-      var _pos = Math.ceil(_part.pos); // can be a float
-      if (i === 0) {
-        _res.staticData.before = xml.slice(_prevPos, _pos);
+      var _pos  = Math.trunc(_part.pos); // can be a float
+      if (i===0) {
+        _res.staticData.before = xml.slice(_prevPos, _pos).replace(MARKER_PRESENCE_CHAR_REGEX,'');
       }
       if (_prevPart.array === 'start') {
         _arrayDepth++;
-        _prevPart.after = xml.slice(_prevPos, _pos);
+        _prevPart.after = xml.slice(_prevPos, _pos).replace(MARKER_PRESENCE_CHAR_REGEX,'');
         _prevPart.depth = _arrayDepth;
         _currentlyVisitedArray.push(_prevPart.obj);
         _lastVisitedArrayPos = descriptor[_prevPart.obj].position;
@@ -395,7 +440,7 @@ var extracter = {
       if (_part.array === 'end') {
         descriptor[_part.obj].depth = _arrayDepth;
         _arrayDepth--;
-        _part.before = xml.slice(_prevPos, _pos);
+        _part.before = xml.slice(_prevPos, _pos ).replace(MARKER_PRESENCE_CHAR_REGEX,'');
       }
       // move or remove part which are not in the repetition section
       for (var j = _currentlyVisitedArray.length - 1; j >= 0; j--) {
@@ -403,14 +448,14 @@ var extracter = {
         var _arrayPos = descriptor[_lastVisitedArrayName].position;
         var _partParents = descriptor[_part.obj].parents;
         // if the part belongs to the current visited array, keep it in this array.
-        if (_part.pos <= _arrayPos.end && (_part.depth === _arrayDepth && _lastVisitedArrayName === _part.obj || _part.array === 'start' || _part.array === 'end' || _partParents.indexOf(_lastVisitedArrayName) !== -1)) {
+        if (_part.pos <= _arrayPos.end && (_part.depth === _arrayDepth &&  _lastVisitedArrayName === _part.obj || _part.array === 'start' || _part.array === 'end' || _partParents.indexOf(_lastVisitedArrayName) !== -1)) {
           break;
         }
         if (_part.pos > _arrayPos.start && _part.pos < _arrayPos.end && _partParents.indexOf(_lastVisitedArrayName) === -1 && _lastVisitedArrayName !== _part.obj) {
           _part.moveTo = _lastVisitedArrayName;
           break;
         }
-        else if (_part.pos > _arrayPos.end && _part.pos < _arrayPos.endOdd) {
+        else if (Math.trunc(_part.pos) > Math.trunc(_arrayPos.end) && Math.trunc(_part.pos) < Math.trunc(_arrayPos.endOdd)) {
           _part.toDelete = true; // remove all parts which are in odd zone of arrays
           break;
         }
@@ -419,36 +464,43 @@ var extracter = {
       if (_arrayDepth > 0 && _part.pos >= _lastVisitedArrayPos.endOdd && _part.pos > _lastVisitedArrayPos.end) {
         _currentlyVisitedArray.pop();
         if (_currentlyVisitedArray.length > 0) {
-          _lastVisitedArrayPos = descriptor[_currentlyVisitedArray[_currentlyVisitedArray.length - 1]].position;
+          _lastVisitedArrayPos = descriptor[_currentlyVisitedArray[_currentlyVisitedArray.length-1]].position;
         }
         else {
           _lastVisitedArrayPos = {};
         }
       }
       if (!_prevPart.array && _part.array !== 'end') {
-        _prevPart.after = xml.slice(_prevPos, _pos);
+        let _xmlPart = xml.slice(_prevPos, _pos ).replace(MARKER_PRESENCE_CHAR_REGEX,'');
+        // if toDelete = true, we are in the odd part of the array. This xml part will be removed
+        if (_prevPart.toDelete !== true) {
+          _prevPart.after = _xmlPart;
+        }
+        else if (_part.toDelete !== true) {
+          _part.before = _xmlPart;
+        }
       }
       if (_prevPart.array && _prevPart.array === 'end') {
-        if (_prevPos !== _pos) {
+        if (_prevPos!==_pos) {
           // if we have two adjacent arrays, keep the xml which is between in two arrys
-          if (_part.array && _part.array === 'start') {
-            descriptor[_part.obj].before = xml.slice(_prevPos, _pos);
+          if (_part.array && _part.array==='start') {
+            descriptor[_part.obj].before =  xml.slice(_prevPos, _pos).replace(MARKER_PRESENCE_CHAR_REGEX,'');
           }
           else {
-            _part.before = xml.slice(_prevPos, _pos);
+            _part.before = xml.slice(_prevPos, _pos).replace(MARKER_PRESENCE_CHAR_REGEX,'');
           }
         }
       }
       // if there is a dead zone (ex. odd part of the array), jump to the next part
       if (_part.array === 'end') {
-        _prevPos = descriptor[_part.obj].position.endOdd;
+        _prevPos = Math.trunc(descriptor[_part.obj].position.endOdd);
       }
       else if (_pos > _prevPos) {
         _prevPos = _pos;
       }
       // the end, put the last xml part in the staticData object
-      if (i === _allSortedParts.length - 1) {
-        _res.staticData.after = xml.slice(_prevPos);
+      if (i === _allSortedParts.length-1) {
+        _res.staticData.after = xml.slice(_prevPos).replace(MARKER_PRESENCE_CHAR_REGEX,'');
       }
       _prevPart = _part;
     }
@@ -462,7 +514,7 @@ var extracter = {
    * @param  {object} descriptor : descriptor returned by splitXml
    * @return {object} descriptor with the hierarchy array
    */
-  buildSortedHierarchy: function (descriptor) {
+  buildSortedHierarchy : function (descriptor) {
     var _sortedNodes = [];
 
     // find all node which have parents
@@ -477,7 +529,7 @@ var extracter = {
     // sort by depth
     _treeLeaves.sort(sortByDepth);
 
-    function insertIfNotExists(table, node) {
+    function insertIfNotExists (table, node) {
       for (var i = 0; i < table.length; i++) {
         if (table[i] === node) {
           return;
@@ -487,10 +539,10 @@ var extracter = {
     }
     for (var i = 0; i < _treeLeaves.length; i++) {
       var _obj = _treeLeaves[i];
-      for (var j = _obj.parents.length - 1; j >= 0; j--) {
+      for (var j = _obj.parents.length - 1; j >= 0 ; j--) {
         insertIfNotExists(_sortedNodes, _obj.parents[j]);
       }
-      insertIfNotExists(_sortedNodes, _obj.objName);
+      insertIfNotExists(_sortedNodes,  _obj.objName);
     }
     descriptor.hierarchy = _sortedNodes;
     return descriptor;
@@ -502,9 +554,9 @@ var extracter = {
    * @param  {object} descriptor : descriptor returned by splitXml
    * @return {object} descriptor with modified xmlParts
    */
-  deleteAndMoveNestedParts: function (descriptor) {
+  deleteAndMoveNestedParts : function (descriptor) {
     for (var _objName in descriptor.dynamicData) {
-      var _xmlParts = descriptor.dynamicData[_objName].xmlParts;
+      var _xmlParts =  descriptor.dynamicData[_objName].xmlParts;
       for (var i = 0; i < _xmlParts.length; i++) {
         var _part = _xmlParts[i];
         if (_part.moveTo !== undefined) {
@@ -531,7 +583,9 @@ module.exports = extracter;
  * @param  {Object} descriptor descriptor generated by the function splitMarkers
  * @return {Array}
  */
-function findAndSetValidPositionOfConditionalBlocks(xml, descriptor) {
+function findAndSetValidPositionOfConditionalBlocks (xml, descriptor) {
+  // TODO performance: flattenXML should be called earlier and used also to find array positions
+  var _xmlFlattened = parser.flattenXML(xml);
   for (var _objName in descriptor) {
     var _xmlParts = descriptor[_objName].xmlParts;
     var _conditionalBlockDetected = [];
@@ -549,23 +603,26 @@ function findAndSetValidPositionOfConditionalBlocks(xml, descriptor) {
       }
       else if (_lastFormatter.indexOf('showEnd') !== -1 || _lastFormatter.indexOf('hideEnd') !== -1) {
         var _beginPart = _conditionalBlockDetected.pop();
-        var _safeXMLBlocks = parser.findSafeConditionalBlockPosition(xml, _beginPart.pos, _part.pos);
-        // TODO manage when nothing can be removed
+        if (_beginPart === undefined) {
+          throw new Error('Missing at least one showBegin or hideBegin');
+        }
+        // convert XML string into an array and find the array index of beginTextIndex
+        var _safeXMLBlocks = parser.findSafeConditionalBlockPosition(_xmlFlattened, _beginPart.pos, _part.pos);
         if (_safeXMLBlocks.length > 0) {
-          _newParts.push({ obj: _beginPart.obj, formatters: [], pos: _safeXMLBlocks[0][0] - 0.2 });
-          _beginPart.pos = _safeXMLBlocks[0][0] - 0.1;
-          _part.pos = _safeXMLBlocks[0][1] - 0.2;
-          _newParts.push({ obj: _beginPart.obj, formatters: [], pos: _safeXMLBlocks[0][1] - 0.1 });
+          _beginPart.posOrigin = _beginPart.pos;
+          _part.posOrigin      = _part.pos;
+          _beginPart.pos       = _safeXMLBlocks[0][0];
+          _part.pos            = _safeXMLBlocks[0][1];
           for (var j = 1; j < _safeXMLBlocks.length; j++) {
             var _safeXMLBlock = _safeXMLBlocks[j];
-            // generate empty part before everything to avoid breaking XML with arrays
-            _newParts.push({ obj: _beginPart.obj, formatters: [], pos: _safeXMLBlock[0] - 0.2 });
-            _newParts.push({ attr: _beginPart.attr, obj: _beginPart.obj, formatters: _beginPart.formatters, pos: _safeXMLBlock[0] - 0.1 });
-            _newParts.push({ attr: _beginPart.attr, obj: _beginPart.obj, formatters: _part.formatters, pos: _safeXMLBlock[1] - 0.2 });
-            _newParts.push({ obj: _beginPart.obj, formatters: [], pos: _safeXMLBlock[1] - 0.1 });
+            _newParts.push({ obj : _beginPart.obj, formatters : _beginPart.formatters, attr : _beginPart.attr, pos : _safeXMLBlock[0], posOrigin : _beginPart.posOrigin, conditions : _beginPart.conditions });
+            _newParts.push({ obj : _beginPart.obj, formatters : _part.formatters     , attr : _beginPart.attr, pos : _safeXMLBlock[1], posOrigin : _part.posOrigin     , conditions : _beginPart.conditions });
           }
         }
       }
+    }
+    if (_conditionalBlockDetected.length > 0) {
+      throw new Error('Missing at least one showEnd or hideEnd');
     }
     descriptor[_objName].xmlParts = descriptor[_objName].xmlParts.concat(_newParts);
   }
@@ -578,25 +635,24 @@ function findAndSetValidPositionOfConditionalBlocks(xml, descriptor) {
  * @param  {Object} descriptor descriptor generated by the function splitMarkers
  * @return {Array}             modify descriptor and return an array of odd parts
  */
-function findAndSetExactPositionOfArrays(xml, descriptor) {
+function findAndSetExactPositionOfArrays (xml, descriptor) {
   var _oddZones = [];
   for (var _objName in descriptor) {
     var _obj = descriptor[_objName];
-    if (_obj.type === 'array' && _obj.iterators.length > 0) {
+    if (_obj.type === 'array' && _obj.iterators.length>0) {
       var _roughPosStart = descriptor[_objName].position.start;
       var _roughPosEnd = descriptor[_objName].position.end;
-      var _subString = xml.slice(_roughPosStart, _roughPosEnd);
+      var _subString = xml.slice(Math.trunc(_roughPosStart), Math.trunc(_roughPosEnd));
       var _pivot = parser.findPivot(_subString);
       // TODO test if the _pivot is not null !!!!!!!!!!!!
       // absolute position of the pivot
-      if (!_pivot) continue;
       _pivot.part2Start.pos += _roughPosStart;
       _pivot.part1End.pos += _roughPosStart;
       var _realArrayPosition = parser.findRepetitionPosition(xml, _pivot, _roughPosStart);
       // TODO test if the repetition is found !!!!!!!!!!!!
-      descriptor[_objName].position = { start: _realArrayPosition.startEven, end: _realArrayPosition.endEven, /* 'startOdd':_realArrayPosition.startOdd,*/ endOdd: _realArrayPosition.endOdd };
-      descriptor[_objName].xmlParts.push({ obj: _objName, array: 'start', pos: _realArrayPosition.startEven });
-      descriptor[_objName].xmlParts.push({ obj: _objName, array: 'end', pos: _realArrayPosition.endEven /* ,'next':_realArrayPosition.endOdd*/ });
+      descriptor[_objName].position = {start : _realArrayPosition.startEven, end : _realArrayPosition.endEven, endOdd : _realArrayPosition.endOdd};
+      descriptor[_objName].xmlParts.push({obj : _objName, array : 'start', pos : _realArrayPosition.startEven, posOrigin : _roughPosStart});
+      descriptor[_objName].xmlParts.push({obj : _objName, array : 'end'  , pos : _realArrayPosition.endEven  , posOrigin : _roughPosEnd });
       _oddZones.push([_realArrayPosition.endEven, _realArrayPosition.endOdd]);
     }
   }
@@ -608,7 +664,7 @@ function findAndSetExactPositionOfArrays(xml, descriptor) {
  * @param  {Object} descriptor descriptor generated by the function splitMarkers
  * @return {Array}             all parts in one array
  */
-function getAllParts(descriptor) {
+function getAllParts (descriptor) {
   var _allSortedParts = [];
   for (var _objName in descriptor) {
     var _xmlParts = descriptor[_objName].xmlParts;
@@ -625,7 +681,7 @@ function getAllParts(descriptor) {
  * @param  {Object} descriptor descriptor generated by the function splitMarkers
  * @return {Object}            the keys are a list of parent's names
  */
-function getAllNodeWhichHaveParents(descriptor) {
+function getAllNodeWhichHaveParents (descriptor) {
   var _hasChildren = {};
   for (var _objName in descriptor.dynamicData) {
     var _obj = descriptor.dynamicData[_objName];
@@ -640,18 +696,18 @@ function getAllNodeWhichHaveParents(descriptor) {
  * @param  {Object} descriptor  descriptor generated by the function splitMarkers
  * @return {Array}              array of leaves
  */
-function getAllLeaves(hasChildren, descriptor) {
+function getAllLeaves (hasChildren, descriptor) {
   var _treeLeaves = [];
   for (var _objName in descriptor.dynamicData) {
     if (hasChildren[_objName] === undefined) {
       var _obj = descriptor.dynamicData[_objName];
       var _tempObj = {
-        objName: _objName,
-        parent: _obj.parent,
-        type: _obj.type,
-        parents: [], // TODO use _obj.parents
-        hasOnlyObjects: (_obj.type !== 'array') ? true : false,
-        branchName: ''
+        objName        : _objName,
+        parent         : _obj.parent,
+        type           : _obj.type,
+        parents        : [], // TODO use _obj.parents
+        hasOnlyObjects : (_obj.type !== 'array') ? true : false,
+        branchName     : ''
       };
       _treeLeaves.push(_tempObj);
     }
@@ -664,25 +720,32 @@ function getAllLeaves(hasChildren, descriptor) {
  * @param  {Array}  treeLeaves  array returned by getAllLeaves
  * @param  {Object} descriptor  descriptor generated by the function splitMarkers
  */
-function computeDepthAndBranchNames(treeLeaves, descriptor) {
+function computeDepthAndBranchNames (treeLeaves, descriptor) {
   // for each leaf, go up in the tree until we reach the root
   for (var i = 0; i < treeLeaves.length; i++) {
     var _obj = treeLeaves[i];
     var _objParentName = _obj.parent;
     var _nbParents = 0;
+    var _firstParentDepth = 0;
     // go up in the hierarchy until we reach the root
     while (_objParentName !== '') {
       _obj.parents.push(_objParentName);
       var _objParent = descriptor.dynamicData[_objParentName];
       if (_objParent.type === 'array') {
         _obj.hasOnlyObjects = false;
+        if (_firstParentDepth === 0) {
+          _firstParentDepth = _objParent.depth;
+        }
       }
-      _obj.branchName = _objParentName + ' ' + _obj.branchName; // the whitespace is the only character which cannot exist in the name of a node
+      _obj.branchName = _objParentName + ' ' +_obj.branchName; // the whitespace is the only character which cannot exist in the name of a node
       _objParentName = _objParent.parent;
-      _nbParents++; // count the number of parents
+      _nbParents++;
     }
-    // store the depth of the leaf (= nb parents)
-    _obj.depth = _nbParents;
+    // store the depth of the leaf (= the depth of the first parent array)
+    if (_obj.depth === undefined ) {
+      _obj.depth = _firstParentDepth;
+    }
+    _obj.jsonDepth = _nbParents;
   }
 }
 
@@ -692,7 +755,7 @@ function computeDepthAndBranchNames(treeLeaves, descriptor) {
  * @param  {Object} b one object of treeLeaves
  * @return {Integer}  1, -1, 0
  */
-function sortByDepth(a, b) {
+function sortByDepth (a,b) {
   // if the branch contains only objects, put it first
   if (a.hasOnlyObjects === true && b.hasOnlyObjects === false) {
     return -1;
@@ -701,7 +764,7 @@ function sortByDepth(a, b) {
     return 1;
   }
   // if the two items do not share the same branch at all, sort them by branch name.
-  if (a.branchName.indexOf(b.branchName) === -1 && b.branchName.indexOf(a.branchName) === -1) {
+  if (a.branchName.indexOf(b.branchName) === -1 &&  b.branchName.indexOf(a.branchName) === -1) {
     if (a.branchName > b.branchName) {
       return 1;
     }
@@ -722,6 +785,12 @@ function sortByDepth(a, b) {
   }
   else if (a.type === 'array' && b.type !== 'array') {
     return 1;
+  }
+  if (a.jsonDepth > b.jsonDepth) {
+    return 1;
+  }
+  else if (a.jsonDepth < b.jsonDepth) {
+    return -1;
   }
 
   return 0;
