@@ -30,7 +30,11 @@ module.exports  = {
     vor       : {extension : 'vor'  , desc : 'StarWriter 5.0 Template'                                       , format : 'StarWriter 5.0 Vorlage/Template'              },
     vor4      : {extension : 'vor'  , desc : 'StarWriter 4.0 Template'                                       , format : 'StarWriter 4.0 Vorlage/Template'              },
     vor3      : {extension : 'vor'  , desc : 'StarWriter 3.0 Template'                                       , format : 'StarWriter 3.0 Vorlage/Template'              },
-    xhtml     : {extension : 'html' , desc : 'XHTML Document'                                                , format : 'XHTML Writer File'                            }
+    xhtml     : {extension : 'html' , desc : 'XHTML Document'                                                , format : 'XHTML Writer File'                            },
+    jpg       : {extension : 'jpg'  , desc : 'Joint Photographic Experts Group'                              , format : 'writer_jpg_Export'                            },
+    jpeg      : {extension : 'jpeg' , desc : 'Joint Photographic Experts Group'                              , format : 'writer_jpg_Export'                            },
+    png       : {extension : 'png'  , desc : 'Portable Network Graphic'                                      , format : 'writer_png_Export'                            },
+    epub      : {extension : 'epub' , desc : 'Electronic Publication'                                        , format : 'EPUB'                                         }
   },
   web : {
     etext     : {extension : 'txt'  , desc : 'Text Encoded (OpenOffice.org Writer/Web)'                      , format : 'Text (encoded) (StarWriter/Web)'              },
