@@ -1,13 +1,26 @@
 var fs = require('fs');
 var path = require('path');
 var assert = require('assert');
+var crypto = require('crypto');
 var params = require('./params');
 var exec = require('child_process').exec;
+var debug = require('debug')('carbone:helper');
 
 var sameUIDCounter = 0;
 var prevTimestamp = 0;
 
+// Random values to pre-allocate
+const randomPool      = Buffer.alloc(256);
+var randomPoolPointer = randomPool.length;
+
 var helper = {
+
+  /**
+   * This factor is used to modify asserts in performance unit tests.
+   * If those tests are executed on slow machines, this factor is higher
+   * to reduce required level of performance in the test.
+   */
+  CPU_PERFORMANCE_FACTOR : 1.0,
 
   /**
    * Generate a unique id
@@ -28,6 +41,55 @@ var helper = {
   },
 
   /**
+   * Generate a constant-length base64 random string without forbidden characters for filesystem
+   *
+   * [/+] are replaced by [-_]
+   *
+   * Length : 22 characters
+   * It is 5 times slower than getUID above but a lot more secure to generate public filename
+   *
+   * It uses Cryptographically Secure Pseudo-Random Number Generator, with unbiased (ie secure) tranformation
+   * Same collision probability as standard 128 bits uuid-v4
+   * Doc: https://gist.github.com/joepie91/7105003c3b26e65efcea63f3db82dfba
+   *
+   * @return {String}   The random string
+   */
+  RANDOM_STRING_LENGTH : 22,
+  getRandomString      : function () {
+    if (randomPoolPointer > randomPool.length - 16) {
+      // It is alot simpler to use sync version
+      // Otherwose, we must maintain a table of callback
+      // This sync function is called 1/16 times
+      crypto.randomFillSync(randomPool);
+      randomPoolPointer = 0;
+    }
+    let _randomStr = randomPool.toString('base64', randomPoolPointer, (randomPoolPointer += 16));
+    return _randomStr.slice(0, -2).replace(/\//g,'_').replace(/\+/g,'-');
+  },
+
+
+  /**
+   * Encode any string to a base64 string, safe for generating POSIX compatible filenames
+   * https://fr.wikipedia.org/wiki/Base64
+   *
+   * @param  {String}  str The string
+   * @return {String}      A safe filename
+   */
+  encodeSafeFilename : function (str) {
+    return Buffer.from(str||'').toString('base64').replace(/=/g,'').replace(/\//g,'_').replace(/\+/g,'-');
+  },
+
+  /**
+   * Decode a base64 safe filename to the original string
+   *
+   * @param  {String}  filename  The filename
+   * @return {String}            Return original string
+   */
+  decodeSafeFilename : function (filename) {
+    return Buffer.from((filename||'').replace(/_/g,'/').replace(/-/g,'+'), 'base64').toString('utf8');
+  },
+
+  /**
    * Beautiful assert between two objects
    * Be careful, it will inspect with a depth of 100
    *
@@ -35,16 +97,9 @@ var helper = {
    * @param {object} expected : expected
    * @return {type} throw an error if the two objects are different
    */
-  assert : function (result, expected, message) {
-    try {
-      assert.equal(JSON.stringify(result, null, 2), JSON.stringify(expected, null, 2), message);
-    }
-    catch (e) {
-      e.showDiff = true;
-      throw e;
-    }
+  assert : function (result, expected) {
+    assert.strictEqual(JSON.stringify(result, null, 2), JSON.stringify(expected , null, 2));
   },
-
   /**
    * Read all files of a directory
    *
@@ -66,6 +121,19 @@ var helper = {
 
   cleanJavascriptVariable : function (attributeName) {
     return  attributeName.replace(/[^a-zA-Z0-9$_]/g, '_');
+  },
+
+  /**
+   * Remove quotes from string
+   *
+   * @param  {Strring} str  string with or without quotes
+   * @return {String}       string without surrounding quotes
+   */
+  removeQuote : function (str) {
+    if (typeof(str) === 'string') {
+      return str.replace(/^ *['"]?/, '').replace(/['"]? *$/, '');
+    }
+    return str;
   },
 
   /**
@@ -125,7 +193,8 @@ var helper = {
     for (var i = 0; i < _attrs.length; i++) {
       var _attr = _attrs[i];
       if (_currentObj[_attr] === undefined) {
-        return '[[C_ERROR]] '+_attr+' not defined';
+        debug('[[C_ERROR]] '+_attr+' not defined');
+        return '';
       }
       _currentObj = _currentObj[_attr];
     }
@@ -170,7 +239,6 @@ var helper = {
         var _filename = path.join(dir, _list[i]);
         var _stat = fs.statSync(_filename);
         var _baseName = path.basename(_filename); // get the base name in order to eliminate folder which starts by a point.
-
         if (_stat.isFile() && _regExp.test(_baseName)) {
           // if this is a file, push it in the table
           _files.push(_filename);
@@ -353,8 +421,36 @@ var helper = {
       c++;
     }
     return (s1.length + s2.length) / 2 - lcs;
-  }
+  },
 
+  /**
+   * Measure CPU performance on this machine and update helper.CPU_PERFORMANCE_FACTOR
+   *
+   * This function is called in tests.
+   */
+  updatePerformanceFactor : function () {
+    const _referenceTime = 20; // time
+    const _nbRows = 1000000;
+    let _data = [];
+    // do 4 loops, and divide the time by 4 to get a rougly constant mean
+    const _start = process.hrtime();
+    for (let i = 0; i < _nbRows; i++) { _data.push( Math.floor(Math.random() * 1000)); } // eslint-disable-line
+    for (let i = 0; i < _nbRows; i++) { _data[i] += Math.floor(Math.random() * 1000);  } // eslint-disable-line
+    for (let i = 0; i < _nbRows; i++) { _data[i] += Math.floor(Math.random() * 1000);  } // eslint-disable-line
+    for (let i = 0; i < _nbRows; i++) { _data[i] += Math.floor(Math.random() * 1000);  } // eslint-disable-line
+    const _diff = process.hrtime(_start);
+    const _elapsed = parseInt(((_diff[0] * 1e9 + _diff[1]) / 1e6) / 4, 10);
+    helper.CPU_PERFORMANCE_FACTOR = _elapsed / _referenceTime;
+
+    // print _data to make sure V8 does not remove useless code
+    console.log('\nRaw CPU perf test : ' + _elapsed + ' ms ('+ _data[_nbRows/2] + ')');
+    console.log('CPU_PERFORMANCE_FACTOR = '+ helper.CPU_PERFORMANCE_FACTOR);
+  }
 };
+
+// Update Performance factor for unit tests
+if (typeof global.it === 'function' && typeof global.describe === 'function') {
+  helper.updatePerformanceFactor();
+}
 
 module.exports = helper;
