@@ -6,26 +6,25 @@ var params = require('./params');
 var helper = require('./helper');
 var format = require('./format');
 var builder = require('./builder');
-var parser = require('./parser');
+var input = require('./input');
 var preprocessor = require('./preprocessor');
 var translator = require('./translator');
 var converter = require('./converter');
-var dayjs = require('dayjs');
-var locale = require('../formatters/_locale.js');
 var debug = require('debug')('carbone');
-var xmljs = require('xml-js')
-const { Worker, parentPort, workerData } = require('worker_threads');
+var dayjs = require('dayjs');
+var locales = require('../formatters/_locale');
+var xmljs = require('xml-js');
 
 var carbone = {
 
-  formatters: {},
-
   /**
-   * This function is NOT asynchrone (It may create the template or temp directory synchronously)
+   * This function is NOT asynchronous (It may create the template or temp directory synchronously)
    * @param {Object} options {
    *                           tempPath     : system temp directory by default
    *                           templatePath : it will create the directory if it does not exists
-   *                           lang         : set default lang of carbone, can be overwrite in carbone.render options
+   *                           renderPath   : where rendered files are temporary saved. It will create the directory if it does not exists
+   *                           lang         : set default lang of carbone, can be overwrite in carbone.render options.lang
+   *                           timezone     : set default timezone of carbone, can be overwrite in carbone.render options.timezone
    *                           translations : overwrite carbone translations object
    *                           currencySource : currency of data, it depends on the locale if empty
    *                           currencyTarget : default target currency when the formatter convCurr is used without target
@@ -56,11 +55,14 @@ var carbone = {
     if (options.tempPath !== undefined && fs.existsSync(params.tempPath) === false) {
       fs.mkdirSync(params.tempPath, '0755');
     }
+    if (fs.existsSync(params.renderPath) === false) {
+      fs.mkdirSync(params.renderPath, '0755');
+    }
     if (options.factories !== undefined || options.startFactory !== undefined) {
       converter.init();
     }
-    // update dayjs lang if it changes
-    dayjs.locale(params.lang);
+    dayjs.tz.setDefault(params.timezone);
+    dayjs.locale(params.lang.toLowerCase());
   },
 
   /**
@@ -73,12 +75,17 @@ var carbone = {
 
     params.tempPath = _tmpDir;
     params.templatePath = process.cwd();
+    params.renderPath = path.join(params.tempPath, 'carbone_render');
     params.factories = 1;
     params.attempts = 2;
     params.startFactory = false;
+    params.factoryMemoryFileSize = 1;
+    params.factoryMemoryThreshold = 50;
+    params.converterFactoryTimeout = 60000;
     params.uidPrefix = 'c';
     params.pipeNamePrefix = '_carbone';
     params.lang = 'en';
+    params.timezone = 'Europe/Paris';
     params.translations = {};
     params.currencySource = '';
     params.currencyTarget = '';
@@ -107,7 +114,7 @@ var carbone = {
    */
   addFormatters: function (customFormatters) {
     for (var f in customFormatters) {
-      carbone.formatters[f] = customFormatters[f];
+      input.formatters[f] = customFormatters[f];
     }
   },
 
@@ -120,6 +127,7 @@ var carbone = {
     var _fullPath = path.join(params.templatePath, fileId);
     fs.unlink(_fullPath, callback);
   },
+
 
   /**
    * Return the list of possible conversion format
@@ -140,14 +148,24 @@ var carbone = {
     return _res;
   },
 
+  /**
+   * Render XML directly
+   *
+   * @param {String}        xml          The XML
+   * @param {Object|Array}  data         The data
+   * @param {Object}        optionsRaw   The options raw
+   * @param {Function}      callbackRaw  The callback raw
+   */
   renderXML: function (xml, data, optionsRaw, callbackRaw) {
-    parseOptions(optionsRaw, callbackRaw, function (options, callback) {
+    input.parseOptions(optionsRaw, callbackRaw, function (options, callback) {
+      // Clean XML tags inside Carbone markers and translate
+      xml = preprocessor.preParseXML(xml, options);
       return builder.buildXML(xml, data, options, callback);
     });
   },
-
   /**
    * Renders a template with given datas and return result to the callback function
+   *
    * @param {String}       templatePath : file name of the template (or absolute path)
    * @param {Object|Array} data : Datas to be inserted in the template represented by the {d.****}
    * @param {Object}       optionsRaw [optional] : {
@@ -156,16 +174,21 @@ var carbone = {
    *                          'extension'    : 'odt' || undefined Specify the template extension
    *                          'variableStr'  : ''    pre-declared variables,
    *                          'lang'         : overwrite default lang. Ex. "fr"
+   *                          'timezone'     : set timezone for date formatters (Europe/Paris) by default
    *                          'translations' : overwrite all loaded translations {fr: {}, en: {}, es: {}}
    *                          'enum'         : { ORDER_STATUS : ['open', 'close', 'sent']
    *                          'currencySource'   : currency of data, 'EUR'
    *                          'currencyTarget' : default target currency when the formatter convCurr is used without target
    *                          'currencyRates'  : rates, based on EUR { EUR : 1, USD : 1.14 }
+   *                          'hardRefresh'  : (default: false) if true, LibreOffice is used to render and refresh the content of the report at the end of Carbone process
+   *                          'renderPrefix' : If defined, it returns a path instead of a buffer, and it adds this prefix in the filename
+   *                                           The filename will contains also the report name URL Encoded
    *                       }
-   * @param {Function}     callbackRaw(err, res, reportName) : Function called after generation with the result
+   * @param {Function}     callbackRaw(err, bufferOrPath, reportName) : Function called after generation with the result
    */
   renderCallback: function (templatePath, data, optionsRaw, callbackRaw) {
-    parseOptions(optionsRaw, callbackRaw, function (options, callback) {
+    if (!optionsRaw) optionsRaw = {};
+    input.parseOptions(optionsRaw, callbackRaw, function (options, callback) {
       // open the template (unzip if necessary)
       file.openTemplate(templatePath, function (err, template) {
         if (err) {
@@ -175,19 +198,26 @@ var carbone = {
         var _extension = file.detectType(template);
         // It takes the user defined one, or use the file type.
         options.extension = optionsRaw.extension || _extension;
-
         if (options.extension === null) {
           return callback('Unknown input file type. It should be a docx, xlsx, pptx, odt, ods, odp, xhtml, html or an xml file');
         }
-        if (options.extension === options.convertTo) {
-          options.convertTo = null; // avoid calling LibreOffice if the output file type is the same as the input file type
+        // check and clean convertTo object, options.convertTo contains a clean version of optionsRaw.convertTo
+        var _error = input.parseConvertTo(options, optionsRaw.convertTo);
+        if (_error) {
+          return callback(_error);
         }
         template.reportName = options.reportName;
         template.extension = options.extension;
-        preprocessor.execute(template, function (err, template) {
+
+
+        preprocessor.execute(template, options, function (err, template) {
           if (err) {
             return callback(err, null);
           }
+
+          // obtaining images from data for generation
+          let images = getImagesFromData(data);
+
           try {
             if (template.extension == 'xlsx') {
               // delete sheets containing pivot tables
@@ -206,20 +236,25 @@ var carbone = {
               rowsShift(template.files, data);
               // print area shift
               printAreaShift(template.files, data);
+
+
+
+              for (let fileIndex = 0; fileIndex < template.files.length; fileIndex++) {
+                template.files[fileIndex].data = preprocessor.removeRowCounterInWorksheet(template.files[fileIndex].data);
+              }
             }
 
           }
           catch (err) {
             return callback(err);
           }
-          // obtaining images from data for generation
-          let images = getImagesFromData(data);
+
+
           // parse all files of the template
           walkFiles(template, data, options, 0, function (err, report) {
             if (err) {
               return callback(err, null);
             }
-            // вставка изображений в книгу
             if (template.extension == 'xlsx') putImagesToSheets(report, images);
             if (template.extension == 'docx') putImagesToDocument(report, images);
             if (template.extension == 'docx') insertHTML(report, data);
@@ -229,28 +264,12 @@ var carbone = {
               if (err) {
                 return callback(err, null);
               }
-              if (typeof (options.convertTo) === 'string' && options.convertTo !== '') {
-                let optionsConvert = {
-                  extension: options.extension
-                };
-                convert(result, options.convertTo, optionsConvert, function (err, result) {
-                  callback(err, result, report.reportName);
-                });
-              }
-              else if (typeof (options.convertTo) === 'object'
-                && options.convertTo !== null
-                && typeof (options.convertTo.formatName) === 'string' && options.convertTo.formatName !== '') {
-                let optionsConvert = {
-                  formatOptions: options.convertTo.formatOptionsRaw ? options.convertTo.formatOptionsRaw : options.convertTo.formatOptions ? options.convertTo.formatOptions : {},
-                  extension: options.extension
-                };
-                convert(result, options.convertTo.formatName, optionsConvert, function (err, result) {
-                  callback(err, result, report.reportName);
-                });
-              }
-              else {
-                callback(null, result, report.reportName);
-              }
+              convert(result, report.reportName, options, function (err, bufferOrFile) {
+                if (report.reportName === undefined && typeof bufferOrFile === 'string') {
+                  report.reportName = path.basename(bufferOrFile);
+                }
+                callback(err, bufferOrFile, report.reportName, (options.isDebugActive === true ? options.debugInfo : null));
+              });
             });
           });
         });
@@ -284,27 +303,38 @@ var carbone = {
       return this.render(templatePath, data, callback);
     })
   },
-
-
   render: function (templatePath, data, optionsRaw, callbackRaw) {
     if (typeof optionsRaw === 'function' || typeof callbackRaw === 'function') {
       return this.renderCallback(templatePath, data, optionsRaw, callbackRaw);
     }
     return this.renderPromise(templatePath, data, optionsRaw);
   },
+  /**
+   * Decodes a rendered filename.
+   *
+   * When carbone.render is called with the options renderPrefix, the callback returns a path instead of a buffer
+   * The filename is built like this (3 distinct parts), with only alphanumeric characters to be able to write it on the disk safely
+   *
+   * <prefix><22-random-chars><encodedReportName.extension>
+   *
+   * This function decodes the part `<encodedReportName.extension>` `
+   *
+   * @param  {String}   pathOrFilename  The path or filename
+   * @param  {Integer}  prefixLength    The prefix length used in options.renderPrefix
+   * @return {Object}   {
+   *                      extension  : 'pdf',
+   *                      reportName : 'decoded filename'
+   *                    }
+   */
+  decodeRenderedFilename: function (pathOrFilename, prefixLength = 0) {
+    var _filename = path.basename(pathOrFilename);
+    var _extension = path.extname(_filename);
+    var _onlyReportName = _filename.slice(prefixLength + helper.RANDOM_STRING_LENGTH, -_extension.length);
 
-  renderParallel: function (templatePath, data, optionsRaw) {
-    return new Promise((resolve, reject) => {
-      const worker = new Worker(__filename, {
-        workerData: { templatePath, data, optionsRaw },
-      });
-      worker.on('message', resolve);
-      worker.on('error', reject);
-      worker.on('exit', (code) => {
-        if (code !== 0)
-          reject(new Error(`Worker stopped with exit code ${code}`));
-      });
-    });
+    return {
+      reportName: helper.decodeSafeFilename(_onlyReportName),
+      extension: _extension.slice(1)
+    };
   },
 
   /**
@@ -330,24 +360,31 @@ var carbone = {
 
   /**
    * Convert a file format to another
+   *
    * @param  {Buffer}   data      raw data returned by fs.readFile (no utf-8)
-   * @param  {String}   convertTo output format. Example: 'csv', 'xls', 'xlsx'
-   * @param  {Object}   options   Example for CSV files:
+   * @param  {Object}   options   Same as carbone.render
    *                              {
-   *                                formatOptions : {
-   *                                  fieldSeparator  : ',',
-   *                                  textDelimiter   : '"',
-   *                                  characterSet    : '76',   // utf8  https://wiki.openoffice.org/wiki/Documentation/DevGuide/Spreadsheets/Filter_Options
-   *                                }
-   *                                extension       : 'csv'   // extension of source file returned by path.extname(filename). It helps LibreOffice to understand the source format (mandatory for CSV files)
+   *                                convertTo : pdf || {formatName, formatOptions}
+   *                                extension : 'csv'  // extension of input file returned by path.extname(filename).
+   *                                                   // It helps LibreOffice to understand the source format (mandatory for CSV files)
    *                              }
    * @param  {Function} callback  (err, result) result is a buffer (file converted)
    */
-  convert: function (data, convertTo, options, callback) {
-    convert(data, convertTo, options, callback);
-  }
+  convert: function (fileBuffer, optionsRaw, callbackRaw) {
+    input.parseOptions(optionsRaw, callbackRaw, function (options, callback) {
+      options.extension = optionsRaw.extension;
+      if (options.extension === null) {
+        return callback('Unknown input file type. options.extension should be equals to docx, xlsx, pptx, odt, ods, odp, xhtml, html or an xml');
+      }
+      var _error = input.parseConvertTo(options, optionsRaw.convertTo);
+      if (_error) {
+        return callback(_error);
+      }
+      convert(fileBuffer, undefined, options, callback);
+    });
+  },
 
-
+  formatters: input.formatters
 };
 
 /** ***************************************************************************************************************/
@@ -514,14 +551,16 @@ function scaleFunctionData(files, data) {
             newSplittedPos[1] = newSplittedPos[1].replace(to.row + offset, to.row + offset + scaleNumber);
           }
           let newPos = newSplittedPos.join(':');
-          let formulaCalcRelation = (files[calcChainFileIndex].data.match(new RegExp(`<c r="${formula.cell}".*?i="${sheetFileIndex.sheetId}".*?\/>`)) || [])[0];
-          if (formulaCalcRelation) {
-            let formulaRelPosition = getNumberFromXlsxPosition(formula.cell);
-            let previousCalcChainInterpolations = interpolations.filter(interpolation => interpolation.value && interpolation.number + 1 < formulaRelPosition.row);
-            let calcChainInterpolationsOffset = 0;
-            previousCalcChainInterpolations.forEach(prevInterpolation => calcChainInterpolationsOffset += data[prevInterpolation.value].length - 2);
-            let newFormulaCalcRelation = formulaCalcRelation.replace(formula.cell, formula.cell.replace(formulaRelPosition.row, formulaRelPosition.row + calcChainInterpolationsOffset));
-            files[calcChainFileIndex].data = files[calcChainFileIndex].data.replace(formulaCalcRelation, newFormulaCalcRelation);
+          if (files[calcChainFileIndex] && files[calcChainFileIndex].data) {
+            let formulaCalcRelation = (files[calcChainFileIndex].data.match(new RegExp(`<c r="${formula.cell}".*?i="${sheetFileIndex.sheetId}".*?\/>`)) || [])[0];
+            if (formulaCalcRelation) {
+              let formulaRelPosition = getNumberFromXlsxPosition(formula.cell);
+              let previousCalcChainInterpolations = interpolations.filter(interpolation => interpolation.value && interpolation.number + 1 < formulaRelPosition.row);
+              let calcChainInterpolationsOffset = 0;
+              previousCalcChainInterpolations.forEach(prevInterpolation => calcChainInterpolationsOffset += data[prevInterpolation.value].length - 2);
+              let newFormulaCalcRelation = formulaCalcRelation.replace(formula.cell, formula.cell.replace(formulaRelPosition.row, formulaRelPosition.row + calcChainInterpolationsOffset));
+              files[calcChainFileIndex].data = files[calcChainFileIndex].data.replace(formulaCalcRelation, newFormulaCalcRelation);
+            }
           }
           files[sheetFileIndex.sheetFileIndex].data = files[sheetFileIndex.sheetFileIndex].data.replace(pos, newPos);
         });
@@ -1148,43 +1187,8 @@ function getImagePosition(sheet, imagePlaceHolder) {
   return imageParams;
 }
 
-function parseOptions(options, callbackFn, callback) {
-  if (typeof (options) === 'function') {
-    callbackFn = options;
-    options = {};
-  }
-  // analyze pre-declared variables in the Object "options"
-  parser.findVariables(options.variableStr, function (err, str, variables) {
-    var _options = {
-      enum: options.enum,
-      currency: {},
-      lang: (options.lang || params.lang).toLowerCase(), // TODO test toLowerCase
-      translations: options.translations || params.translations,
-      complement: options.complement,
-      reportName: options.reportName,
-      convertTo: options.convertTo,
-      extension: options.extension,
-      formatters: carbone.formatters,
-      existingVariables: variables
-    };
-    var _currency = _options.currency;
-    var _locale = locale[_options.lang] || locale.en;
-    var _currencyFromLocale = _locale.currency.code;
-    _currency.source = options.currencySource || params.currencySource;
-    _currency.target = options.currencyTarget || params.currencyTarget;
-    _currency.rates = options.currencyRates || params.currencyRates;
-    if (!_currency.source) {
-      _currency.source = _currencyFromLocale;
-    }
-    if (!_currency.target) {
-      _currency.target = _currencyFromLocale;
-    }
-    return callback(_options, callbackFn);
-  });
-}
-
 /**
- * Parse and compute xml for all files of the template
+ * Parse and compute XML for all files of the template
  * @param  {Object}   template     template file returned by file.js
  * @param  {Object}   data         data to insert
  * @param  {Object}   options      {'complement', 'variables', ...}
@@ -1222,105 +1226,93 @@ function walkFiles(template, data, options, currentIndex, callback) {
   }
 }
 
-/**
- * Extract the option values and convert it to LibreOffice format
- * https://wiki.openoffice.org/wiki/Documentation/DevGuide/Spreadsheets/Filter_Options
- * @param  {String || Object} data
- * @return {String}
- */
-function extractValuesFromOptions(data) {
-  // most of the case, there is no options
-  if (data === '' || data === null || data === undefined) {
-    data = {};
-  }
-  /*
-    Default values
-      Separator : ','
-      Delimiter : '"'
-      Encoding  : UTF-8
-   */
-  var _optionsArray = ['44', '34', '76'];
-  if (typeof (data) === 'string') {
-    var _dataToArray = data.split(',');
-    for (var i = 0; i < _dataToArray.length; i++) {
-      _optionsArray[i] = _dataToArray[i].trim();
-    }
-  }
-  else if (typeof (data) === 'object') {
-    _optionsArray[0] = (data.fieldSeparator || ',').charCodeAt(0);
-    _optionsArray[1] = (data.textDelimiter || '"').charCodeAt(0);
-    _optionsArray[2] = data.characterSet || '0';
-    if (data.numberOfFirstLine) {
-      _optionsArray[3] = data.numberOfFirstLine;
-    }
-    if (data.cellFormat) {
-      _optionsArray[4] = data.cellFormat;
-    }
-  }
-  return _optionsArray.join(',');
-}
 
 /**
- * Verify the files extension type based on the format.js
+ * { function_description }
  *
- * @param  {String} extensionIn extension type coming from the template file.
- * @param  {String} extensionOut extension type expected to be converted.
- * @return {Boolean} Return true if the extensions format are not matching, otherwhise false.
+ * @param {Buffer}    inputFileBuffer  The input file buffer
+ * @param {String}    customReportName user-defined report name computed by Carbone
+ * @param {Function}  options          options coming from input.parseOptions and input.parseConvertTo
+ *                                     {
+ *                                       convertTo : {
+ *                                         extension  : 'pdf',
+ *                                         format     : 'writer_pdf_Export' // coming from lib/format.js
+ *                                         optionsStr : '44,34,76',         // only for CSV
+ *                                         filters    : {                   // only for PDF, JPG, ...
+ *                                           ReduceImageResolution : true
+ *                                         }
+ *                                       }
+ *                                       extension    : 'odt' || Force input template extension
+ *                                       hardRefresh  : (default: false) if true, LibreOffice is used to render and refresh the content of the report at the end of Carbone process
+ *                                       renderPrefix : If defined, it add a prefix to the report name
+ *                                     }
+ * @param {Function}  callback(err, bufferOrPath) return a path if renderPrefix is defined, a buffer otherwise
  */
-function checkDocTypeAndOutputExtensions(extensionIn, extensionOut) {
-  for (var docType in format) {
-    if (format[docType][extensionIn] !== undefined
-      && format[docType][extensionOut] !== undefined) {
-      return docType;
-    }
-  }
-  return null;
-}
+function convert(inputFileBuffer, customReportName, options, callback) {
+  const _convertTo = options.convertTo;
+  const _hasConversion = options.extension !== _convertTo.extension || options.hardRefresh === true;
+  const _isReturningBuffer = options.renderPrefix === undefined || options.renderPrefix === null;
 
-/**
- * Convert a file using LibreOffice.
- * @param  {Buffer|String}   data
- * @param  {String}          convertTo
- * @param  {Object}          options (optional) {
- *                                formatName', 'formatOptions'}
- * @param  {Function}        callback(err, data)
- */
-function convert(data, convertTo, options, callback) {
-  if (typeof (options) === 'function') {
-    callback = options;
-    options = {};
+  // If there is no conversion, and there is no renderPrefix, we return the buffer directly
+  if (_hasConversion === false && _isReturningBuffer === true) {
+    return callback(null, inputFileBuffer);
   }
 
-  if (!options.extension) {
-    return callback('options.extension must be set to detect input file type', null);
-  }
-  const docType = checkDocTypeAndOutputExtensions(options.extension, convertTo);
-  if (docType === null) {
-    return callback(`Format "${options.extension}" can't be converted to "${convertTo}".`);
-  }
-  var _fileFormatOutput = format[docType][convertTo].format;
-  var _filePathInput = path.join(params.tempPath, helper.getUID()) + '.' + options.extension;
-  var _formatOptions = extractValuesFromOptions(options.formatOptions);
+  // generate a unique random & safe filename
+  const _renderPrefix = (options.renderPrefix || '').replace(/[^0-9a-z-]/gi, '');
+  const _randomNamePart = helper.getRandomString();
+  const _customReportName = customReportName !== undefined ? customReportName : 'report';
+  const _renderFilename = _renderPrefix + _randomNamePart + helper.encodeSafeFilename(_customReportName) + '.' + _convertTo.extension;
+  const _renderFile = path.join(params.renderPath, _renderFilename);
 
-  fs.writeFile(_filePathInput, data, function (err) {
+  // no conversion, but return a path
+  if (_hasConversion === false) {
+    return fs.writeFile(_renderFile, inputFileBuffer, function (err) {
+      if (err) {
+        debug('Cannot write rendered file on disk' + err);
+        return callback('Cannot write rendered file on disk', null);
+      }
+      return callback(null, _renderFile);
+    });
+  }
+
+  // A conversion is necessary, generate a intermediate file for the converter
+  const _intermediateFilename = _renderPrefix + _randomNamePart + '_tmp.' + options.extension;
+  const _intermediateFile = path.join(params.renderPath, _intermediateFilename);
+  fs.writeFile(_intermediateFile, inputFileBuffer, function (err) {
     if (err) {
-      return callback(err);
+      debug('Cannot write rendered file on disk' + err);
+      return callback('Cannot write rendered file on disk', null);
     }
-    converter.convertFile(_filePathInput, _fileFormatOutput, _formatOptions, function (errConvert, data) {
-      fs.unlink(_filePathInput, function (err) {
-        debug('cannot remove file ' + err);
+    // call the converter and tell him to generate directly the wanted filename
+    converter.convertFile(_intermediateFile, _convertTo.format, _convertTo.optionsStr, _renderFile, function (errConvert, outputFile) {
+      fs.unlink(_intermediateFile, function (err) {
+        if (err) {
+          debug('Cannot remove intermediate file before conversion ' + err);
+        }
       });
-      callback(errConvert, data);
+      if (errConvert) {
+        return callback(errConvert, null);
+      }
+      if (_isReturningBuffer === false) {
+        return callback(null, outputFile);
+      }
+      fs.readFile(outputFile, function (err, outputBuffer) {
+        fs.unlink(outputFile, function (err) {
+          if (err) {
+            debug('Cannot remove rendered file ' + err);
+          }
+        });
+        if (err) {
+          debug('Cannot returned file buffer ' + err);
+          return callback('Cannot returned file buffer', null);
+        }
+        callback(null, outputBuffer);
+      });
     });
   });
 }
 
-if (workerData && workerData.templatePath) {
-  carbone.render(workerData.templatePath, workerData.data, workerData.optionsRaw)
-    .then(result => {
-      parentPort.postMessage(result);
-    })
-}
 
 // add default formatters
 carbone.addFormatters(require('../formatters/array.js'));
@@ -1329,4 +1321,170 @@ carbone.addFormatters(require('../formatters/date.js'));
 carbone.addFormatters(require('../formatters/number.js'));
 carbone.addFormatters(require('../formatters/string.js'));
 translator.loadTranslations(params.templatePath);
+
+// We must include all locales like this for PKG
+require('dayjs/locale/af.js');
+require('dayjs/locale/am.js');
+require('dayjs/locale/ar-dz.js');
+require('dayjs/locale/ar-kw.js');
+require('dayjs/locale/ar-ly.js');
+require('dayjs/locale/ar-ma.js');
+require('dayjs/locale/ar-sa.js');
+require('dayjs/locale/ar-tn.js');
+require('dayjs/locale/ar.js');
+require('dayjs/locale/az.js');
+require('dayjs/locale/be.js');
+require('dayjs/locale/bg.js');
+require('dayjs/locale/bi.js');
+require('dayjs/locale/bm.js');
+require('dayjs/locale/bn.js');
+require('dayjs/locale/bo.js');
+require('dayjs/locale/br.js');
+require('dayjs/locale/bs.js');
+require('dayjs/locale/ca.js');
+require('dayjs/locale/cs.js');
+require('dayjs/locale/cv.js');
+require('dayjs/locale/cy.js');
+require('dayjs/locale/da.js');
+require('dayjs/locale/de-at.js');
+require('dayjs/locale/de-ch.js');
+require('dayjs/locale/de.js');
+require('dayjs/locale/dv.js');
+require('dayjs/locale/el.js');
+require('dayjs/locale/en-au.js');
+require('dayjs/locale/en-ca.js');
+require('dayjs/locale/en-gb.js');
+require('dayjs/locale/en-ie.js');
+require('dayjs/locale/en-il.js');
+require('dayjs/locale/en-in.js');
+require('dayjs/locale/en-nz.js');
+require('dayjs/locale/en-sg.js');
+require('dayjs/locale/en-tt.js');
+require('dayjs/locale/en.js');
+require('dayjs/locale/eo.js');
+require('dayjs/locale/es-do.js');
+require('dayjs/locale/es-pr.js');
+require('dayjs/locale/es-us.js');
+require('dayjs/locale/es.js');
+require('dayjs/locale/et.js');
+require('dayjs/locale/eu.js');
+require('dayjs/locale/fa.js');
+require('dayjs/locale/fi.js');
+require('dayjs/locale/fo.js');
+require('dayjs/locale/fr-ca.js');
+require('dayjs/locale/fr-ch.js');
+require('dayjs/locale/fr.js');
+require('dayjs/locale/fy.js');
+require('dayjs/locale/ga.js');
+require('dayjs/locale/gd.js');
+require('dayjs/locale/gl.js');
+require('dayjs/locale/gom-latn.js');
+require('dayjs/locale/gu.js');
+require('dayjs/locale/he.js');
+require('dayjs/locale/hi.js');
+require('dayjs/locale/hr.js');
+require('dayjs/locale/ht.js');
+require('dayjs/locale/hu.js');
+require('dayjs/locale/hy-am.js');
+require('dayjs/locale/id.js');
+require('dayjs/locale/is.js');
+require('dayjs/locale/it-ch.js');
+require('dayjs/locale/it.js');
+require('dayjs/locale/ja.js');
+require('dayjs/locale/jv.js');
+require('dayjs/locale/ka.js');
+require('dayjs/locale/kk.js');
+require('dayjs/locale/km.js');
+require('dayjs/locale/kn.js');
+require('dayjs/locale/ko.js');
+require('dayjs/locale/ku.js');
+require('dayjs/locale/ky.js');
+require('dayjs/locale/lb.js');
+require('dayjs/locale/lo.js');
+require('dayjs/locale/lt.js');
+require('dayjs/locale/lv.js');
+require('dayjs/locale/me.js');
+require('dayjs/locale/mi.js');
+require('dayjs/locale/mk.js');
+require('dayjs/locale/ml.js');
+require('dayjs/locale/mn.js');
+require('dayjs/locale/mr.js');
+require('dayjs/locale/ms-my.js');
+require('dayjs/locale/ms.js');
+require('dayjs/locale/mt.js');
+require('dayjs/locale/my.js');
+require('dayjs/locale/nb.js');
+require('dayjs/locale/ne.js');
+require('dayjs/locale/nl-be.js');
+require('dayjs/locale/nl.js');
+require('dayjs/locale/nn.js');
+require('dayjs/locale/oc-lnc.js');
+require('dayjs/locale/pa-in.js');
+require('dayjs/locale/pl.js');
+require('dayjs/locale/pt-br.js');
+require('dayjs/locale/pt.js');
+require('dayjs/locale/ro.js');
+require('dayjs/locale/ru.js');
+require('dayjs/locale/rw.js');
+require('dayjs/locale/sd.js');
+require('dayjs/locale/se.js');
+require('dayjs/locale/si.js');
+require('dayjs/locale/sk.js');
+require('dayjs/locale/sl.js');
+require('dayjs/locale/sq.js');
+require('dayjs/locale/sr-cyrl.js');
+require('dayjs/locale/sr.js');
+require('dayjs/locale/ss.js');
+require('dayjs/locale/sv.js');
+require('dayjs/locale/sw.js');
+require('dayjs/locale/ta.js');
+require('dayjs/locale/te.js');
+require('dayjs/locale/tet.js');
+require('dayjs/locale/tg.js');
+require('dayjs/locale/th.js');
+require('dayjs/locale/tk.js');
+require('dayjs/locale/tl-ph.js');
+require('dayjs/locale/tlh.js');
+require('dayjs/locale/tr.js');
+require('dayjs/locale/tzl.js');
+require('dayjs/locale/tzm-latn.js');
+require('dayjs/locale/tzm.js');
+require('dayjs/locale/ug-cn.js');
+require('dayjs/locale/uk.js');
+require('dayjs/locale/ur.js');
+require('dayjs/locale/uz-latn.js');
+require('dayjs/locale/uz.js');
+require('dayjs/locale/vi.js');
+require('dayjs/locale/yo.js');
+require('dayjs/locale/zh-cn.js');
+require('dayjs/locale/zh-hk.js');
+require('dayjs/locale/zh-tw.js');
+require('dayjs/locale/zh.js');
+
+// if DayJS does not have a locale defined with country code, define it
+// For example "de-de" does not exists in DaysJS, but "de" exists.
+// So add locale "de-de" in DaysJS
+for (let _locale in locales) {
+  if (dayjs.Ls[_locale] === undefined) {
+    let _localeWithoutCountry = _locale.replace(/-\S+/g, '');
+    if (dayjs.Ls[_localeWithoutCountry] !== undefined) {
+      dayjs.locale(_locale, dayjs.Ls[_localeWithoutCountry]);
+    }
+  }
+}
+
+dayjs.extend(require('dayjs/plugin/advancedFormat'));    // Support Quarter, ...
+dayjs.extend(require('dayjs/plugin/localizedFormat'));   // Support L LL LLLL
+dayjs.extend(require('dayjs/plugin/customParseFormat')); // Support custom format as input
+dayjs.extend(require('dayjs/plugin/utc'));
+dayjs.extend(require('dayjs-timezone-iana-plugin'));
+dayjs.extend(require('dayjs/plugin/isoWeek'));
+
+dayjs.tz.setDefault('Europe/Paris');
+dayjs.locale('en');
+
+if (fs.existsSync(params.renderPath) === false) {
+  fs.mkdirSync(params.renderPath, '0755');
+}
+
 module.exports = carbone;
