@@ -2,7 +2,7 @@ const params = require('./params');
 const format = require('./format');
 const parser = require('./parser');
 const locale = require('../formatters/_locale.js');
-const formatters = {};
+const formatters = Object.create(null); // Remove  __proto__ and constructor attributes. Mitigates prototype pollution attacks.
 
 /**
  * Parse options coming from user-side. Clean it and generate a safe options object for internal use
@@ -41,7 +41,9 @@ function parseOptions (userDefinedOptions, callbackFn, callback) {
       formatters        : formatters,
       existingVariables : variables,
       hardRefresh       : userDefinedOptions.hardRefresh || false,
-      renderPrefix      : userDefinedOptions.renderPrefix
+      isDebugActive     : userDefinedOptions.isDebugActive || false,     // show info about the template
+      renderPrefix      : userDefinedOptions.renderPrefix,
+      debugInfo         : { markers : [] } // filled and returned after POST /render if isDebugActive === true
     };
     var _currency           = _options.currency;
     var _locale             = locale[_options.lang] || locale.en;
@@ -112,6 +114,10 @@ function parseConvertTo (options, userDefinedConvertTo) {
     case 'pdf':
       // set options.convertTo.filters and return null of there is no error
       return checkAndSetOptionsForPDF(_optionsOutput, options.convertTo);
+    case 'png':
+    case 'jpg':
+      // set options.convertTo.filters and return null of there is no error
+      return checkAndSetOptionsForImage(_optionsOutput, options.convertTo);
     default:
       break;
   }
@@ -168,6 +174,18 @@ function checkAndSetOptionsForPDF (userDefinedPDFFilters, convertToObj) {
     userDefinedPDFFilters.ReduceImageResolution = false;
   }
   convertToObj.filters = userDefinedPDFFilters;
+  return null;
+}
+
+/**
+ * Parse options of userDefinedImageFilters coming from user side and update internal options.convertTo
+ *
+ * @param      {Object}  userDefinedImageFilters  Image filters
+ * @param      {Object}  convertToObj             updates convertTo.optionsStr
+ * @return     {String}                           null of there is no error, a string if there is an error
+ */
+function checkAndSetOptionsForImage (userDefinedImageFilters, convertToObj) {
+  convertToObj.filters = userDefinedImageFilters;
   return null;
 }
 
