@@ -1,4 +1,5 @@
 var os = require('os');
+var path = require('path');
 
 // manage node 0.8 / 0.10 differences
 var nodeVersion = process.versions.node.split('.');
@@ -6,32 +7,43 @@ var tmpDir = (parseInt(nodeVersion[0], 10) === 0 && parseInt(nodeVersion[1], 10)
 
 module.exports  = {
   /* Temp directory */
-  tempPath         : tmpDir,
+  tempPath                : tmpDir,
+  /* Sub directory automatically created in tempPath to store temporal render */
+  pythonPath              : path.join(__dirname, 'converter.py'),
+  renderPath              : path.join(tmpDir, 'carbone_render'),
   /* Template path */
-  templatePath     : process.cwd(),
+  templatePath            : process.cwd(),
   /* Number of LibreOffice + Python factory to start by default. One factory = 2 threads */
-  factories        : 1,
+  factories               : 1,
   /* If LibreOffice fails to convert one document, how many times we re-try to convert this file? */
-  attempts         : 3,
+  attempts                : 1,
   /* If true, it will start LibreOffice + Python factory immediately (true by default if the carbone server is used).
      If false, it will start LibreOffice + Python factory only when at least one document conversion is needed.*/
-  startFactory     : false,
+  startFactory            : false,
+  /* approximated LibreOffice memory leak per report convered, unit: MegaBytes. Set the value to 0 to disable */
+  factoryMemoryFileSize   : 1,
+  /* max percentage of memory used by one LibreOffice process. Set the value to 0 to disable */
+  factoryMemoryThreshold  : 50,
+  /* Timeout used to kill a factory if a report is converting for a long time. Set the value to 0 to disable. (unit: ms) */
+  converterFactoryTimeout : 60000,
   /* The method helper.getUID() add this prefix in the uid */
-  uidPrefix        : 'c',
+  uidPrefix               : 'c',
   /* If multiple factories are used, the pipe name is generated automatically to avoid conflicts */
-  pipeNamePrefix   : '_carbone',
+  pipeNamePrefix          : '_carbone',
   /* list of file parsed for translation and find tools */
-  extensionParsed  : '(odt|ods|odp|xlsx|docx|pptx|xml|html)',
-  /* lang of carbone and dayjs.js */
-  lang             : 'en',
+  extensionParsed         : '(odt|ods|odp|xlsx|docx|pptx|xml|html)',
+  /* lang of carbone and moment.js */
+  lang                    : 'en',
+  /* default output timezone for dates */
+  timezone                : 'Europe/Paris',
   /* all locales are loaded in memory */
-  translations     : {},
+  translations            : {},
   /* currency of data, it depends on the locale if empty */
-  currencySource   : '',
+  currencySource          : '',
   /* default target currency when the formatter convCurr is used without target. It depends on the locale if empty */
-  currencyTarget   : '',
+  currencyTarget          : '',
   /* currency rates, always based on EUR. So EUR should always equals "1" */
-  currencyRates    : {
+  currencyRates           : {
     EUR : 1,
     USD : 1.1403,
     JPY : 123.20,
@@ -67,4 +79,3 @@ module.exports  = {
     ZAR : 16.1175
   }
 };
-
