@@ -1,6 +1,18 @@
 var helper = require('./helper');
+var condition = require('../formatters/condition');
+const MARKER_PRESENCE_CHAR = '\uFFFF';
 
 var parser = {
+
+  /**
+   * Determines whether the specified one marker is a carbone marker.
+   *
+   * @param {String}  oneMarker  One marker
+   */
+  isCarboneMarker : function (oneMarker) {
+    const _isCarboneMarkerRegexp = /^\{?\s*(?:[cdt]\s*[.[:(])|^\{?\s*[#$]|^\{?\s*bindColor/;
+    return _isCarboneMarkerRegexp.test(oneMarker);
+  },
 
   /**
    * Find all markers in the xml.
@@ -16,7 +28,54 @@ var parser = {
     var that = this;
     var _allMarkers = [];
     var _previousMarkerLength = 0;
-    var _isCarboneMarkerRegexp = /^\{?\s*(?:[cd]\s*[.[:])|\$/;
+    var _previousMarkerPos = -1;
+    var _nbMarkerPresenceChar = 0;
+    var _markerPresenceChar = MARKER_PRESENCE_CHAR;
+    var _cleanedXml = xml;
+
+    // Capture { } markers again and extract them
+    _cleanedXml = _cleanedXml.replace(/\{([^{]+?)\}/g, function (m, marker, offset) {
+      if (that.isCarboneMarker(marker) === false) {
+        return m;
+      }
+      var _cleanedMarker = that.removeWhitespace(that.cleanMarker(marker));
+      _markerPresenceChar = '';
+      if (condition._isConditionalBlockEndMarker(_cleanedMarker) === false && condition._isConditionalBlockBeginMarker(_cleanedMarker) === false) {
+        // Replace the marker by a special character for all markers, except begin/end conditional blocks. This character is removed after.
+        // This is the simplest method to manage this case "<p>{d.id}{d.id:showBegin}a</p>{d.id:showEnd}" -> "<p>\uFFFFa</p>"
+        // It prevents the if-block system deleting the whole paragraph "<p></p>" when there is a marker just after <p>
+        ++_nbMarkerPresenceChar;
+        _markerPresenceChar = MARKER_PRESENCE_CHAR;
+      }
+      var _pos = offset - _previousMarkerLength + _nbMarkerPresenceChar;
+      if (_pos === Math.trunc(_previousMarkerPos)) {
+        // Avoid exactly the same XML position to sort all parts at the end of the process
+        // Why adding 1/64? to avoid rounding problems of floats (http://0.30000000000000004.com)
+        // It let us the possibility to have 64 markers at the same position. It should be enough for all cases.
+        _pos = _previousMarkerPos + 1/64;
+      }
+      var _obj = {
+        pos  : _pos,
+        name : '_root.'+_cleanedMarker
+      };
+      _allMarkers.push(_obj);
+      _previousMarkerLength += (marker.length + 2);  // 2 equals the number of braces '{' or '}'
+      _previousMarkerPos = _pos;
+      return _markerPresenceChar;
+    });
+    process.nextTick(function () {
+      callback(null, _cleanedXml, _allMarkers);
+    });
+  },
+  /**
+   * @description Remove XML tags inside markers
+   * @param   {String} xml
+   * @returns {String} Return a clean XML
+   */
+  removeXMLInsideMarkers : function (xml) {
+    if (!xml || typeof xml !== 'string') {
+      return xml;
+    }
     var _cleanedXml = xml.replace(/(\r\n|\n|\r)/g,' ');
     // Markers inside xml tag break the system which find markers because we can have nested markers.
     // So we find all xml tags and replace { } which are inside xml by a temporary character
@@ -28,11 +87,15 @@ var parser = {
     _cleanedXml = _cleanedXml.replace(/(\{[^{]+?\})/g, function (m, markerWithXml) {
       // is this marker contains xml? like this {d. <xml> <tr> menu}  </xml> </tr>
       if (/</.test(markerWithXml) === true) {
-        let _markerOnly = that.extractMarker(markerWithXml);
-        let _xmlOnly    = that.cleanXml(markerWithXml);
+        let _markerOnly = parser.extractMarker(markerWithXml);
+        let _xmlOnly    = parser.cleanXml(markerWithXml);
         // verify if it is really a Carbone marker, otherwise do nothing
-        if (_isCarboneMarkerRegexp.test(_markerOnly) === false) {
+        if (parser.isCarboneMarker(_markerOnly) === false) {
           return m;
+        }
+        // put the conditional block beginEnd/hideEnd markers at the end of the brace to remove everything in-between
+        if (condition._isConditionalBlockEndMarker(_markerOnly) === true) {
+          return _xmlOnly + _markerOnly;
         }
         // separate clearly the marker and the xml
         return _markerOnly + _xmlOnly;
@@ -42,27 +105,8 @@ var parser = {
     // put back markers inside XML
     // eslint-disable-next-line no-control-regex
     _cleanedXml = _cleanedXml.replace(/\u0000/g, '{').replace(/\uFFFF/g, '}');
-
-    // Capture { } markers again and extract them
-    _cleanedXml = _cleanedXml.replace(/\{([^{]+?)\}/g, function (m, marker, offset) {
-      if (_isCarboneMarkerRegexp.test(marker) === false) {
-        return m;
-      }
-      var _cleanedMarker = that.removeWhitespace(that.cleanMarker(marker));
-      var _obj = {
-        pos  : offset-_previousMarkerLength,
-        name : '_root.'+_cleanedMarker
-      };
-      _allMarkers.push(_obj);
-      _previousMarkerLength += (marker.length + 2);  // 2 equals the number of braces '{' or '}'
-      return '';
-    }).replace(/'/g, '\\\'');
-
-    process.nextTick(function () {
-      callback(null, _cleanedXml, _allMarkers);
-    });
+    return _cleanedXml;
   },
-
   /**
    * Extract markers from the xml
    *
@@ -169,12 +213,12 @@ var parser = {
    *                              'lang'         : selected lang: 'fr',
    *                              'translations' : all translations: {fr: {}, en: {}, es: {}}
    *                            }
-   * @param  {Function} callback(err, xmlTranslated) : xml with all marker {t()} replaced by the traduction
+   * @param  {Function} callback Optional callback (err, xmlTranslated) : xml with all marker {t()} replaced by the traduction
    */
   translate : function (xml, options, callback) {
     var _translations = options.translations !== undefined ? options.translations[options.lang] : {};
     if (typeof(xml) !== 'string') {
-      return callback(null, xml);
+      return callback ? callback(null, xml) : xml;
     }
     // capture {t  } markers. The part "(?:\s*<[^>]*>\s*)*" is used to eleminate xml tags
     var _cleanedXml = xml.replace(/\{\s*((?:\s*<[^>]*>\s*)*t[\s\S]+?)\}/g, function (m, text) {
@@ -200,7 +244,7 @@ var parser = {
         return m;
       }
     });
-    callback(null, _cleanedXml);
+    return callback ? callback(null, _cleanedXml) : _cleanedXml;
   },
 
   /**
@@ -241,7 +285,7 @@ var parser = {
             _sortedParams.push({index : i, name : _params[i]});
           }
           _sortedParams.sort(function (a,b) {
-            return a.name.length < b.name.length;
+            return b.name.length - a.name.length;
           });
           // replace all parameters by _$0_, _$1_, _$2_, ..., in the code
           for (var p = 0; p < _sortedParams.length; p++) {
@@ -249,7 +293,7 @@ var parser = {
             _code = _code.replace(new RegExp('\\'+_param.name,'g'), '_$'+_param.index+'_');
           }
         }
-        var _regex = new RegExp('\\$'+_variable+'(?:\\(([\\s\\S]+?)\\))?', 'g');
+        var _regex = new RegExp('\\$'+_variable+'(?!\\w)(?:\\(([\\s\\S]+?)\\))?', 'g');
         existingVariables.push({
           name  : _variable,
           code  : _code,
@@ -360,7 +404,7 @@ var parser = {
     var _prevTagPosEnd = 0;
     var _openingTagIndex = {};
     var _lastOpeningTagIndexBeforeStop = {};
-    var _similarTag =  new RegExp('<(/)?\\b[^/|^>]*(/)?>', 'g');
+    var _similarTag =  new RegExp('<(/)?[^>]+?(/)?>', 'g');
     var _tag = _similarTag.exec(leftSideXml);
     while (_tag !== null) {
       if (_tag[1]==='/') {
@@ -402,7 +446,7 @@ var parser = {
     var _startTagCount = 0;
     var _endTagCount = 0;
     var _endTagPos = -1;
-    var _similarTag =  new RegExp('<(/)?\\b[^/|^>]*(/)?>', 'g'); // reset the regex
+    var _similarTag =  new RegExp('<(/)?[^>]+?(/)?>', 'g'); // reset the regex
     var _tag = _similarTag.exec(rightSideXml);
     while (_tag !== null) {
       if (_tag[1]==='/') {
@@ -556,110 +600,239 @@ var parser = {
   },
 
   /**
+   * Transform XML in a flat array of objects
+   *
+   * @param  {String} xml
+   * @return {Array}
+   */
+  flattenXML : function (xml) {
+    var _array = [];
+    if (typeof (xml) !== 'string') {
+      return [];
+    }
+    var _xmlTagRegExp = /<([^>]+)>/g;
+    var _tag = _xmlTagRegExp.exec(xml);
+    var _depth = 0;
+    var _prevLastIndex = 0;
+    var _openingTagPos = [];
+    if (_tag === null) {
+      _array.push({
+        id        : 0,
+        index     : 0,
+        lastIndex : xml.length,
+        depth     : 0
+      });
+    }
+    while (_tag !== null) {
+      if ( _prevLastIndex < _tag.index) {
+        // non XML part
+        _array.push({
+          id        : _array.length,
+          index     : _prevLastIndex,
+          lastIndex : _tag.index,
+          depth     : _depth
+        });
+      }
+      var _xmlTag = {
+        id        : _array.length,
+        index     : _tag.index,
+        lastIndex : _xmlTagRegExp.lastIndex,
+        depth     : _depth
+      };
+      // closing tag
+      if (_tag[1].startsWith('/') === true) {
+        _xmlTag.depth = _depth--;
+        var _correspondingOpenPos = _openingTagPos.pop();
+        if (_correspondingOpenPos === undefined) {
+          throw new Error('XML not valid');
+        }
+        _correspondingOpenPos.match = _xmlTag.id;
+        _xmlTag.match = _correspondingOpenPos.id;
+      }
+      // is not a self-closing tag, so it is an opening tag
+      else if (_tag[1].endsWith('/') === false) {
+        _xmlTag.depth = _depth++;
+        _xmlTag.match = -1; // temp
+        _openingTagPos.push(_xmlTag);
+      }
+      _array.push(_xmlTag);
+      _prevLastIndex = _xmlTagRegExp.lastIndex;
+      _tag = _xmlTagRegExp.exec(xml);
+    }
+    return _array;
+  },
+
+  /**
+   * Detect and move the if-block beginning and ending if possible
+   *
+   * For example "<p>{d.id:showBegin}blabla</p>{d.id:showEnd}" becomes "{d.id:showBegin}<p>blabla</p>{d.id:showEnd}"
+   * It pushes/pulls the begin/end beyond the limit in order to include surrounded XML tags
+   *
+   * @param  {Array}   xmlTree              List of all XML tag, returned by parser.flattenXML
+   * @param  {Integer} beginFlattenXMLIndex where is the beginTextIndex in the xmlTree
+   * @param  {Integer} beginTextIndex       char position of if-block's beginning
+   * @param  {Integer} endTextIndex         char position of if-block's ending
+   * @return {Object}                       new begin/end position for if-block
+   */
+  moveConditionalBlockBeginEnd : function (xmlTree, beginFlattenXMLIndex, beginTextIndex, endTextIndex) {
+    const _dummyObject     = {};  // constant, empty object to avoid instanciation in the loop below
+    var _newBeginTreeIndex = beginFlattenXMLIndex;
+    var _newBeginTextIndex = beginTextIndex;
+    var _newEndTextIndex   = endTextIndex;
+
+    for (var i = beginFlattenXMLIndex; i < xmlTree.length; i++) {
+      var _tag = xmlTree[i];
+      var _matchingTag = _tag.match > -1 ? xmlTree[_tag.match] : _dummyObject;
+      // if this XML part is included in the if-block, and if it is an opening tag
+      if (_matchingTag.lastIndex < endTextIndex && _tag.id < _matchingTag.id) {
+        // jump directly to the end of this currently visited XML (optimization)
+        i = _matchingTag.id;
+      }
+      // if we find a corresponding opening tag next to the first tag of the if-block
+      // move the if-block's beginning earlier
+      if (_matchingTag.lastIndex === _newBeginTextIndex) {
+        _newBeginTextIndex = _matchingTag.index;
+        _newBeginTreeIndex = _matchingTag.id;
+      }
+      // if we find a corresponding ending tag next to the last tag of the if-block
+      // move the if-block's ending further
+      if (_tag.index >= endTextIndex && _matchingTag.index < _tag.index && _matchingTag.index >= beginTextIndex ) {
+        _newEndTextIndex = _tag.lastIndex;
+      }
+      if (_tag.lastIndex > _newEndTextIndex || (_matchingTag.lastIndex < _newBeginTextIndex  && _tag.index > _newEndTextIndex)) {
+        break;
+      }
+    }
+    if (_newBeginTreeIndex !== beginFlattenXMLIndex) {
+      _newBeginTextIndex = xmlTree[_newBeginTreeIndex].index;
+    }
+    return [_newBeginTreeIndex, _newBeginTextIndex, _newEndTextIndex];
+  },
+
+  /**
    * Find safe position of the conditional block in XML
    *
    * Safe means "does not break XML if the XML block between beginning and end is removed"
    *
-   * @param  {String}  xml              Template XML
+   * @param  {String}  xmlFlattened     flattened XML returned by flattenXML
    * @param  {Integer} startSearchIndex beginning index of the conditional block
    * @param  {Integer} endSearchIndex   ending index of the conditional block
    * @return {Array}                    [newBegining, newEnding]
    */
-  findSafeConditionalBlockPosition : function (xml, startSearchIndex, endSearchIndex) {
-    var _tagRegExp =  new RegExp('<(/)?\\b[^/|^>]*(/)?>', 'g'); // reset the regex
-    _tagRegExp.lastIndex = startSearchIndex;
-    var _tag = _tagRegExp.exec(xml);
-    var _openingTagPos = [];
-    var _lastTagEndingPos = startSearchIndex;
+  findSafeConditionalBlockPosition : function (xmlFlattened, beginTextIndex, endTextIndex) {
+    beginTextIndex = Math.trunc(beginTextIndex);
+    endTextIndex   = Math.trunc(endTextIndex);
     var _validCandidates = [];
-    var _currentXMLDepth = 0;
-    var _newCandidate = null;
-    var _lastCandidate = null;
+    var _newCandidate    = null;
+    var _lastCandidate   = [];
+    const _dummyObject   = {};  // constant, empty object to avoid instanciation in the loop below
 
+    // TODO: we can improve performance of we store beginFlattenXMLIndex for each XML part
+    // We could avoid this loop for each call of findSafeConditionalBlockPosition
+    var _beginFlattenXMLIndex = 0;
+    while (_beginFlattenXMLIndex < xmlFlattened.length) {
+      if (xmlFlattened[_beginFlattenXMLIndex].lastIndex > beginTextIndex) {
+        break;
+      }
+      _beginFlattenXMLIndex++;
+    }
+
+    // find better begin/end conditional blocks
+    var _newBeginEndValues = parser.moveConditionalBlockBeginEnd(xmlFlattened, _beginFlattenXMLIndex, beginTextIndex, endTextIndex);
+    _beginFlattenXMLIndex  = _newBeginEndValues[0];
+    beginTextIndex         = Math.trunc(_newBeginEndValues[1]);
+    endTextIndex           = Math.trunc(_newBeginEndValues[2]);
+
+    var _tag = xmlFlattened[_beginFlattenXMLIndex] || {};
+    while (_beginFlattenXMLIndex < xmlFlattened.length) {
+      if (_tag.index >= endTextIndex) {
+        break;
+      }
+      var _matchingTag = _tag.match > -1 ? xmlFlattened[_tag.match] : _dummyObject;
+
+      _beginFlattenXMLIndex++;
+      // if this XML part is included in the if-block, and if it is an opening tag
+      if (_matchingTag.lastIndex <= endTextIndex && _tag.id < _matchingTag.id) {
+        _newCandidate = [_tag.index, _matchingTag.lastIndex];
+        // jump directly to the next valid XML part
+        _beginFlattenXMLIndex = _matchingTag.id + 1;
+      }
+      // no XML
+      else if (_tag.match === undefined) {
+        _newCandidate = [_tag.index, _tag.lastIndex];
+      }
+
+      if (_newCandidate !== null) {
+        // merge adjacent valid XML part
+        if (_newCandidate[0] === _lastCandidate[1]) {
+          _lastCandidate[1] = _newCandidate[1];
+        }
+        else if (_newCandidate[1] === _lastCandidate[0]) {
+          _lastCandidate[0] = _newCandidate[0];
+        }
+        else {
+          // or add new part
+          _validCandidates.push(_newCandidate);
+          _lastCandidate = _newCandidate;
+        }
+        _newCandidate = null;
+      }
+      _tag = xmlFlattened[_beginFlattenXMLIndex];
+    }
     // edge case: when there is no XML tags, return directly
-    if (_tag === null || _tag.index >= endSearchIndex) {
-      return [[startSearchIndex, endSearchIndex, 0]];
+    if (xmlFlattened.length === 0) {
+      return [[beginTextIndex, endTextIndex]];
     }
+    if (_validCandidates.length === 0) {
+      return [[endTextIndex, endTextIndex]];
+    }
+    // never go beyond limits
+    if (_lastCandidate[1] > endTextIndex) {
+      _lastCandidate[1] = endTextIndex;
+    }
+    if (_validCandidates[0][0] < beginTextIndex) {
+      _validCandidates[0][0] = beginTextIndex;
+    }
+    return _validCandidates;
+  },
 
-    while (_tag !== null && _tag.index < endSearchIndex ) {
-      if (_newCandidate === null && _lastTagEndingPos < _tag.index) { // non-XML char
-        _newCandidate = [_lastTagEndingPos, _tag.index, _currentXMLDepth];
-      }
-      // closing tag
-      if (_tag[1] === '/') {
-        _currentXMLDepth--;
-        if (_openingTagPos.length > 0 ) {
-          var _beginningPos = _openingTagPos.pop();
-          _newCandidate = [_beginningPos, _tagRegExp.lastIndex, _currentXMLDepth];
-        }
-      }
-      // or self-closing
-      else if ( _tag[2] === '/') {
-        _newCandidate = [_lastTagEndingPos, _tagRegExp.lastIndex, _currentXMLDepth];
-      }
-      // opening tag
-      else {
-        _currentXMLDepth++;
-        _openingTagPos.push(_lastTagEndingPos);
-      }
-      // merge wirth previous detected XML part
-      if ( _newCandidate  !== null
-        && _lastCandidate !== null
-        // when the new XML part is nested in the previous one
-        && (_lastCandidate[0] > _newCandidate[0]
-          // when the new XML part is adjacent or overlaps with the previous part, and is at the same depth
-          ||   _newCandidate[2] === _lastCandidate[2]
-            && _newCandidate[0] <=  _lastCandidate[1]
-        )
-      ) {
-        _lastCandidate[0] = Math.min(_newCandidate[0], _lastCandidate[0]);
-        _lastCandidate[1] = _newCandidate[1];
-        _lastCandidate[2] = _newCandidate[2];
-        _newCandidate = null;
-      }
-      _lastTagEndingPos = _tagRegExp.lastIndex;
-      _tag = _tagRegExp.exec(xml);
 
-      if (_newCandidate) {
-        _validCandidates.push(_newCandidate);
-        _lastCandidate = _newCandidate;
-        _newCandidate = null;
-      }
+  /**
+   * Simple mathematical expression parser without parenthesis
+   *
+   * @param   {String}    mathExpr                 The mathematics expression coming from a formatter calc, add, mul, div, sub
+   * @param   {Function}  safeVariableInjectionFn  The safe variable injection function
+   * @return  {String}    code to inject in builder
+   */
+  parseMathematicalExpression : function (mathExpr, safeVariableInjectionFn) {
+    if (typeof mathExpr !== 'string' || mathExpr.trim() === '') {
+      return '';
     }
-    if (_lastCandidate === null && _lastTagEndingPos <= _tag.index) { // non-XML char
-      _lastCandidate = [_lastTagEndingPos, _tag.index, _currentXMLDepth];
-      _validCandidates.push(_lastCandidate);
-    }
-    if (_lastCandidate[1] > endSearchIndex) {
-      _lastCandidate[1] = endSearchIndex;
-    }
-    // include trailing non XML char, same depth
-    if (_tag !== null && _lastTagEndingPos < _tag.index) {
-      var _newEndingTag = Math.min(_tag.index, endSearchIndex);
-      if (_lastCandidate[2] === _currentXMLDepth) {
-        _lastCandidate[1] =  _newEndingTag;
+    let _currMathExprIndex = mathExpr.length - 1;
+    let _injectedCode = '';
+    let _prevVariable = '';
+    let _prevOperator = '';
+    const _variables = mathExpr.split(/[*+/\-]/);
+    for(let i = _variables.length - 1; i > -1; i--) {
+      const _variable = _variables[i];
+      _currMathExprIndex -= (_variable.length + 1); // add operator length, which is always "1"
+      const _operator = mathExpr[_currMathExprIndex + 1] ?? '';
+      const _trimVariable = _variable.trim();
+      if (_trimVariable.length === 0 && _prevOperator !== '-') {
+        throw Error ('Bad Mathematical Expression in "'+mathExpr+'"');
       }
-      else {
-        if (_newEndingTag > _lastTagEndingPos) {
-          _validCandidates.push([_lastTagEndingPos, _newEndingTag, _currentXMLDepth]);
-        }
+      // accept variable with dashes for example
+      if (/^[0-9\.]/.test(_trimVariable) === false) {
+        _prevVariable = _operator + _trimVariable + _prevVariable;
+        continue;
       }
+      const _fullVariable = _trimVariable + _prevVariable;
+      const _safeVarCode = safeVariableInjectionFn(_fullVariable);
+      _prevVariable = '';
+      _injectedCode = _operator + 'parseFloat(' + _safeVarCode + ')' + _injectedCode;
+      _prevOperator = _operator;
     }
-
-    _lastCandidate = null;
-    var _newCandidates = [];
-    for (var i = 0; i < _validCandidates.length; i++) {
-      var _validCandidate = _validCandidates[i];
-      // merge consecutive valid XML-part
-      if (_lastCandidate !== null && _validCandidate[2] === _lastCandidate[2] && _validCandidate[0] <= _lastCandidate[1]) {
-        _lastCandidate[1] = _validCandidate[1];
-      }
-      else {
-        _newCandidates.push(_validCandidate);
-        _lastCandidate = _validCandidate;
-      }
-    }
-    return _newCandidates;
+    return _injectedCode;
   }
 
 };
