@@ -1,5 +1,6 @@
-var path = require('path');
+var path  = require('path');
 var helper = require('./helper');
+var parser = require('./parser');
 
 
 var preprocessor = {
@@ -9,9 +10,16 @@ var preprocessor = {
    * @param  {Object}   template
    * @param  {Function} callback
    */
-  execute: function (template, callback) {
+  execute : function (template, options, callback) {
+    if (!(callback instanceof Function)) {
+      callback = options;
+      options = {};
+    }
     if (template === null || template.files === undefined) {
       return callback(null, template);
+    }
+    for (let i = 0, len = template.files.length; i < len; i++) {
+      template.files[i].data = preprocessor.preParseXML(template.files[i].data, options);
     }
     for (var i = -1; i < template.embeddings.length; i++) {
       var _mainOrEmbeddedTemplate = template.filename;
@@ -24,10 +32,14 @@ var preprocessor = {
       }
       switch (_fileType) {
         case 'xlsx':
+          preprocessor.removeCalcChain(template, _parentFilter);
           preprocessor.convertSharedStringToInlineString(template, _parentFilter);
           break;
         case 'ods':
           preprocessor.convertNumberMarkersIntoNumericFormat(template);
+          break;
+        case 'odt':
+          preprocessor.removeSoftPageBreak(template);
           break;
         default:
           break;
@@ -37,12 +49,49 @@ var preprocessor = {
   },
 
   /**
+   * Loop through XML files to clean tags inside markers and to translate
+   *
+   * @param {Object} template
+   */
+  preParseXML : function (xml, options) {
+    // find translation markers {t()} and translate before other processing
+    xml = parser.translate(xml, options);
+    // Remove XML tags inside markers
+    xml = parser.removeXMLInsideMarkers(xml);
+    return xml;
+  },
+  /**
+   * Remove soft page break in ODT document as we modify them
+   *
+   * Search title "text:use-soft-page-breaks" in
+   * http://docs.oasis-open.org/office/v1.2/os/OpenDocument-v1.2-os-part1.html#__RefHeading__1415190_253892949
+   *
+   * TODO: Should we do it in Word?
+   *
+   * @param  {Object} template (modified)
+   * @return {Object}          template
+   */
+  removeSoftPageBreak : function (template) {
+    for (var i = 0; i < template.files.length; i++) {
+      var _file = template.files[i];
+      if (/content\.xml$/.test(_file.name) === true) {
+        _file.data = _file.data
+          .replace(/text:use-soft-page-breaks="true"/g, '')
+          .replace(/<text:soft-page-break\/>/g, '')
+          .replace(/<text:soft-page-break><\/text:soft-page-break>/g, '');
+        return template;
+      }
+    }
+    return template;
+  },
+
+  /**
    * [XLSX] Convert shared string to inline string in Excel format in order to be compatible with Carbone algorithm
    * @param  {Object} template     (modified)
    * @param  {String} parentFilter apply the transformation on a specific embedded file
    * @return {Object}              the modified template
    */
-  convertSharedStringToInlineString: function (template, parentFilter) {
+  convertSharedStringToInlineString : function (template, parentFilter) {
     var _sharedStrings = [];
     var _filesToConvert = [];
     var _sharedStringIndex = -1;
@@ -65,7 +114,36 @@ var preprocessor = {
     // once shared string is found, convert files
     for (var f = 0; f < _filesToConvert.length; f++) {
       var _modifiedFile = _filesToConvert[f];
-      _modifiedFile.data = preprocessor.convertToInlineString(_modifiedFile.data, _sharedStrings);
+      _modifiedFile.data = preprocessor.convertToInlineString(_modifiedFile.data, _sharedStrings)
+    }
+    return template;
+  },
+
+  /**
+   * Removes uselss calc chain file.
+   *
+   * https://docs.microsoft.com/en-us/office/open-xml/working-with-the-calculation-chain
+   *
+   * This file is a sort of "cache" for Excel. The Calculation Chain part specifies the order in
+   * which cells in the workbook were last calculated.
+   *
+   * As Carbone will modify the Excel file, this file may become incorrect.
+   * We can safely remove this file to force Excel to recompute this file on opening.
+   *
+   * It removes the "alert" when opening the generated report with Excel
+   *
+   * @param      {<type>}  template      The template
+   * @param      {<type>}  parentFilter  The parent filter
+   * @return     {<type>}  { description_of_the_return_value }
+   */
+  removeCalcChain : function (template, parentFilter) {
+    // parse all files and find shared strings first
+    for (var i = 0; i < template.files.length; i++) {
+      var _file = template.files[i];
+      if (_file.parent === parentFilter && /xl\/calcChain\.xml$/.test(_file.name) === true) {
+        preprocessor.removeOneFile(template, i, parentFilter);
+        break;
+      }
     }
     return template;
   },
@@ -77,12 +155,12 @@ var preprocessor = {
    * @param  {String} parentFilter         filter to modify only an embedded document
    * @return {}                            it modifies the template directly
    */
-  removeOneFile: function (template, indexOfFileToRemove, parentFilter) {
+  removeOneFile : function (template, indexOfFileToRemove, parentFilter) {
     if (indexOfFileToRemove < 0 || indexOfFileToRemove >= template.files.length) {
       return;
     }
     var _fileToRemove = template.files[indexOfFileToRemove];
-    var _dirname = path.dirname(_fileToRemove.name);
+    var _dirname  = path.dirname(_fileToRemove.name);
     var _basename = path.basename(_fileToRemove.name);
     template.files.splice(indexOfFileToRemove, 1);
     for (var i = 0; i < template.files.length; i++) {
@@ -102,12 +180,12 @@ var preprocessor = {
    * @param  {String} sharedStringXml shared string content
    * @return {Array}                  array
    */
-  readSharedString: function (sharedStringXml) {
+  readSharedString : function (sharedStringXml) {
     var _sharedStrings = [];
     if (sharedStringXml === null || sharedStringXml === undefined) {
       return _sharedStrings;
     }
-    var _tagRegex = new RegExp('<si>(.+?)</si>', 'g');
+    var _tagRegex = new RegExp('<si>(.+?)</si>','g');
     var _tag = _tagRegex.exec(sharedStringXml);
     while (_tag !== null) {
       _sharedStrings.push(_tag[1]);
@@ -122,8 +200,8 @@ var preprocessor = {
    * @param  {Array} sharedStrings  shared string
    * @return {String}               updated xml
    */
-  convertToInlineString: function (xml, sharedStrings) {
-    if (typeof (xml) !== 'string') {
+  convertToInlineString : function (xml, sharedStrings) {
+    if (typeof(xml) !== 'string') {
       return xml;
     }
     // find all tags which have attribute t="s" (type = shared string)
@@ -135,7 +213,7 @@ var preprocessor = {
         // replace the index by the string
         var _sharedStringIndex = parseInt(_tab[1], 10);
         var _selectedSharedString = sharedStrings[_sharedStringIndex];
-        const _contentAsNumber = /{.*:formatN\(.*\)}/.exec(_selectedSharedString);
+        const _contentAsNumber = /{[d|c][.].*:formatN\(.*\)}/.exec(_selectedSharedString);
         if (_contentAsNumber instanceof Array && _tab.length > 0) {
           // Convert the marker into number cell type when it using the ':formatN' formatter
           _selectedSharedString = _contentAsNumber[0].replace(/:formatN\(.*\)/, '');
@@ -162,8 +240,8 @@ var preprocessor = {
    * @param  {String} xml sheet
    * @return {String}     sheet updated
    */
-  removeRowCounterInWorksheet: function (xml) {
-    if (typeof (xml) !== 'string') {
+  removeRowCounterInWorksheet : function (xml) {
+    if (typeof(xml) !== 'string') {
       return xml;
     }
     return xml.replace(/<(?:c|row)[^>]*\s(r="\S+")[^>]*>/g, function (m, rowValue) {
@@ -179,7 +257,7 @@ var preprocessor = {
    * @param {Object} template
    * @return
    */
-  convertNumberMarkersIntoNumericFormat: function (template) {
+  convertNumberMarkersIntoNumericFormat : function (template) {
     const _contentFileId = template.files.findIndex(x => x.name === 'content.xml');
     if (_contentFileId > -1 && !!template.files[_contentFileId] === true) {
       template.files[_contentFileId].data = template.files[_contentFileId].data.replace(/<table:table-cell[^<]*>\s*<text:p>[^<]*formatN[^<]*<\/text:p>\s*<\/table:table-cell>/g, function (xml) {
