@@ -51,7 +51,7 @@ function walkReportFiles (template, currentIndex, strToFind, callback) {
         var _marker = _allMarkers[i];
         if (_marker.name !== undefined && _marker.name.indexOf(strToFind) !== -1) {
           nbOccurencies++;
-          console.log('\033[37mfound in \033[0m' + path.basename(template.filename) + '\033[37m -> '+ _marker.name+'\033[0m');
+          console.log('\x1b[37mfound in \x1b[0m' + path.basename(template.filename) + '\x1b[37m -> '+ _marker.name+'\x1b[0m');
         }
       }
       return walkReportFiles(template, ++currentIndex, strToFind, callback);
