@@ -1,4 +1,5 @@
 export function init(options: any, callback: Function): void;
 export function exit(callback: Function): void;
-export function convertFile(inputFile: string, outputType: string, formatOptions: string, callback: Function, returnLink: any, outputFile: any): any;
+export function convertFile(inputFile: string, outputType: string, formatOptions: string, outputFile: string, callback: Function): any;
+export function shouldTheFactoryBeRestarted(params: Objecct, availableMemory: Integer, nbReports: Integer): boolean;
 //# sourceMappingURL=converter.d.ts.map
