@@ -1,7 +1,7 @@
-export function splitMarkers(markers: any[]): any;
+export function splitMarkers(markers: any[]): object;
 export function _cleanUniqueMarkerName(markerName: string): string;
 export function sortXmlParts(xmlParts: any[]): void;
-export function splitXml(xml: string, descriptor: any): any;
-export function buildSortedHierarchy(descriptor: any): any;
-export function deleteAndMoveNestedParts(descriptor: any): any;
+export function splitXml(xml: string, descriptor: object): object;
+export function buildSortedHierarchy(descriptor: object): object;
+export function deleteAndMoveNestedParts(descriptor: object): object;
 //# sourceMappingURL=extracter.d.ts.map
