@@ -1,7 +1,8 @@
 export function isZipped(filePath: string, callback: Function): void;
+export function getTemplateFile(template: any, filename: string): any | null;
 export function unzip(filePath: string, callback: Function): void;
 export function zip(files: any[], callback: Function): void;
-export function openTemplate(templateId: string, callback: Function): void;
+export function openTemplate(templateId: string, callback: Function): any;
 export function detectType(template: any): any;
 export function _isXMLFile(template: any): boolean;
 export function _isXHTMLFile(template: any): boolean;
