@@ -1,1216 +1,1250 @@
 export namespace document {
     namespace bib {
-        const extension: string;
-        const desc: string;
-        const format: string;
+        let extension: string;
+        let desc: string;
+        let format: string;
     }
     namespace doc {
-        const extension_1: string;
+        let extension_1: string;
         export { extension_1 as extension };
-        const desc_1: string;
+        let desc_1: string;
         export { desc_1 as desc };
-        const format_1: string;
+        let format_1: string;
         export { format_1 as format };
     }
     namespace doc6 {
-        const extension_2: string;
+        let extension_2: string;
         export { extension_2 as extension };
-        const desc_2: string;
+        let desc_2: string;
         export { desc_2 as desc };
-        const format_2: string;
+        let format_2: string;
         export { format_2 as format };
     }
     namespace doc95 {
-        const extension_3: string;
+        let extension_3: string;
         export { extension_3 as extension };
-        const desc_3: string;
+        let desc_3: string;
         export { desc_3 as desc };
-        const format_3: string;
+        let format_3: string;
         export { format_3 as format };
     }
     namespace docbook {
-        const extension_4: string;
+        let extension_4: string;
         export { extension_4 as extension };
-        const desc_4: string;
+        let desc_4: string;
         export { desc_4 as desc };
-        const format_4: string;
+        let format_4: string;
         export { format_4 as format };
     }
     namespace docx {
-        const extension_5: string;
+        let extension_5: string;
         export { extension_5 as extension };
-        const desc_5: string;
+        let desc_5: string;
         export { desc_5 as desc };
-        const format_5: string;
+        let format_5: string;
         export { format_5 as format };
     }
     namespace docx7 {
-        const extension_6: string;
+        let extension_6: string;
         export { extension_6 as extension };
-        const desc_6: string;
+        let desc_6: string;
         export { desc_6 as desc };
-        const format_6: string;
+        let format_6: string;
         export { format_6 as format };
     }
     namespace fodt {
-        const extension_7: string;
+        let extension_7: string;
         export { extension_7 as extension };
-        const desc_7: string;
+        let desc_7: string;
         export { desc_7 as desc };
-        const format_7: string;
+        let format_7: string;
         export { format_7 as format };
     }
     namespace html {
-        const extension_8: string;
+        let extension_8: string;
         export { extension_8 as extension };
-        const desc_8: string;
+        let desc_8: string;
         export { desc_8 as desc };
-        const format_8: string;
+        let format_8: string;
         export { format_8 as format };
     }
     namespace latex {
-        const extension_9: string;
+        let extension_9: string;
         export { extension_9 as extension };
-        const desc_9: string;
+        let desc_9: string;
         export { desc_9 as desc };
-        const format_9: string;
+        let format_9: string;
         export { format_9 as format };
     }
     namespace mediawiki {
-        const extension_10: string;
+        let extension_10: string;
         export { extension_10 as extension };
-        const desc_10: string;
+        let desc_10: string;
         export { desc_10 as desc };
-        const format_10: string;
+        let format_10: string;
         export { format_10 as format };
     }
     namespace odt {
-        const extension_11: string;
+        let extension_11: string;
         export { extension_11 as extension };
-        const desc_11: string;
+        let desc_11: string;
         export { desc_11 as desc };
-        const format_11: string;
+        let format_11: string;
         export { format_11 as format };
     }
     namespace ooxml {
-        const extension_12: string;
+        let extension_12: string;
         export { extension_12 as extension };
-        const desc_12: string;
+        let desc_12: string;
         export { desc_12 as desc };
-        const format_12: string;
+        let format_12: string;
         export { format_12 as format };
     }
     namespace ott {
-        const extension_13: string;
+        let extension_13: string;
         export { extension_13 as extension };
-        const desc_13: string;
+        let desc_13: string;
         export { desc_13 as desc };
-        const format_13: string;
+        let format_13: string;
         export { format_13 as format };
     }
     namespace pdb {
-        const extension_14: string;
+        let extension_14: string;
         export { extension_14 as extension };
-        const desc_14: string;
+        let desc_14: string;
         export { desc_14 as desc };
-        const format_14: string;
+        let format_14: string;
         export { format_14 as format };
     }
     namespace pdf {
-        const extension_15: string;
+        let extension_15: string;
         export { extension_15 as extension };
-        const desc_15: string;
+        let desc_15: string;
         export { desc_15 as desc };
-        const format_15: string;
+        let format_15: string;
         export { format_15 as format };
     }
     namespace psw {
-        const extension_16: string;
+        let extension_16: string;
         export { extension_16 as extension };
-        const desc_16: string;
+        let desc_16: string;
         export { desc_16 as desc };
-        const format_16: string;
+        let format_16: string;
         export { format_16 as format };
     }
     namespace rtf {
-        const extension_17: string;
+        let extension_17: string;
         export { extension_17 as extension };
-        const desc_17: string;
+        let desc_17: string;
         export { desc_17 as desc };
-        const format_17: string;
+        let format_17: string;
         export { format_17 as format };
     }
     namespace sdw {
-        const extension_18: string;
+        let extension_18: string;
         export { extension_18 as extension };
-        const desc_18: string;
+        let desc_18: string;
         export { desc_18 as desc };
-        const format_18: string;
+        let format_18: string;
         export { format_18 as format };
     }
     namespace sdw4 {
-        const extension_19: string;
+        let extension_19: string;
         export { extension_19 as extension };
-        const desc_19: string;
+        let desc_19: string;
         export { desc_19 as desc };
-        const format_19: string;
+        let format_19: string;
         export { format_19 as format };
     }
     namespace sdw3 {
-        const extension_20: string;
+        let extension_20: string;
         export { extension_20 as extension };
-        const desc_20: string;
+        let desc_20: string;
         export { desc_20 as desc };
-        const format_20: string;
+        let format_20: string;
         export { format_20 as format };
     }
     namespace stw {
-        const extension_21: string;
+        let extension_21: string;
         export { extension_21 as extension };
-        const desc_21: string;
+        let desc_21: string;
         export { desc_21 as desc };
-        const format_21: string;
+        let format_21: string;
         export { format_21 as format };
     }
     namespace sxw {
-        const extension_22: string;
+        let extension_22: string;
         export { extension_22 as extension };
-        const desc_22: string;
+        let desc_22: string;
         export { desc_22 as desc };
-        const format_22: string;
+        let format_22: string;
         export { format_22 as format };
     }
     namespace text {
-        const extension_23: string;
+        let extension_23: string;
         export { extension_23 as extension };
-        const desc_23: string;
+        let desc_23: string;
         export { desc_23 as desc };
-        const format_23: string;
+        let format_23: string;
         export { format_23 as format };
     }
     namespace txt {
-        const extension_24: string;
+        let extension_24: string;
         export { extension_24 as extension };
-        const desc_24: string;
+        let desc_24: string;
         export { desc_24 as desc };
-        const format_24: string;
+        let format_24: string;
         export { format_24 as format };
     }
     namespace uot {
-        const extension_25: string;
+        let extension_25: string;
         export { extension_25 as extension };
-        const desc_25: string;
+        let desc_25: string;
         export { desc_25 as desc };
-        const format_25: string;
+        let format_25: string;
         export { format_25 as format };
     }
     namespace vor {
-        const extension_26: string;
+        let extension_26: string;
         export { extension_26 as extension };
-        const desc_26: string;
+        let desc_26: string;
         export { desc_26 as desc };
-        const format_26: string;
+        let format_26: string;
         export { format_26 as format };
     }
     namespace vor4 {
-        const extension_27: string;
+        let extension_27: string;
         export { extension_27 as extension };
-        const desc_27: string;
+        let desc_27: string;
         export { desc_27 as desc };
-        const format_27: string;
+        let format_27: string;
         export { format_27 as format };
     }
     namespace vor3 {
-        const extension_28: string;
+        let extension_28: string;
         export { extension_28 as extension };
-        const desc_28: string;
+        let desc_28: string;
         export { desc_28 as desc };
-        const format_28: string;
+        let format_28: string;
         export { format_28 as format };
     }
     namespace xhtml {
-        const extension_29: string;
+        let extension_29: string;
         export { extension_29 as extension };
-        const desc_29: string;
+        let desc_29: string;
         export { desc_29 as desc };
-        const format_29: string;
+        let format_29: string;
         export { format_29 as format };
+    }
+    namespace jpg {
+        let extension_30: string;
+        export { extension_30 as extension };
+        let desc_30: string;
+        export { desc_30 as desc };
+        let format_30: string;
+        export { format_30 as format };
+    }
+    namespace jpeg {
+        let extension_31: string;
+        export { extension_31 as extension };
+        let desc_31: string;
+        export { desc_31 as desc };
+        let format_31: string;
+        export { format_31 as format };
+    }
+    namespace png {
+        let extension_32: string;
+        export { extension_32 as extension };
+        let desc_32: string;
+        export { desc_32 as desc };
+        let format_32: string;
+        export { format_32 as format };
+    }
+    namespace epub {
+        let extension_33: string;
+        export { extension_33 as extension };
+        let desc_33: string;
+        export { desc_33 as desc };
+        let format_33: string;
+        export { format_33 as format };
     }
 }
 export namespace web {
     export namespace etext {
-        const extension_30: string;
-        export { extension_30 as extension };
-        const desc_30: string;
-        export { desc_30 as desc };
-        const format_30: string;
-        export { format_30 as format };
+        let extension_34: string;
+        export { extension_34 as extension };
+        let desc_34: string;
+        export { desc_34 as desc };
+        let format_34: string;
+        export { format_34 as format };
     }
     export namespace html10 {
-        const extension_31: string;
-        export { extension_31 as extension };
-        const desc_31: string;
-        export { desc_31 as desc };
-        const format_31: string;
-        export { format_31 as format };
+        let extension_35: string;
+        export { extension_35 as extension };
+        let desc_35: string;
+        export { desc_35 as desc };
+        let format_35: string;
+        export { format_35 as format };
     }
     export namespace html_1 {
-        const extension_32: string;
-        export { extension_32 as extension };
-        const desc_32: string;
-        export { desc_32 as desc };
-        const format_32: string;
-        export { format_32 as format };
+        let extension_36: string;
+        export { extension_36 as extension };
+        let desc_36: string;
+        export { desc_36 as desc };
+        let format_36: string;
+        export { format_36 as format };
     }
     export { html_1 as html };
     export namespace mediawiki_1 {
-        const extension_33: string;
-        export { extension_33 as extension };
-        const desc_33: string;
-        export { desc_33 as desc };
-        const format_33: string;
-        export { format_33 as format };
+        let extension_37: string;
+        export { extension_37 as extension };
+        let desc_37: string;
+        export { desc_37 as desc };
+        let format_37: string;
+        export { format_37 as format };
     }
     export { mediawiki_1 as mediawiki };
     export namespace pdf_1 {
-        const extension_34: string;
-        export { extension_34 as extension };
-        const desc_34: string;
-        export { desc_34 as desc };
-        const format_34: string;
-        export { format_34 as format };
+        let extension_38: string;
+        export { extension_38 as extension };
+        let desc_38: string;
+        export { desc_38 as desc };
+        let format_38: string;
+        export { format_38 as format };
     }
     export { pdf_1 as pdf };
     export namespace sdw3_1 {
-        const extension_35: string;
-        export { extension_35 as extension };
-        const desc_35: string;
-        export { desc_35 as desc };
-        const format_35: string;
-        export { format_35 as format };
+        let extension_39: string;
+        export { extension_39 as extension };
+        let desc_39: string;
+        export { desc_39 as desc };
+        let format_39: string;
+        export { format_39 as format };
     }
     export { sdw3_1 as sdw3 };
     export namespace sdw4_1 {
-        const extension_36: string;
-        export { extension_36 as extension };
-        const desc_36: string;
-        export { desc_36 as desc };
-        const format_36: string;
-        export { format_36 as format };
+        let extension_40: string;
+        export { extension_40 as extension };
+        let desc_40: string;
+        export { desc_40 as desc };
+        let format_40: string;
+        export { format_40 as format };
     }
     export { sdw4_1 as sdw4 };
     export namespace sdw_1 {
-        const extension_37: string;
-        export { extension_37 as extension };
-        const desc_37: string;
-        export { desc_37 as desc };
-        const format_37: string;
-        export { format_37 as format };
+        let extension_41: string;
+        export { extension_41 as extension };
+        let desc_41: string;
+        export { desc_41 as desc };
+        let format_41: string;
+        export { format_41 as format };
     }
     export { sdw_1 as sdw };
     export namespace txt_1 {
-        const extension_38: string;
-        export { extension_38 as extension };
-        const desc_38: string;
-        export { desc_38 as desc };
-        const format_38: string;
-        export { format_38 as format };
+        let extension_42: string;
+        export { extension_42 as extension };
+        let desc_42: string;
+        export { desc_42 as desc };
+        let format_42: string;
+        export { format_42 as format };
     }
     export { txt_1 as txt };
     export namespace text10 {
-        const extension_39: string;
-        export { extension_39 as extension };
-        const desc_39: string;
-        export { desc_39 as desc };
-        const format_39: string;
-        export { format_39 as format };
+        let extension_43: string;
+        export { extension_43 as extension };
+        let desc_43: string;
+        export { desc_43 as desc };
+        let format_43: string;
+        export { format_43 as format };
     }
     export namespace text_1 {
-        const extension_40: string;
-        export { extension_40 as extension };
-        const desc_40: string;
-        export { desc_40 as desc };
-        const format_40: string;
-        export { format_40 as format };
+        let extension_44: string;
+        export { extension_44 as extension };
+        let desc_44: string;
+        export { desc_44 as desc };
+        let format_44: string;
+        export { format_44 as format };
     }
     export { text_1 as text };
     export namespace vor4_1 {
-        const extension_41: string;
-        export { extension_41 as extension };
-        const desc_41: string;
-        export { desc_41 as desc };
-        const format_41: string;
-        export { format_41 as format };
+        let extension_45: string;
+        export { extension_45 as extension };
+        let desc_45: string;
+        export { desc_45 as desc };
+        let format_45: string;
+        export { format_45 as format };
     }
     export { vor4_1 as vor4 };
     export namespace vor_1 {
-        const extension_42: string;
-        export { extension_42 as extension };
-        const desc_42: string;
-        export { desc_42 as desc };
-        const format_42: string;
-        export { format_42 as format };
+        let extension_46: string;
+        export { extension_46 as extension };
+        let desc_46: string;
+        export { desc_46 as desc };
+        let format_46: string;
+        export { format_46 as format };
     }
     export { vor_1 as vor };
 }
 export namespace spreadsheet {
     export namespace dbf {
-        const extension_43: string;
-        export { extension_43 as extension };
-        const desc_43: string;
-        export { desc_43 as desc };
-        const format_43: string;
-        export { format_43 as format };
+        let extension_47: string;
+        export { extension_47 as extension };
+        let desc_47: string;
+        export { desc_47 as desc };
+        let format_47: string;
+        export { format_47 as format };
     }
     export namespace dif {
-        const extension_44: string;
-        export { extension_44 as extension };
-        const desc_44: string;
-        export { desc_44 as desc };
-        const format_44: string;
-        export { format_44 as format };
+        let extension_48: string;
+        export { extension_48 as extension };
+        let desc_48: string;
+        export { desc_48 as desc };
+        let format_48: string;
+        export { format_48 as format };
     }
     export namespace fods {
-        const extension_45: string;
-        export { extension_45 as extension };
-        const desc_45: string;
-        export { desc_45 as desc };
-        const format_45: string;
-        export { format_45 as format };
+        let extension_49: string;
+        export { extension_49 as extension };
+        let desc_49: string;
+        export { desc_49 as desc };
+        let format_49: string;
+        export { format_49 as format };
     }
     export namespace html_2 {
-        const extension_46: string;
-        export { extension_46 as extension };
-        const desc_46: string;
-        export { desc_46 as desc };
-        const format_46: string;
-        export { format_46 as format };
+        let extension_50: string;
+        export { extension_50 as extension };
+        let desc_50: string;
+        export { desc_50 as desc };
+        let format_50: string;
+        export { format_50 as format };
     }
     export { html_2 as html };
     export namespace ods {
-        const extension_47: string;
-        export { extension_47 as extension };
-        const desc_47: string;
-        export { desc_47 as desc };
-        const format_47: string;
-        export { format_47 as format };
+        let extension_51: string;
+        export { extension_51 as extension };
+        let desc_51: string;
+        export { desc_51 as desc };
+        let format_51: string;
+        export { format_51 as format };
     }
     export namespace ooxml_1 {
-        const extension_48: string;
-        export { extension_48 as extension };
-        const desc_48: string;
-        export { desc_48 as desc };
-        const format_48: string;
-        export { format_48 as format };
+        let extension_52: string;
+        export { extension_52 as extension };
+        let desc_52: string;
+        export { desc_52 as desc };
+        let format_52: string;
+        export { format_52 as format };
     }
     export { ooxml_1 as ooxml };
     export namespace ots {
-        const extension_49: string;
-        export { extension_49 as extension };
-        const desc_49: string;
-        export { desc_49 as desc };
-        const format_49: string;
-        export { format_49 as format };
+        let extension_53: string;
+        export { extension_53 as extension };
+        let desc_53: string;
+        export { desc_53 as desc };
+        let format_53: string;
+        export { format_53 as format };
     }
     export namespace pdf_2 {
-        const extension_50: string;
-        export { extension_50 as extension };
-        const desc_50: string;
-        export { desc_50 as desc };
-        const format_50: string;
-        export { format_50 as format };
+        let extension_54: string;
+        export { extension_54 as extension };
+        let desc_54: string;
+        export { desc_54 as desc };
+        let format_54: string;
+        export { format_54 as format };
     }
     export { pdf_2 as pdf };
     export namespace pxl {
-        const extension_51: string;
-        export { extension_51 as extension };
-        const desc_51: string;
-        export { desc_51 as desc };
-        const format_51: string;
-        export { format_51 as format };
-    }
-    export namespace sdc {
-        const extension_52: string;
-        export { extension_52 as extension };
-        const desc_52: string;
-        export { desc_52 as desc };
-        const format_52: string;
-        export { format_52 as format };
-    }
-    export namespace sdc4 {
-        const extension_53: string;
-        export { extension_53 as extension };
-        const desc_53: string;
-        export { desc_53 as desc };
-        const format_53: string;
-        export { format_53 as format };
-    }
-    export namespace sdc3 {
-        const extension_54: string;
-        export { extension_54 as extension };
-        const desc_54: string;
-        export { desc_54 as desc };
-        const format_54: string;
-        export { format_54 as format };
-    }
-    export namespace slk {
-        const extension_55: string;
+        let extension_55: string;
         export { extension_55 as extension };
-        const desc_55: string;
+        let desc_55: string;
         export { desc_55 as desc };
-        const format_55: string;
+        let format_55: string;
         export { format_55 as format };
     }
-    export namespace stc {
-        const extension_56: string;
+    export namespace sdc {
+        let extension_56: string;
         export { extension_56 as extension };
-        const desc_56: string;
+        let desc_56: string;
         export { desc_56 as desc };
-        const format_56: string;
+        let format_56: string;
         export { format_56 as format };
     }
-    export namespace sxc {
-        const extension_57: string;
+    export namespace sdc4 {
+        let extension_57: string;
         export { extension_57 as extension };
-        const desc_57: string;
+        let desc_57: string;
         export { desc_57 as desc };
-        const format_57: string;
+        let format_57: string;
         export { format_57 as format };
     }
-    export namespace uos {
-        const extension_58: string;
+    export namespace sdc3 {
+        let extension_58: string;
         export { extension_58 as extension };
-        const desc_58: string;
+        let desc_58: string;
         export { desc_58 as desc };
-        const format_58: string;
+        let format_58: string;
         export { format_58 as format };
     }
-    export namespace vor3_1 {
-        const extension_59: string;
+    export namespace slk {
+        let extension_59: string;
         export { extension_59 as extension };
-        const desc_59: string;
+        let desc_59: string;
         export { desc_59 as desc };
-        const format_59: string;
+        let format_59: string;
         export { format_59 as format };
+    }
+    export namespace stc {
+        let extension_60: string;
+        export { extension_60 as extension };
+        let desc_60: string;
+        export { desc_60 as desc };
+        let format_60: string;
+        export { format_60 as format };
+    }
+    export namespace sxc {
+        let extension_61: string;
+        export { extension_61 as extension };
+        let desc_61: string;
+        export { desc_61 as desc };
+        let format_61: string;
+        export { format_61 as format };
+    }
+    export namespace uos {
+        let extension_62: string;
+        export { extension_62 as extension };
+        let desc_62: string;
+        export { desc_62 as desc };
+        let format_62: string;
+        export { format_62 as format };
+    }
+    export namespace vor3_1 {
+        let extension_63: string;
+        export { extension_63 as extension };
+        let desc_63: string;
+        export { desc_63 as desc };
+        let format_63: string;
+        export { format_63 as format };
     }
     export { vor3_1 as vor3 };
     export namespace vor4_2 {
-        const extension_60: string;
-        export { extension_60 as extension };
-        const desc_60: string;
-        export { desc_60 as desc };
-        const format_60: string;
-        export { format_60 as format };
+        let extension_64: string;
+        export { extension_64 as extension };
+        let desc_64: string;
+        export { desc_64 as desc };
+        let format_64: string;
+        export { format_64 as format };
     }
     export { vor4_2 as vor4 };
     export namespace vor_2 {
-        const extension_61: string;
-        export { extension_61 as extension };
-        const desc_61: string;
-        export { desc_61 as desc };
-        const format_61: string;
-        export { format_61 as format };
+        let extension_65: string;
+        export { extension_65 as extension };
+        let desc_65: string;
+        export { desc_65 as desc };
+        let format_65: string;
+        export { format_65 as format };
     }
     export { vor_2 as vor };
     export namespace xhtml_1 {
-        const extension_62: string;
-        export { extension_62 as extension };
-        const desc_62: string;
-        export { desc_62 as desc };
-        const format_62: string;
-        export { format_62 as format };
+        let extension_66: string;
+        export { extension_66 as extension };
+        let desc_66: string;
+        export { desc_66 as desc };
+        let format_66: string;
+        export { format_66 as format };
     }
     export { xhtml_1 as xhtml };
     export namespace csv {
-        const extension_63: string;
-        export { extension_63 as extension };
-        const desc_63: string;
-        export { desc_63 as desc };
-        const format_63: string;
-        export { format_63 as format };
-    }
-    export namespace xls {
-        const extension_64: string;
-        export { extension_64 as extension };
-        const desc_64: string;
-        export { desc_64 as desc };
-        const format_64: string;
-        export { format_64 as format };
-    }
-    export namespace xls5 {
-        const extension_65: string;
-        export { extension_65 as extension };
-        const desc_65: string;
-        export { desc_65 as desc };
-        const format_65: string;
-        export { format_65 as format };
-    }
-    export namespace xls95 {
-        const extension_66: string;
-        export { extension_66 as extension };
-        const desc_66: string;
-        export { desc_66 as desc };
-        const format_66: string;
-        export { format_66 as format };
-    }
-    export namespace xlt {
-        const extension_67: string;
+        let extension_67: string;
         export { extension_67 as extension };
-        const desc_67: string;
+        let desc_67: string;
         export { desc_67 as desc };
-        const format_67: string;
+        let format_67: string;
         export { format_67 as format };
     }
-    export namespace xlt5 {
-        const extension_68: string;
+    export namespace xls {
+        let extension_68: string;
         export { extension_68 as extension };
-        const desc_68: string;
+        let desc_68: string;
         export { desc_68 as desc };
-        const format_68: string;
+        let format_68: string;
         export { format_68 as format };
     }
-    export namespace xlt95 {
-        const extension_69: string;
+    export namespace xls5 {
+        let extension_69: string;
         export { extension_69 as extension };
-        const desc_69: string;
+        let desc_69: string;
         export { desc_69 as desc };
-        const format_69: string;
+        let format_69: string;
         export { format_69 as format };
     }
-    export namespace xlsx {
-        const extension_70: string;
+    export namespace xls95 {
+        let extension_70: string;
         export { extension_70 as extension };
-        const desc_70: string;
+        let desc_70: string;
         export { desc_70 as desc };
-        const format_70: string;
+        let format_70: string;
         export { format_70 as format };
+    }
+    export namespace xlt {
+        let extension_71: string;
+        export { extension_71 as extension };
+        let desc_71: string;
+        export { desc_71 as desc };
+        let format_71: string;
+        export { format_71 as format };
+    }
+    export namespace xlt5 {
+        let extension_72: string;
+        export { extension_72 as extension };
+        let desc_72: string;
+        export { desc_72 as desc };
+        let format_72: string;
+        export { format_72 as format };
+    }
+    export namespace xlt95 {
+        let extension_73: string;
+        export { extension_73 as extension };
+        let desc_73: string;
+        export { desc_73 as desc };
+        let format_73: string;
+        export { format_73 as format };
+    }
+    export namespace xlsx {
+        let extension_74: string;
+        export { extension_74 as extension };
+        let desc_74: string;
+        export { desc_74 as desc };
+        let format_74: string;
+        export { format_74 as format };
     }
 }
 export namespace graphics {
     export namespace bmp {
-        const extension_71: string;
-        export { extension_71 as extension };
-        const desc_71: string;
-        export { desc_71 as desc };
-        const format_71: string;
-        export { format_71 as format };
-    }
-    export namespace emf {
-        const extension_72: string;
-        export { extension_72 as extension };
-        const desc_72: string;
-        export { desc_72 as desc };
-        const format_72: string;
-        export { format_72 as format };
-    }
-    export namespace eps {
-        const extension_73: string;
-        export { extension_73 as extension };
-        const desc_73: string;
-        export { desc_73 as desc };
-        const format_73: string;
-        export { format_73 as format };
-    }
-    export namespace fodg {
-        const extension_74: string;
-        export { extension_74 as extension };
-        const desc_74: string;
-        export { desc_74 as desc };
-        const format_74: string;
-        export { format_74 as format };
-    }
-    export namespace gif {
-        const extension_75: string;
+        let extension_75: string;
         export { extension_75 as extension };
-        const desc_75: string;
+        let desc_75: string;
         export { desc_75 as desc };
-        const format_75: string;
+        let format_75: string;
         export { format_75 as format };
     }
-    export namespace html_3 {
-        const extension_76: string;
+    export namespace emf {
+        let extension_76: string;
         export { extension_76 as extension };
-        const desc_76: string;
+        let desc_76: string;
         export { desc_76 as desc };
-        const format_76: string;
+        let format_76: string;
         export { format_76 as format };
     }
-    export { html_3 as html };
-    export namespace jpg {
-        const extension_77: string;
+    export namespace eps {
+        let extension_77: string;
         export { extension_77 as extension };
-        const desc_77: string;
+        let desc_77: string;
         export { desc_77 as desc };
-        const format_77: string;
+        let format_77: string;
         export { format_77 as format };
     }
-    export namespace met {
-        const extension_78: string;
+    export namespace fodg {
+        let extension_78: string;
         export { extension_78 as extension };
-        const desc_78: string;
+        let desc_78: string;
         export { desc_78 as desc };
-        const format_78: string;
+        let format_78: string;
         export { format_78 as format };
     }
-    export namespace odd {
-        const extension_79: string;
+    export namespace gif {
+        let extension_79: string;
         export { extension_79 as extension };
-        const desc_79: string;
+        let desc_79: string;
         export { desc_79 as desc };
-        const format_79: string;
+        let format_79: string;
         export { format_79 as format };
     }
-    export namespace otg {
-        const extension_80: string;
+    export namespace html_3 {
+        let extension_80: string;
         export { extension_80 as extension };
-        const desc_80: string;
+        let desc_80: string;
         export { desc_80 as desc };
-        const format_80: string;
+        let format_80: string;
         export { format_80 as format };
     }
-    export namespace pbm {
-        const extension_81: string;
+    export { html_3 as html };
+    export namespace jpg_1 {
+        let extension_81: string;
         export { extension_81 as extension };
-        const desc_81: string;
+        let desc_81: string;
         export { desc_81 as desc };
-        const format_81: string;
+        let format_81: string;
         export { format_81 as format };
     }
-    export namespace pct {
-        const extension_82: string;
+    export { jpg_1 as jpg };
+    export namespace met {
+        let extension_82: string;
         export { extension_82 as extension };
-        const desc_82: string;
+        let desc_82: string;
         export { desc_82 as desc };
-        const format_82: string;
+        let format_82: string;
         export { format_82 as format };
     }
-    export namespace pdf_3 {
-        const extension_83: string;
+    export namespace odd {
+        let extension_83: string;
         export { extension_83 as extension };
-        const desc_83: string;
+        let desc_83: string;
         export { desc_83 as desc };
-        const format_83: string;
+        let format_83: string;
         export { format_83 as format };
+    }
+    export namespace otg {
+        let extension_84: string;
+        export { extension_84 as extension };
+        let desc_84: string;
+        export { desc_84 as desc };
+        let format_84: string;
+        export { format_84 as format };
+    }
+    export namespace pbm {
+        let extension_85: string;
+        export { extension_85 as extension };
+        let desc_85: string;
+        export { desc_85 as desc };
+        let format_85: string;
+        export { format_85 as format };
+    }
+    export namespace pct {
+        let extension_86: string;
+        export { extension_86 as extension };
+        let desc_86: string;
+        export { desc_86 as desc };
+        let format_86: string;
+        export { format_86 as format };
+    }
+    export namespace pdf_3 {
+        let extension_87: string;
+        export { extension_87 as extension };
+        let desc_87: string;
+        export { desc_87 as desc };
+        let format_87: string;
+        export { format_87 as format };
     }
     export { pdf_3 as pdf };
     export namespace pgm {
-        const extension_84: string;
-        export { extension_84 as extension };
-        const desc_84: string;
-        export { desc_84 as desc };
-        const format_84: string;
-        export { format_84 as format };
-    }
-    export namespace png {
-        const extension_85: string;
-        export { extension_85 as extension };
-        const desc_85: string;
-        export { desc_85 as desc };
-        const format_85: string;
-        export { format_85 as format };
-    }
-    export namespace ppm {
-        const extension_86: string;
-        export { extension_86 as extension };
-        const desc_86: string;
-        export { desc_86 as desc };
-        const format_86: string;
-        export { format_86 as format };
-    }
-    export namespace ras {
-        const extension_87: string;
-        export { extension_87 as extension };
-        const desc_87: string;
-        export { desc_87 as desc };
-        const format_87: string;
-        export { format_87 as format };
-    }
-    export namespace std {
-        const extension_88: string;
+        let extension_88: string;
         export { extension_88 as extension };
-        const desc_88: string;
+        let desc_88: string;
         export { desc_88 as desc };
-        const format_88: string;
+        let format_88: string;
         export { format_88 as format };
     }
-    export namespace svg {
-        const extension_89: string;
+    export namespace png_1 {
+        let extension_89: string;
         export { extension_89 as extension };
-        const desc_89: string;
+        let desc_89: string;
         export { desc_89 as desc };
-        const format_89: string;
+        let format_89: string;
         export { format_89 as format };
     }
-    export namespace svm {
-        const extension_90: string;
+    export { png_1 as png };
+    export namespace ppm {
+        let extension_90: string;
         export { extension_90 as extension };
-        const desc_90: string;
+        let desc_90: string;
         export { desc_90 as desc };
-        const format_90: string;
+        let format_90: string;
         export { format_90 as format };
     }
-    export namespace swf {
-        const extension_91: string;
+    export namespace ras {
+        let extension_91: string;
         export { extension_91 as extension };
-        const desc_91: string;
+        let desc_91: string;
         export { desc_91 as desc };
-        const format_91: string;
+        let format_91: string;
         export { format_91 as format };
     }
-    export namespace sxd {
-        const extension_92: string;
+    export namespace std {
+        let extension_92: string;
         export { extension_92 as extension };
-        const desc_92: string;
+        let desc_92: string;
         export { desc_92 as desc };
-        const format_92: string;
+        let format_92: string;
         export { format_92 as format };
     }
-    export namespace sxd3 {
-        const extension_93: string;
+    export namespace svg {
+        let extension_93: string;
         export { extension_93 as extension };
-        const desc_93: string;
+        let desc_93: string;
         export { desc_93 as desc };
-        const format_93: string;
+        let format_93: string;
         export { format_93 as format };
     }
-    export namespace sxd5 {
-        const extension_94: string;
+    export namespace svm {
+        let extension_94: string;
         export { extension_94 as extension };
-        const desc_94: string;
+        let desc_94: string;
         export { desc_94 as desc };
-        const format_94: string;
+        let format_94: string;
         export { format_94 as format };
     }
-    export namespace sxw_1 {
-        const extension_95: string;
+    export namespace swf {
+        let extension_95: string;
         export { extension_95 as extension };
-        const desc_95: string;
+        let desc_95: string;
         export { desc_95 as desc };
-        const format_95: string;
+        let format_95: string;
         export { format_95 as format };
+    }
+    export namespace sxd {
+        let extension_96: string;
+        export { extension_96 as extension };
+        let desc_96: string;
+        export { desc_96 as desc };
+        let format_96: string;
+        export { format_96 as format };
+    }
+    export namespace sxd3 {
+        let extension_97: string;
+        export { extension_97 as extension };
+        let desc_97: string;
+        export { desc_97 as desc };
+        let format_97: string;
+        export { format_97 as format };
+    }
+    export namespace sxd5 {
+        let extension_98: string;
+        export { extension_98 as extension };
+        let desc_98: string;
+        export { desc_98 as desc };
+        let format_98: string;
+        export { format_98 as format };
+    }
+    export namespace sxw_1 {
+        let extension_99: string;
+        export { extension_99 as extension };
+        let desc_99: string;
+        export { desc_99 as desc };
+        let format_99: string;
+        export { format_99 as format };
     }
     export { sxw_1 as sxw };
     export namespace tiff {
-        const extension_96: string;
-        export { extension_96 as extension };
-        const desc_96: string;
-        export { desc_96 as desc };
-        const format_96: string;
-        export { format_96 as format };
+        let extension_100: string;
+        export { extension_100 as extension };
+        let desc_100: string;
+        export { desc_100 as desc };
+        let format_100: string;
+        export { format_100 as format };
     }
     export namespace vor_3 {
-        const extension_97: string;
-        export { extension_97 as extension };
-        const desc_97: string;
-        export { desc_97 as desc };
-        const format_97: string;
-        export { format_97 as format };
+        let extension_101: string;
+        export { extension_101 as extension };
+        let desc_101: string;
+        export { desc_101 as desc };
+        let format_101: string;
+        export { format_101 as format };
     }
     export { vor_3 as vor };
     export namespace vor3_2 {
-        const extension_98: string;
-        export { extension_98 as extension };
-        const desc_98: string;
-        export { desc_98 as desc };
-        const format_98: string;
-        export { format_98 as format };
+        let extension_102: string;
+        export { extension_102 as extension };
+        let desc_102: string;
+        export { desc_102 as desc };
+        let format_102: string;
+        export { format_102 as format };
     }
     export { vor3_2 as vor3 };
     export namespace wmf {
-        const extension_99: string;
-        export { extension_99 as extension };
-        const desc_99: string;
-        export { desc_99 as desc };
-        const format_99: string;
-        export { format_99 as format };
+        let extension_103: string;
+        export { extension_103 as extension };
+        let desc_103: string;
+        export { desc_103 as desc };
+        let format_103: string;
+        export { format_103 as format };
     }
     export namespace xhtml_2 {
-        const extension_100: string;
-        export { extension_100 as extension };
-        const desc_100: string;
-        export { desc_100 as desc };
-        const format_100: string;
-        export { format_100 as format };
+        let extension_104: string;
+        export { extension_104 as extension };
+        let desc_104: string;
+        export { desc_104 as desc };
+        let format_104: string;
+        export { format_104 as format };
     }
     export { xhtml_2 as xhtml };
     export namespace xpm {
-        const extension_101: string;
-        export { extension_101 as extension };
-        const desc_101: string;
-        export { desc_101 as desc };
-        const format_101: string;
-        export { format_101 as format };
+        let extension_105: string;
+        export { extension_105 as extension };
+        let desc_105: string;
+        export { desc_105 as desc };
+        let format_105: string;
+        export { format_105 as format };
     }
 }
 export namespace presentation {
     export namespace bmp_1 {
-        const extension_102: string;
-        export { extension_102 as extension };
-        const desc_102: string;
-        export { desc_102 as desc };
-        const format_102: string;
-        export { format_102 as format };
+        let extension_106: string;
+        export { extension_106 as extension };
+        let desc_106: string;
+        export { desc_106 as desc };
+        let format_106: string;
+        export { format_106 as format };
     }
     export { bmp_1 as bmp };
     export namespace emf_1 {
-        const extension_103: string;
-        export { extension_103 as extension };
-        const desc_103: string;
-        export { desc_103 as desc };
-        const format_103: string;
-        export { format_103 as format };
+        let extension_107: string;
+        export { extension_107 as extension };
+        let desc_107: string;
+        export { desc_107 as desc };
+        let format_107: string;
+        export { format_107 as format };
     }
     export { emf_1 as emf };
     export namespace eps_1 {
-        const extension_104: string;
-        export { extension_104 as extension };
-        const desc_104: string;
-        export { desc_104 as desc };
-        const format_104: string;
-        export { format_104 as format };
+        let extension_108: string;
+        export { extension_108 as extension };
+        let desc_108: string;
+        export { desc_108 as desc };
+        let format_108: string;
+        export { format_108 as format };
     }
     export { eps_1 as eps };
     export namespace fodp {
-        const extension_105: string;
-        export { extension_105 as extension };
-        const desc_105: string;
-        export { desc_105 as desc };
-        const format_105: string;
-        export { format_105 as format };
+        let extension_109: string;
+        export { extension_109 as extension };
+        let desc_109: string;
+        export { desc_109 as desc };
+        let format_109: string;
+        export { format_109 as format };
     }
     export namespace gif_1 {
-        const extension_106: string;
-        export { extension_106 as extension };
-        const desc_106: string;
-        export { desc_106 as desc };
-        const format_106: string;
-        export { format_106 as format };
+        let extension_110: string;
+        export { extension_110 as extension };
+        let desc_110: string;
+        export { desc_110 as desc };
+        let format_110: string;
+        export { format_110 as format };
     }
     export { gif_1 as gif };
     export namespace html_4 {
-        const extension_107: string;
-        export { extension_107 as extension };
-        const desc_107: string;
-        export { desc_107 as desc };
-        const format_107: string;
-        export { format_107 as format };
+        let extension_111: string;
+        export { extension_111 as extension };
+        let desc_111: string;
+        export { desc_111 as desc };
+        let format_111: string;
+        export { format_111 as format };
     }
     export { html_4 as html };
-    export namespace jpg_1 {
-        const extension_108: string;
-        export { extension_108 as extension };
-        const desc_108: string;
-        export { desc_108 as desc };
-        const format_108: string;
-        export { format_108 as format };
+    export namespace jpg_2 {
+        let extension_112: string;
+        export { extension_112 as extension };
+        let desc_112: string;
+        export { desc_112 as desc };
+        let format_112: string;
+        export { format_112 as format };
     }
-    export { jpg_1 as jpg };
+    export { jpg_2 as jpg };
     export namespace met_1 {
-        const extension_109: string;
-        export { extension_109 as extension };
-        const desc_109: string;
-        export { desc_109 as desc };
-        const format_109: string;
-        export { format_109 as format };
+        let extension_113: string;
+        export { extension_113 as extension };
+        let desc_113: string;
+        export { desc_113 as desc };
+        let format_113: string;
+        export { format_113 as format };
     }
     export { met_1 as met };
     export namespace odg {
-        const extension_110: string;
-        export { extension_110 as extension };
-        const desc_110: string;
-        export { desc_110 as desc };
-        const format_110: string;
-        export { format_110 as format };
+        let extension_114: string;
+        export { extension_114 as extension };
+        let desc_114: string;
+        export { desc_114 as desc };
+        let format_114: string;
+        export { format_114 as format };
     }
     export namespace pbm_1 {
-        const extension_111: string;
-        export { extension_111 as extension };
-        const desc_111: string;
-        export { desc_111 as desc };
-        const format_111: string;
-        export { format_111 as format };
+        let extension_115: string;
+        export { extension_115 as extension };
+        let desc_115: string;
+        export { desc_115 as desc };
+        let format_115: string;
+        export { format_115 as format };
     }
     export { pbm_1 as pbm };
     export namespace pct_1 {
-        const extension_112: string;
-        export { extension_112 as extension };
-        const desc_112: string;
-        export { desc_112 as desc };
-        const format_112: string;
-        export { format_112 as format };
+        let extension_116: string;
+        export { extension_116 as extension };
+        let desc_116: string;
+        export { desc_116 as desc };
+        let format_116: string;
+        export { format_116 as format };
     }
     export { pct_1 as pct };
     export namespace pdf_4 {
-        const extension_113: string;
-        export { extension_113 as extension };
-        const desc_113: string;
-        export { desc_113 as desc };
-        const format_113: string;
-        export { format_113 as format };
+        let extension_117: string;
+        export { extension_117 as extension };
+        let desc_117: string;
+        export { desc_117 as desc };
+        let format_117: string;
+        export { format_117 as format };
     }
     export { pdf_4 as pdf };
     export namespace pgm_1 {
-        const extension_114: string;
-        export { extension_114 as extension };
-        const desc_114: string;
-        export { desc_114 as desc };
-        const format_114: string;
-        export { format_114 as format };
+        let extension_118: string;
+        export { extension_118 as extension };
+        let desc_118: string;
+        export { desc_118 as desc };
+        let format_118: string;
+        export { format_118 as format };
     }
     export { pgm_1 as pgm };
-    export namespace png_1 {
-        const extension_115: string;
-        export { extension_115 as extension };
-        const desc_115: string;
-        export { desc_115 as desc };
-        const format_115: string;
-        export { format_115 as format };
+    export namespace png_2 {
+        let extension_119: string;
+        export { extension_119 as extension };
+        let desc_119: string;
+        export { desc_119 as desc };
+        let format_119: string;
+        export { format_119 as format };
     }
-    export { png_1 as png };
+    export { png_2 as png };
     export namespace ppm_1 {
-        const extension_116: string;
-        export { extension_116 as extension };
-        const desc_116: string;
-        export { desc_116 as desc };
-        const format_116: string;
-        export { format_116 as format };
+        let extension_120: string;
+        export { extension_120 as extension };
+        let desc_120: string;
+        export { desc_120 as desc };
+        let format_120: string;
+        export { format_120 as format };
     }
     export { ppm_1 as ppm };
     export namespace pwp {
-        const extension_117: string;
-        export { extension_117 as extension };
-        const desc_117: string;
-        export { desc_117 as desc };
-        const format_117: string;
-        export { format_117 as format };
+        let extension_121: string;
+        export { extension_121 as extension };
+        let desc_121: string;
+        export { desc_121 as desc };
+        let format_121: string;
+        export { format_121 as format };
     }
     export namespace ras_1 {
-        const extension_118: string;
-        export { extension_118 as extension };
-        const desc_118: string;
-        export { desc_118 as desc };
-        const format_118: string;
-        export { format_118 as format };
+        let extension_122: string;
+        export { extension_122 as extension };
+        let desc_122: string;
+        export { desc_122 as desc };
+        let format_122: string;
+        export { format_122 as format };
     }
     export { ras_1 as ras };
     export namespace sda {
-        const extension_119: string;
-        export { extension_119 as extension };
-        const desc_119: string;
-        export { desc_119 as desc };
-        const format_119: string;
-        export { format_119 as format };
+        let extension_123: string;
+        export { extension_123 as extension };
+        let desc_123: string;
+        export { desc_123 as desc };
+        let format_123: string;
+        export { format_123 as format };
     }
     export namespace sdd {
-        const extension_120: string;
-        export { extension_120 as extension };
-        const desc_120: string;
-        export { desc_120 as desc };
-        const format_120: string;
-        export { format_120 as format };
+        let extension_124: string;
+        export { extension_124 as extension };
+        let desc_124: string;
+        export { desc_124 as desc };
+        let format_124: string;
+        export { format_124 as format };
     }
     export namespace sdd3 {
-        const extension_121: string;
-        export { extension_121 as extension };
-        const desc_121: string;
-        export { desc_121 as desc };
-        const format_121: string;
-        export { format_121 as format };
+        let extension_125: string;
+        export { extension_125 as extension };
+        let desc_125: string;
+        export { desc_125 as desc };
+        let format_125: string;
+        export { format_125 as format };
     }
     export namespace sdd4 {
-        const extension_122: string;
-        export { extension_122 as extension };
-        const desc_122: string;
-        export { desc_122 as desc };
-        const format_122: string;
-        export { format_122 as format };
+        let extension_126: string;
+        export { extension_126 as extension };
+        let desc_126: string;
+        export { desc_126 as desc };
+        let format_126: string;
+        export { format_126 as format };
     }
     export namespace sxd_1 {
-        const extension_123: string;
-        export { extension_123 as extension };
-        const desc_123: string;
-        export { desc_123 as desc };
-        const format_123: string;
-        export { format_123 as format };
+        let extension_127: string;
+        export { extension_127 as extension };
+        let desc_127: string;
+        export { desc_127 as desc };
+        let format_127: string;
+        export { format_127 as format };
     }
     export { sxd_1 as sxd };
     export namespace sti {
-        const extension_124: string;
-        export { extension_124 as extension };
-        const desc_124: string;
-        export { desc_124 as desc };
-        const format_124: string;
-        export { format_124 as format };
+        let extension_128: string;
+        export { extension_128 as extension };
+        let desc_128: string;
+        export { desc_128 as desc };
+        let format_128: string;
+        export { format_128 as format };
     }
     export namespace svg_1 {
-        const extension_125: string;
-        export { extension_125 as extension };
-        const desc_125: string;
-        export { desc_125 as desc };
-        const format_125: string;
-        export { format_125 as format };
+        let extension_129: string;
+        export { extension_129 as extension };
+        let desc_129: string;
+        export { desc_129 as desc };
+        let format_129: string;
+        export { format_129 as format };
     }
     export { svg_1 as svg };
     export namespace svm_1 {
-        const extension_126: string;
-        export { extension_126 as extension };
-        const desc_126: string;
-        export { desc_126 as desc };
-        const format_126: string;
-        export { format_126 as format };
+        let extension_130: string;
+        export { extension_130 as extension };
+        let desc_130: string;
+        export { desc_130 as desc };
+        let format_130: string;
+        export { format_130 as format };
     }
     export { svm_1 as svm };
     export namespace swf_1 {
-        const extension_127: string;
-        export { extension_127 as extension };
-        const desc_127: string;
-        export { desc_127 as desc };
-        const format_127: string;
-        export { format_127 as format };
+        let extension_131: string;
+        export { extension_131 as extension };
+        let desc_131: string;
+        export { desc_131 as desc };
+        let format_131: string;
+        export { format_131 as format };
     }
     export { swf_1 as swf };
     export namespace sxi {
-        const extension_128: string;
-        export { extension_128 as extension };
-        const desc_128: string;
-        export { desc_128 as desc };
-        const format_128: string;
-        export { format_128 as format };
+        let extension_132: string;
+        export { extension_132 as extension };
+        let desc_132: string;
+        export { desc_132 as desc };
+        let format_132: string;
+        export { format_132 as format };
     }
     export namespace tiff_1 {
-        const extension_129: string;
-        export { extension_129 as extension };
-        const desc_129: string;
-        export { desc_129 as desc };
-        const format_129: string;
-        export { format_129 as format };
+        let extension_133: string;
+        export { extension_133 as extension };
+        let desc_133: string;
+        export { desc_133 as desc };
+        let format_133: string;
+        export { format_133 as format };
     }
     export { tiff_1 as tiff };
     export namespace uop {
-        const extension_130: string;
-        export { extension_130 as extension };
-        const desc_130: string;
-        export { desc_130 as desc };
-        const format_130: string;
-        export { format_130 as format };
+        let extension_134: string;
+        export { extension_134 as extension };
+        let desc_134: string;
+        export { desc_134 as desc };
+        let format_134: string;
+        export { format_134 as format };
     }
     export namespace vor_4 {
-        const extension_131: string;
-        export { extension_131 as extension };
-        const desc_131: string;
-        export { desc_131 as desc };
-        const format_131: string;
-        export { format_131 as format };
+        let extension_135: string;
+        export { extension_135 as extension };
+        let desc_135: string;
+        export { desc_135 as desc };
+        let format_135: string;
+        export { format_135 as format };
     }
     export { vor_4 as vor };
     export namespace vor3_3 {
-        const extension_132: string;
-        export { extension_132 as extension };
-        const desc_132: string;
-        export { desc_132 as desc };
-        const format_132: string;
-        export { format_132 as format };
+        let extension_136: string;
+        export { extension_136 as extension };
+        let desc_136: string;
+        export { desc_136 as desc };
+        let format_136: string;
+        export { format_136 as format };
     }
     export { vor3_3 as vor3 };
     export namespace vor4_3 {
-        const extension_133: string;
-        export { extension_133 as extension };
-        const desc_133: string;
-        export { desc_133 as desc };
-        const format_133: string;
-        export { format_133 as format };
+        let extension_137: string;
+        export { extension_137 as extension };
+        let desc_137: string;
+        export { desc_137 as desc };
+        let format_137: string;
+        export { format_137 as format };
     }
     export { vor4_3 as vor4 };
     export namespace vor5 {
-        const extension_134: string;
-        export { extension_134 as extension };
-        const desc_134: string;
-        export { desc_134 as desc };
-        const format_134: string;
-        export { format_134 as format };
+        let extension_138: string;
+        export { extension_138 as extension };
+        let desc_138: string;
+        export { desc_138 as desc };
+        let format_138: string;
+        export { format_138 as format };
     }
     export namespace wmf_1 {
-        const extension_135: string;
-        export { extension_135 as extension };
-        const desc_135: string;
-        export { desc_135 as desc };
-        const format_135: string;
-        export { format_135 as format };
+        let extension_139: string;
+        export { extension_139 as extension };
+        let desc_139: string;
+        export { desc_139 as desc };
+        let format_139: string;
+        export { format_139 as format };
     }
     export { wmf_1 as wmf };
     export namespace xhtml_3 {
-        const extension_136: string;
-        export { extension_136 as extension };
-        const desc_136: string;
-        export { desc_136 as desc };
-        const format_136: string;
-        export { format_136 as format };
+        let extension_140: string;
+        export { extension_140 as extension };
+        let desc_140: string;
+        export { desc_140 as desc };
+        let format_140: string;
+        export { format_140 as format };
     }
     export { xhtml_3 as xhtml };
     export namespace xpm_1 {
-        const extension_137: string;
-        export { extension_137 as extension };
-        const desc_137: string;
-        export { desc_137 as desc };
-        const format_137: string;
-        export { format_137 as format };
+        let extension_141: string;
+        export { extension_141 as extension };
+        let desc_141: string;
+        export { desc_141 as desc };
+        let format_141: string;
+        export { format_141 as format };
     }
     export { xpm_1 as xpm };
     export namespace odp {
-        const extension_138: string;
-        export { extension_138 as extension };
-        const desc_138: string;
-        export { desc_138 as desc };
-        const format_138: string;
-        export { format_138 as format };
-    }
-    export namespace otp {
-        const extension_139: string;
-        export { extension_139 as extension };
-        const desc_139: string;
-        export { desc_139 as desc };
-        const format_139: string;
-        export { format_139 as format };
-    }
-    export namespace potm {
-        const extension_140: string;
-        export { extension_140 as extension };
-        const desc_140: string;
-        export { desc_140 as desc };
-        const format_140: string;
-        export { format_140 as format };
-    }
-    export namespace pot {
-        const extension_141: string;
-        export { extension_141 as extension };
-        const desc_141: string;
-        export { desc_141 as desc };
-        const format_141: string;
-        export { format_141 as format };
-    }
-    export namespace pptx {
-        const extension_142: string;
+        let extension_142: string;
         export { extension_142 as extension };
-        const desc_142: string;
+        let desc_142: string;
         export { desc_142 as desc };
-        const format_142: string;
+        let format_142: string;
         export { format_142 as format };
     }
-    export namespace pps {
-        const extension_143: string;
+    export namespace otp {
+        let extension_143: string;
         export { extension_143 as extension };
-        const desc_143: string;
+        let desc_143: string;
         export { desc_143 as desc };
-        const format_143: string;
+        let format_143: string;
         export { format_143 as format };
     }
-    export namespace ppt {
-        const extension_144: string;
+    export namespace potm {
+        let extension_144: string;
         export { extension_144 as extension };
-        const desc_144: string;
+        let desc_144: string;
         export { desc_144 as desc };
-        const format_144: string;
+        let format_144: string;
         export { format_144 as format };
+    }
+    export namespace pot {
+        let extension_145: string;
+        export { extension_145 as extension };
+        let desc_145: string;
+        export { desc_145 as desc };
+        let format_145: string;
+        export { format_145 as format };
+    }
+    export namespace pptx {
+        let extension_146: string;
+        export { extension_146 as extension };
+        let desc_146: string;
+        export { desc_146 as desc };
+        let format_146: string;
+        export { format_146 as format };
+    }
+    export namespace pps {
+        let extension_147: string;
+        export { extension_147 as extension };
+        let desc_147: string;
+        export { desc_147 as desc };
+        let format_147: string;
+        export { format_147 as format };
+    }
+    export namespace ppt {
+        let extension_148: string;
+        export { extension_148 as extension };
+        let desc_148: string;
+        export { desc_148 as desc };
+        let format_148: string;
+        export { format_148 as format };
     }
 }
 //# sourceMappingURL=format.d.ts.map
