@@ -1,49 +1,55 @@
 declare var tmpDir: any;
 export { tmpDir as tempPath };
-export declare const templatePath: any;
-export declare const factories: number;
-export declare const attempts: number;
-export declare const startFactory: boolean;
-export declare const uidPrefix: string;
-export declare const pipeNamePrefix: string;
-export declare const extensionParsed: string;
-export declare const lang: string;
-export declare const translations: {};
-export declare const currencySource: string;
-export declare const currencyTarget: string;
+export declare let pythonPath: any;
+export declare let renderPath: any;
+export declare let templatePath: any;
+export declare let factories: number;
+export declare let attempts: number;
+export declare let startFactory: boolean;
+export declare let factoryMemoryFileSize: number;
+export declare let factoryMemoryThreshold: number;
+export declare let converterFactoryTimeout: number;
+export declare let uidPrefix: string;
+export declare let pipeNamePrefix: string;
+export declare let extensionParsed: string;
+export declare let lang: string;
+export declare let timezone: string;
+export declare let translations: {};
+export declare let currencySource: string;
+export declare let currencyTarget: string;
 export declare namespace currencyRates {
-    const EUR: number;
-    const USD: number;
-    const JPY: number;
-    const BGN: number;
-    const CZK: number;
-    const DKK: number;
-    const GBP: number;
-    const HUF: number;
-    const PLN: number;
-    const RON: number;
-    const SEK: number;
-    const CHF: number;
-    const ISK: number;
-    const NOK: number;
-    const HRK: number;
-    const RUB: number;
-    const TRY: number;
-    const AUD: number;
-    const BRL: number;
-    const CAD: number;
-    const CNY: number;
-    const HKD: number;
-    const IDR: number;
-    const ILS: number;
-    const INR: number;
-    const KRW: number;
-    const MXN: number;
-    const MYR: number;
-    const NZD: number;
-    const PHP: number;
-    const SGD: number;
-    const THB: number;
-    const ZAR: number;
+    let EUR: number;
+    let USD: number;
+    let JPY: number;
+    let BGN: number;
+    let CZK: number;
+    let DKK: number;
+    let GBP: number;
+    let HUF: number;
+    let PLN: number;
+    let RON: number;
+    let SEK: number;
+    let CHF: number;
+    let ISK: number;
+    let NOK: number;
+    let HRK: number;
+    let RUB: number;
+    let TRY: number;
+    let AUD: number;
+    let BRL: number;
+    let CAD: number;
+    let CNY: number;
+    let HKD: number;
+    let IDR: number;
+    let ILS: number;
+    let INR: number;
+    let KRW: number;
+    let MXN: number;
+    let MYR: number;
+    let NZD: number;
+    let PHP: number;
+    let SGD: number;
+    let THB: number;
+    let ZAR: number;
 }
 //# sourceMappingURL=params.d.ts.map
