@@ -1,5 +1,8 @@
-export function execute(template: any, callback: Function): any;
+export function execute(template: any, options: any, callback: Function): any;
+export function preParseXML(xml: any, options: any): any;
+export function removeSoftPageBreak(template: any): any;
 export function convertSharedStringToInlineString(template: any, parentFilter: string): any;
+export function removeCalcChain(template: <type>() => any, parentFilter: <type>() => any): <type>() => any;
 export function removeOneFile(template: any, indexOfFileToRemove: Integer, parentFilter: string): any;
 export function readSharedString(sharedStringXml: string): any[];
 export function convertToInlineString(xml: string, sharedStrings: any[]): string;
